Struct competitive::algorithm::XorBasis

pub struct XorBasis { /* private fields */ }

Basis of xor operation.

Implementations

impl XorBasis

pub fn new() -> Self

Create a empty space.

pub fn reduce(&self, x: u64) -> (u64, u64)

Return (reduced basis, coordinate). Coordinate means if i-th bit is 1, x was reduced by i-th inserted basis.

Examples found in repository?

crates/competitive/src/algorithm/xorbasis.rs (line 27)

26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56

    pub fn insert(&mut self, x: u64) -> bool {
        let (y, coord) = self.reduce(x);
        if y != 0 {
            self.bases.push((y, coord, x));
        }
        y != 0
    }
    /// Return coordinate if element can be consisted by current basis.
    pub fn find(&self, x: u64) -> Option<u64> {
        let (y, coord) = self.reduce(x);
        if y == 0 {
            Some(coord)
        } else {
            None
        }
    }
    /// Return coordinate if element can be consisted by current basis.
    pub fn basis(&self, x: u64) -> Option<Vec<u64>> {
        let (y, coord) = self.reduce(x);
        if y == 0 {
            Some(
                self.bases
                    .iter()
                    .enumerate()
                    .filter_map(|(i, (_, _, b))| if coord & 1 << i != 0 { Some(*b) } else { None })
                    .collect(),
            )
        } else {
            None
        }
    }

pub fn insert(&mut self, x: u64) -> bool

Return true if inserted element cannot be consisted by current basis and be added as a new basis. Return false if inserted element can be consisted by current basis.

Examples found in repository?

crates/competitive/src/algorithm/xorbasis.rs (line 62)

59
60
61
62
63
64
65

    fn from_iter<T: IntoIterator<Item = u64>>(iter: T) -> Self {
        let mut basis = XorBasis::default();
        for x in iter {
            basis.insert(x);
        }
        basis
    }

pub fn find(&self, x: u64) -> Option<u64>

Return coordinate if element can be consisted by current basis.

pub fn basis(&self, x: u64) -> Option<Vec<u64>>

Return coordinate if element can be consisted by current basis.