competitive::data_structure::splay_operations::with_parent

Function rotate

unsafe fn rotate<Spec, Data>(
    node: BstNodePtr<<Spec as BstSpec>::Data, <Spec as BstSpec>::Parent>,
)
where
    Spec: BstSpec<Data = Data, Parent = WithParent<Data>>,

Examples found in repository?

crates/competitive/src/data_structure/splay_operations.rs (line 105)

70    pub unsafe fn splay<Spec, Data>(node: NodePtr<Spec>) -> NodePtr<Spec>
71    where
72        Spec: BstSpec<Data = Data, Parent = WithParent<Data>>,
73    {
74        let mut inline_stack = [const { MaybeUninit::uninit() }; 64];
75        let mut inline_len = 0;
76        let mut overflow_stack = Vec::new();
77        let mut current = node;
78        loop {
79            if inline_len < inline_stack.len() {
80                inline_stack[inline_len].write(current);
81                inline_len += 1;
82            } else {
83                overflow_stack.push(current);
84            }
85            match unsafe { internal_parent::<Spec, Data>(current) } {
86                Ok((parent, _)) => current = parent,
87                Err(_) => break,
88            }
89        }
90        for &node in overflow_stack.iter().rev() {
91            unsafe { Spec::top_down(BstDataMutRef::new_unchecked(node)) };
92        }
93        while inline_len > 0 {
94            inline_len -= 1;
95            unsafe {
96                Spec::top_down(BstDataMutRef::new_unchecked(
97                    *inline_stack[inline_len].assume_init_ref(),
98                ));
99            }
100        }
101
102        while let Ok((parent, node_direction)) = unsafe { internal_parent::<Spec, Data>(node) } {
103            if let Ok((_, parent_direction)) = unsafe { internal_parent::<Spec, Data>(parent) } {
104                if node_direction == parent_direction {
105                    unsafe { rotate::<Spec, Data>(parent) };
106                } else {
107                    unsafe { rotate::<Spec, Data>(node) };
108                }
109            }
110            unsafe { rotate::<Spec, Data>(node) };
111        }
112        unsafe { Spec::bottom_up(BstDataMutRef::new_unchecked(node)) };
113        current
114    }
115
116    /// Moves `node` to the root by propagating only the nodes involved in each rotation and
117    /// returns the previous root.
118    ///
119    /// # Safety
120    ///
121    /// `node` and every pointer reachable through its auxiliary-parent chain must refer to live
122    /// nodes of the same tree. Propagating an ancestor after its descendant must be valid for
123    /// `Spec`.
124    #[inline(always)]
125    pub unsafe fn splay_with_local_top_down<Spec, Data>(node: NodePtr<Spec>) -> NodePtr<Spec>
126    where
127        Spec: BstSpec<Data = Data, Parent = WithParent<Data>>,
128    {
129        let mut current = node;
130        unsafe { Spec::top_down(BstDataMutRef::new_unchecked(node)) };
131        while let Ok((parent, node_direction)) = unsafe { internal_parent::<Spec, Data>(node) } {
132            match unsafe { internal_parent::<Spec, Data>(parent) } {
133                Ok((grandparent, parent_direction)) => {
134                    current = grandparent;
135                    unsafe {
136                        Spec::top_down(BstDataMutRef::new_unchecked(grandparent));
137                        Spec::top_down(BstDataMutRef::new_unchecked(parent));
138                        Spec::top_down(BstDataMutRef::new_unchecked(node));
139                    }
140                    if node_direction == parent_direction {
141                        unsafe { rotate::<Spec, Data>(parent) };
142                    } else {
143                        unsafe { rotate::<Spec, Data>(node) };
144                    }
145                }
146                Err(_) => {
147                    current = parent;
148                    unsafe {
149                        Spec::top_down(BstDataMutRef::new_unchecked(parent));
150                        Spec::top_down(BstDataMutRef::new_unchecked(node));
151                    }
152                }
153            }
154            unsafe { rotate::<Spec, Data>(node) };
155        }
156        unsafe { Spec::bottom_up(BstDataMutRef::new_unchecked(node)) };
157        current
158    }