Struct competitive::data_structure::LazySegmentTreeMap

pub struct LazySegmentTreeMap<M>
where
    M: MonoidAction,
    M::Act: PartialEq,
{ /* private fields */ }

Implementations

impl<M> LazySegmentTreeMap<M>
where M: MonoidAction, M::Act: PartialEq,

pub fn new(n: usize) -> Self

pub fn update(&mut self, l: usize, r: usize, x: M::Act)

pub fn fold(&mut self, l: usize, r: usize) -> M::Agg

Examples found in repository?

crates/competitive/src/data_structure/lazy_segment_tree_map.rs (line 170)

169
170
171
172
173
174

    pub fn get(&mut self, k: usize) -> M::Agg {
        self.fold(k, k + 1)
    }
    pub fn fold_all(&mut self) -> M::Agg {
        self.fold(0, self.n)
    }

pub fn set(&mut self, k: usize, x: M::Agg)

pub fn get(&mut self, k: usize) -> M::Agg

pub fn fold_all(&mut self) -> M::Agg

pub fn position_acc<P>(&mut self, l: usize, r: usize, p: P) -> Option<usize>
where P: Fn(&M::Agg) -> bool,

Returns the first index that satisfies a accumlative predicate.

pub fn rposition_acc<P>(&mut self, l: usize, r: usize, p: P) -> Option<usize>
where P: Fn(&M::Agg) -> bool,

Returns the last index that satisfies a accumlative predicate.