Module competitive::data_structures::fenwick_tree

Structs

FenwickTree
In a max bit tree or Fenwick Tree, get(i) will return the largest element e that has been added to the bit tree with set(j, e), where j <= i. Initially all positions have the value T::default(). Note that a set cannot be 'undone' by inserting a smaller element at the same index. Time Complexity: O(n) to build a new tree or O(log n) for get() and set() operations, where n = tree.len().

MaxOp

SumOp

Traits

PrefixOp
Fenwick tree prefix operator

Type Definitions

MaxBitTree
Fenwick tree specialized for prefix-max

SumBitTree
Fenwick tree specialized for prefix-sum