Function competitive::graph::diktstra

pub fn diktstra<W>(graph: &ListGraph<W>, start: usize) -> (Vec<W>, Vec<usize>) where
    W: Copy + One + Zero + PartialEq + PartialOrd + NumCast + Bounded + Debug, 

Diktstra
O(|E+V|log(|V|))
let E be edge number, let V be vertex number. return distance from start and prev nodes information. restore_path function can create shortest path from start to goal from prev nodes.

use competitive::graph::*;
let vec = vec![
    (0, 1, 1),
    (0, 2, 4),
    (2, 0, 1),
    (1, 2, 2),
    (3, 1, 1),
    (3, 2, 5),
];
let graph: ListGraph<isize> = ListGraph::weighted_from(vec, 4, 0, Direction::DiGraph);

let (w, prev_nodes) = diktstra(&graph, 1);
assert_eq!(w, vec![3, 0, 2, std::isize::MAX]);
assert_eq!(restore_path(1, 0, &prev_nodes), vec![1, 2, 0]);