Struct competitive::tools::Xorshift

pub struct Xorshift { /* private fields */ }

Implementations

impl Xorshift

pub fn gen<T, R>(&mut self, spec: R) -> T
where R: RandomSpec<T>,

Examples found in repository?

crates/competitive/src/tools/random_generator.rs (line 75)

74
75
76
77
78
79
80
81
82
83
84
85
86
87
88
89
90
91
92
93
94
95
96
97
98
99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122
123
124
125
126
127
128
129
130
131
132
133
134
135
136
137
138
139
140
141
142
143
144
145
146
147
148
149
150
151
152
153
154
155
156
157
158
159
160
161
162
163
164
165
166
167
168
169
170
171
172
173
174
175
176
177
178
179
180
181
182
183
184
185
186
187
188
189
190
191
192
193
194
195
196
197
198
199
200
201
202
203
204
205
206
207
208
209
210
211
212
213
214
215
216
217
218
219
220
221
222
223
224
225
226
227
228
229
230
231
232
233
234
235
236
237
238
239
240
241
242
243
244
245
246
247
248
249
250
251
252
253
254
255
256
257
258
259
260

    fn rand(&self, rng: &mut Xorshift) -> i128 {
        rng.gen::<u128, _>(..) as i128
    }
}

macro_rules! impl_random_spec_ranges {
    ($($u:ident $i:ident)*) => {
        $(
            impl RandomSpec<$u> for Range<$u> {
                fn rand(&self, rng: &mut Xorshift) -> $u {
                    assert!(self.start < self.end);
                    let len = self.end - self.start;
                    (self.start + rng.gen::<$u, _>(..) % len)
                }
            }
            impl RandomSpec<$i> for Range<$i> {
                fn rand(&self, rng: &mut Xorshift) -> $i {
                    assert!(self.start < self.end);
                    let len = self.end.abs_diff(self.start);
                    self.start.wrapping_add_unsigned(rng.gen::<$u, _>(..) % len)
                }
            }
            impl RandomSpec<$u> for RangeFrom<$u> {
                fn rand(&self, rng: &mut Xorshift) -> $u {
                    let len = ($u::MAX - self.start).wrapping_add(1);
                    let x = rng.gen::<$u, _>(..);
                    self.start + if len != 0 { x % len } else { x }
                }
            }
            impl RandomSpec<$i> for RangeFrom<$i> {
                fn rand(&self, rng: &mut Xorshift) -> $i {
                    let len = ($i::MAX.abs_diff(self.start)).wrapping_add(1);
                    let x = rng.gen::<$u, _>(..);
                    self.start.wrapping_add_unsigned(if len != 0 { x % len } else { x })
                }
            }
            impl RandomSpec<$u> for RangeInclusive<$u> {
                fn rand(&self, rng: &mut Xorshift) -> $u {
                    assert!(self.start() <= self.end());
                    let len = (self.end() - self.start()).wrapping_add(1);
                    let x = rng.gen::<$u, _>(..);
                    self.start() + if len != 0 { x % len } else { x }
                }
            }
            impl RandomSpec<$i> for RangeInclusive<$i> {
                fn rand(&self, rng: &mut Xorshift) -> $i {
                    assert!(self.start() <= self.end());
                    let len = (self.end().abs_diff(*self.start())).wrapping_add(1);
                    let x = rng.gen::<$u, _>(..);
                    self.start().wrapping_add_unsigned(if len != 0 { x % len } else { x })
                }
            }
            impl RandomSpec<$u> for RangeTo<$u> {
                fn rand(&self, rng: &mut Xorshift) -> $u {
                    let len = self.end;
                    rng.gen::<$u, _>(..) % len
                }
            }
            impl RandomSpec<$i> for RangeTo<$i> {
                fn rand(&self, rng: &mut Xorshift) -> $i {
                    let len = self.end.abs_diff($i::MIN);
                    $i::MIN.wrapping_add_unsigned(rng.gen::<$u, _>(..) % len)
                }
            }
            impl RandomSpec<$u> for RangeToInclusive<$u> {
                fn rand(&self, rng: &mut Xorshift) -> $u {
                    let len = (self.end).wrapping_add(1);
                    let x = rng.gen::<$u, _>(..);
                    if len != 0 { x % len } else { x }
                }
            }
            impl RandomSpec<$i> for RangeToInclusive<$i> {
                fn rand(&self, rng: &mut Xorshift) -> $i {
                    let len = (self.end.abs_diff($i::MIN)).wrapping_add(1);
                    let x = rng.gen::<$u, _>(..);
                    $i::MIN.wrapping_add_unsigned(if len != 0 { x % len } else { x })
                }
            }
        )*
    };
}
impl_random_spec_ranges!(u8 i8 u16 i16 u32 i32 u64 i64 u128 i128 usize isize);

macro_rules! impl_random_spec_tuple {
    ($($T:ident)*, $($R:ident)*, $($v:ident)*) => {
        impl<$($T),*, $($R),*> RandomSpec<($($T,)*)> for ($($R,)*)
        where
            $($R: RandomSpec<$T>),*
        {
            fn rand(&self, rng: &mut Xorshift) -> ($($T,)*) {
                let ($($v,)*) = self;
                ($(($v).rand(rng),)*)
            }
        }
    };
}
impl_random_spec_tuple!(A, RA, a);
impl_random_spec_tuple!(A B, RA RB, a b);
impl_random_spec_tuple!(A B C, RA RB RC, a b c);
impl_random_spec_tuple!(A B C D, RA RB RC RD, a b c d);
impl_random_spec_tuple!(A B C D E, RA RB RC RD RE, a b c d e);
impl_random_spec_tuple!(A B C D E F, RA RB RC RD RE RF, a b c d e f);
impl_random_spec_tuple!(A B C D E F G, RA RB RC RD RE RF RG, a b c d e f g);
impl_random_spec_tuple!(A B C D E F G H, RA RB RC RD RE RF RG RH, a b c d e f g h);
impl_random_spec_tuple!(A B C D E F G H I, RA RB RC RD RE RF RG RH RI, a b c d e f g h i);
impl_random_spec_tuple!(A B C D E F G H I J, RA RB RC RD RE RF RG RH RI RJ, a b c d e f g h i j);

macro_rules! impl_random_spec_primitive {
    ($($t:ty)*) => {
        $(impl RandomSpec<$t> for $t {
            fn rand(&self, _rng: &mut Xorshift) -> $t {
                *self
            }
        })*
    };
}
impl_random_spec_primitive!(() u8 u16 u32 u64 u128 usize i8 i16 i32 i64 i128 isize bool char);

impl<T, R> RandomSpec<T> for &R
where
    R: RandomSpec<T>,
{
    fn rand(&self, rng: &mut Xorshift) -> T {
        <R as RandomSpec<T>>::rand(self, rng)
    }
}
impl<T, R> RandomSpec<T> for &mut R
where
    R: RandomSpec<T>,
{
    fn rand(&self, rng: &mut Xorshift) -> T {
        <R as RandomSpec<T>>::rand(self, rng)
    }
}

#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
/// Left-close Right-open No Empty Segment
pub struct NotEmptySegment<T>(pub T);
impl<T> RandomSpec<(usize, usize)> for NotEmptySegment<T>
where
    T: RandomSpec<usize>,
{
    fn rand(&self, rng: &mut Xorshift) -> (usize, usize) {
        let n = rng.gen(&self.0) as u64;
        let k = randint_uniform(rng, n);
        let l = randint_uniform(rng, n - k) as usize;
        (l, l + k as usize + 1)
    }
}

#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
pub struct RandRange<Q, T> {
    data: Q,
    _marker: PhantomData<fn() -> T>,
}
impl<Q, T> RandRange<Q, T> {
    pub fn new(data: Q) -> Self {
        Self {
            data,
            _marker: PhantomData,
        }
    }
}
impl<Q, T> RandomSpec<(Bound<T>, Bound<T>)> for RandRange<Q, T>
where
    Q: RandomSpec<T>,
    T: Ord,
{
    fn rand(&self, rng: &mut Xorshift) -> (Bound<T>, Bound<T>) {
        let mut l = rng.gen(&self.data);
        let mut r = rng.gen(&self.data);
        if l > r {
            swap(&mut l, &mut r);
        }
        (
            match rng.rand(3) {
                0 => Bound::Excluded(l),
                1 => Bound::Included(l),
                _ => Bound::Unbounded,
            },
            match rng.rand(3) {
                0 => Bound::Excluded(r),
                1 => Bound::Included(r),
                _ => Bound::Unbounded,
            },
        )
    }

Hide additional examples

crates/competitive/src/tree/generator.rs (line 8)

7
8
9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70
71
72

    fn rand(&self, rng: &mut Xorshift) -> UndirectedSparseGraph {
        let n = rng.gen(&self.0);
        let edges = from_prufer_sequence(
            n,
            &rng.gen_iter(0..n)
                .take(n.saturating_sub(2))
                .collect::<Vec<usize>>(),
        );
        UndirectedSparseGraph::from_edges(n, edges)
    }
}

pub struct PathTree<T>(pub T);

impl<T: RandomSpec<usize>> RandomSpec<UndirectedSparseGraph> for PathTree<T> {
    fn rand(&self, rng: &mut Xorshift) -> UndirectedSparseGraph {
        let n = rng.gen(&self.0);
        let edges = (1..n).map(|u| (u - 1, u)).collect();
        UndirectedSparseGraph::from_edges(n, edges)
    }
}

pub struct StarTree<T>(pub T);

impl<T: RandomSpec<usize>> RandomSpec<UndirectedSparseGraph> for StarTree<T> {
    fn rand(&self, rng: &mut Xorshift) -> UndirectedSparseGraph {
        let n = rng.gen(&self.0);
        let edges = (1..n).map(|u| (0, u)).collect();
        UndirectedSparseGraph::from_edges(n, edges)
    }
}

pub struct MixedTree<T>(pub T);

impl<T: RandomSpec<usize>> RandomSpec<UndirectedSparseGraph> for MixedTree<T> {
    fn rand(&self, rng: &mut Xorshift) -> UndirectedSparseGraph {
        fn rand_inner(n: usize, rng: &mut Xorshift) -> Vec<(usize, usize)> {
            let mut edges = Vec::with_capacity(n.saturating_sub(1));
            if n >= 2 {
                let k = rng.gen(1..n);
                for n in [k, n - k].iter().cloned() {
                    let ty = rng.rand(6);
                    edges.extend(match ty {
                        0 => from_prufer_sequence(
                            n,
                            &rng.gen_iter(0..n)
                                .take(n.saturating_sub(2))
                                .collect::<Vec<usize>>(),
                        ),
                        1 => (1..n).map(|u| (u - 1, u)).collect(),
                        2 => (1..n).map(|u| (0, u)).collect(),
                        _ => rand_inner(n, rng),
                    });
                }
                for (u, v) in edges[k - 1..].iter_mut() {
                    *u += k;
                    *v += k;
                }
                edges.push((rng.gen(0..k), rng.gen(k..n)));
            }
            edges
        }
        let n = rng.gen(&self.0);
        let edges = rand_inner(n, rng);
        UndirectedSparseGraph::from_edges(n, edges)
    }

pub fn gen_iter<T, R>(&mut self, spec: R) -> RandIter<'_, T, R>
where R: RandomSpec<T>,

Examples found in repository?

crates/competitive/src/tree/generator.rs (line 11)

7
8
9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68

    fn rand(&self, rng: &mut Xorshift) -> UndirectedSparseGraph {
        let n = rng.gen(&self.0);
        let edges = from_prufer_sequence(
            n,
            &rng.gen_iter(0..n)
                .take(n.saturating_sub(2))
                .collect::<Vec<usize>>(),
        );
        UndirectedSparseGraph::from_edges(n, edges)
    }
}

pub struct PathTree<T>(pub T);

impl<T: RandomSpec<usize>> RandomSpec<UndirectedSparseGraph> for PathTree<T> {
    fn rand(&self, rng: &mut Xorshift) -> UndirectedSparseGraph {
        let n = rng.gen(&self.0);
        let edges = (1..n).map(|u| (u - 1, u)).collect();
        UndirectedSparseGraph::from_edges(n, edges)
    }
}

pub struct StarTree<T>(pub T);

impl<T: RandomSpec<usize>> RandomSpec<UndirectedSparseGraph> for StarTree<T> {
    fn rand(&self, rng: &mut Xorshift) -> UndirectedSparseGraph {
        let n = rng.gen(&self.0);
        let edges = (1..n).map(|u| (0, u)).collect();
        UndirectedSparseGraph::from_edges(n, edges)
    }
}

pub struct MixedTree<T>(pub T);

impl<T: RandomSpec<usize>> RandomSpec<UndirectedSparseGraph> for MixedTree<T> {
    fn rand(&self, rng: &mut Xorshift) -> UndirectedSparseGraph {
        fn rand_inner(n: usize, rng: &mut Xorshift) -> Vec<(usize, usize)> {
            let mut edges = Vec::with_capacity(n.saturating_sub(1));
            if n >= 2 {
                let k = rng.gen(1..n);
                for n in [k, n - k].iter().cloned() {
                    let ty = rng.rand(6);
                    edges.extend(match ty {
                        0 => from_prufer_sequence(
                            n,
                            &rng.gen_iter(0..n)
                                .take(n.saturating_sub(2))
                                .collect::<Vec<usize>>(),
                        ),
                        1 => (1..n).map(|u| (u - 1, u)).collect(),
                        2 => (1..n).map(|u| (0, u)).collect(),
                        _ => rand_inner(n, rng),
                    });
                }
                for (u, v) in edges[k - 1..].iter_mut() {
                    *u += k;
                    *v += k;
                }
                edges.push((rng.gen(0..k), rng.gen(k..n)));
            }
            edges
        }

impl Xorshift

pub fn new_with_seed(seed: u64) -> Self

Examples found in repository?

crates/competitive/src/tools/xorshift.rs (line 18)

17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62

    pub fn new() -> Self {
        Xorshift::new_with_seed(RandomState::new().build_hasher().finish())
    }

    pub fn rand64(&mut self) -> u64 {
        self.y ^= self.y << 5;
        self.y ^= self.y >> 17;
        self.y ^= self.y << 11;
        self.y
    }

    pub fn rand(&mut self, k: u64) -> u64 {
        self.rand64() % k
    }

    pub fn rands(&mut self, k: u64, n: usize) -> Vec<u64> {
        repeat_with(|| self.rand(k)).take(n).collect()
    }

    pub fn randf(&mut self) -> f64 {
        const UPPER_MASK: u64 = 0x3FF0_0000_0000_0000;
        const LOWER_MASK: u64 = 0x000F_FFFF_FFFF_FFFF;
        let x = self.rand64();
        let tmp = UPPER_MASK | (x & LOWER_MASK);
        let result: f64 = f64::from_bits(tmp);
        f64::from_bits(f64::to_bits(result - 1.0) ^ x >> 63)
    }

    pub fn gen_bool(&mut self, p: f64) -> bool {
        self.randf() < p
    }

    pub fn shuffle<T>(&mut self, slice: &mut [T]) {
        let mut n = slice.len();
        while n > 1 {
            let i = self.rand(n as _) as usize;
            n -= 1;
            slice.swap(i, n);
        }
    }
}

impl Default for Xorshift {
    fn default() -> Self {
        Xorshift::new_with_seed(0x2b99_2ddf_a232_49d6)
    }

Hide additional examples

crates/competitive/src/tree/tree_hash.rs (line 47)

44
45
46
47
48
49

    pub fn with_seed(seed: u64) -> Self {
        Self {
            rv: Vec::new(),
            rng: Xorshift::new_with_seed(seed),
        }
    }

crates/competitive/src/heuristic/simurated_annealing.rs (line 30)

19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37

    fn default() -> Self {
        let now = std::time::Instant::now();
        let log_table = (0..Self::LOG_TABLE_SIZE)
            .map(|i| ((i * 2 + 1) as f64 / (Self::LOG_TABLE_SIZE * 2) as f64).ln())
            .collect();
        Self {
            iter_count: 0,
            now,
            time: 0.,
            temperture: 3e3,
            log_table,
            rand: Xorshift::new_with_seed(Self::SEED),
            is_maximize: true,
            start_temp: 3e3,
            end_temp: 1e-8,
            time_limit: 1.99,
            update_interval: 0xff,
        }
    }

pub fn new() -> Self

Examples found in repository?

crates/competitive/src/string/rolling_hash.rs (line 14)

13
14
15
16

    fn init(len: usize) {
        let mut rng = Xorshift::new();
        Self::init_with_rng(len, &mut rng);
    }

Hide additional examples

crates/competitive/src/tree/tree_hash.rs (line 41)

38
39
40
41
42
43

    pub fn new() -> Self {
        Self {
            rv: Vec::new(),
            rng: Xorshift::new(),
        }
    }

crates/competitive/src/math/discrete_logarithm.rs (line 283)

271
272
273
274
275
276
277
278
279
280
281
282
283
284
285
286
287
288
289
290
291
292
293
294
295
296
297
298
299
300
301
302
303
304
305
306
307
308
309
310
311
312

    fn index_calculus(&self, a: u64, br_primes: &[BarrettReduction<u64>]) -> Option<u64> {
        let p = self.p;
        let ord = self.ord;
        let br = BarrettReduction::<u128>::new(p as u128);
        let a = br.rem(a as _) as u64;
        if a == 1 {
            return Some(0);
        }
        if p == 2 {
            return None;
        }

        let mut rng = Xorshift::new();
        let mut row = vec![0u64; br_primes.len()];
        let mut kk = rng.rand(ord - 1) + 1;
        let mut gkk = self.prec.pow(kk, &br);
        let mut k = kk;
        let mut gk = br.rem(gkk as u128 * a as u128) as u64;
        loop {
            if factorize_smooth(gk, &mut row, br_primes) {
                let mut res = ord - k;
                for (&c, &r) in self.coeff.iter().zip(&row) {
                    for _ in 0..r {
                        res += c;
                        if res >= ord {
                            res -= ord;
                        }
                    }
                }
                return Some(res);
            }
            if k + kk < ord {
                k += kk;
                gk = br.rem(gk as u128 * gkk as u128) as u64;
            } else {
                kk = rng.rand(ord - 1) + 1;
                gkk = self.prec.pow(kk, &br);
                k = kk;
                gk = br.rem(gkk as u128 * a as u128) as u64;
            }
        }
    }

pub fn rand64(&mut self) -> u64

Examples found in repository?

crates/competitive/src/tools/xorshift.rs (line 29)

28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43

    pub fn rand(&mut self, k: u64) -> u64 {
        self.rand64() % k
    }

    pub fn rands(&mut self, k: u64, n: usize) -> Vec<u64> {
        repeat_with(|| self.rand(k)).take(n).collect()
    }

    pub fn randf(&mut self) -> f64 {
        const UPPER_MASK: u64 = 0x3FF0_0000_0000_0000;
        const LOWER_MASK: u64 = 0x000F_FFFF_FFFF_FFFF;
        let x = self.rand64();
        let tmp = UPPER_MASK | (x & LOWER_MASK);
        let result: f64 = f64::from_bits(tmp);
        f64::from_bits(f64::to_bits(result - 1.0) ^ x >> 63)
    }

Hide additional examples

crates/competitive/src/tools/random_generator.rs (line 70)

69
70
71
72
73
74
75
76
77
78
79
80
81
82
83
84
85
86
87
88
89
90
91
92
93
94
95
96
97
98
99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122
123
124
125
126
127
128
129
130
131
132
133
134
135
136
137
138
139
140
141
142
143
144
145
146
147
148
149
150
151
152
153
154
155
156
157
158
159
160
161
162
163
164
165
166
167
168
169
170
171
172
173
174
175
176
177
178
179
180
181
182
183
184
185
186
187
188
189
190
191
192
193
194
195
196
197
198
199
200
201
202
203
204
205
206
207
208
209
210
211
212
213
214
215
216
217
218
219
220
221
222
223
224
225
226
227
228
229
230
231
232
233
234
235
236
237
238
239
240
241
242
243
244
245
246
247
248
249
250
251
252
253
254
255
256
257
258
259
260
261
262
263
264
265
266
267
268
269
270

    fn rand(&self, rng: &mut Xorshift) -> u128 {
        (rng.rand64() as u128) << 64 | rng.rand64() as u128
    }
}
impl RandomSpec<i128> for RangeFull {
    fn rand(&self, rng: &mut Xorshift) -> i128 {
        rng.gen::<u128, _>(..) as i128
    }
}

macro_rules! impl_random_spec_ranges {
    ($($u:ident $i:ident)*) => {
        $(
            impl RandomSpec<$u> for Range<$u> {
                fn rand(&self, rng: &mut Xorshift) -> $u {
                    assert!(self.start < self.end);
                    let len = self.end - self.start;
                    (self.start + rng.gen::<$u, _>(..) % len)
                }
            }
            impl RandomSpec<$i> for Range<$i> {
                fn rand(&self, rng: &mut Xorshift) -> $i {
                    assert!(self.start < self.end);
                    let len = self.end.abs_diff(self.start);
                    self.start.wrapping_add_unsigned(rng.gen::<$u, _>(..) % len)
                }
            }
            impl RandomSpec<$u> for RangeFrom<$u> {
                fn rand(&self, rng: &mut Xorshift) -> $u {
                    let len = ($u::MAX - self.start).wrapping_add(1);
                    let x = rng.gen::<$u, _>(..);
                    self.start + if len != 0 { x % len } else { x }
                }
            }
            impl RandomSpec<$i> for RangeFrom<$i> {
                fn rand(&self, rng: &mut Xorshift) -> $i {
                    let len = ($i::MAX.abs_diff(self.start)).wrapping_add(1);
                    let x = rng.gen::<$u, _>(..);
                    self.start.wrapping_add_unsigned(if len != 0 { x % len } else { x })
                }
            }
            impl RandomSpec<$u> for RangeInclusive<$u> {
                fn rand(&self, rng: &mut Xorshift) -> $u {
                    assert!(self.start() <= self.end());
                    let len = (self.end() - self.start()).wrapping_add(1);
                    let x = rng.gen::<$u, _>(..);
                    self.start() + if len != 0 { x % len } else { x }
                }
            }
            impl RandomSpec<$i> for RangeInclusive<$i> {
                fn rand(&self, rng: &mut Xorshift) -> $i {
                    assert!(self.start() <= self.end());
                    let len = (self.end().abs_diff(*self.start())).wrapping_add(1);
                    let x = rng.gen::<$u, _>(..);
                    self.start().wrapping_add_unsigned(if len != 0 { x % len } else { x })
                }
            }
            impl RandomSpec<$u> for RangeTo<$u> {
                fn rand(&self, rng: &mut Xorshift) -> $u {
                    let len = self.end;
                    rng.gen::<$u, _>(..) % len
                }
            }
            impl RandomSpec<$i> for RangeTo<$i> {
                fn rand(&self, rng: &mut Xorshift) -> $i {
                    let len = self.end.abs_diff($i::MIN);
                    $i::MIN.wrapping_add_unsigned(rng.gen::<$u, _>(..) % len)
                }
            }
            impl RandomSpec<$u> for RangeToInclusive<$u> {
                fn rand(&self, rng: &mut Xorshift) -> $u {
                    let len = (self.end).wrapping_add(1);
                    let x = rng.gen::<$u, _>(..);
                    if len != 0 { x % len } else { x }
                }
            }
            impl RandomSpec<$i> for RangeToInclusive<$i> {
                fn rand(&self, rng: &mut Xorshift) -> $i {
                    let len = (self.end.abs_diff($i::MIN)).wrapping_add(1);
                    let x = rng.gen::<$u, _>(..);
                    $i::MIN.wrapping_add_unsigned(if len != 0 { x % len } else { x })
                }
            }
        )*
    };
}
impl_random_spec_ranges!(u8 i8 u16 i16 u32 i32 u64 i64 u128 i128 usize isize);

macro_rules! impl_random_spec_tuple {
    ($($T:ident)*, $($R:ident)*, $($v:ident)*) => {
        impl<$($T),*, $($R),*> RandomSpec<($($T,)*)> for ($($R,)*)
        where
            $($R: RandomSpec<$T>),*
        {
            fn rand(&self, rng: &mut Xorshift) -> ($($T,)*) {
                let ($($v,)*) = self;
                ($(($v).rand(rng),)*)
            }
        }
    };
}
impl_random_spec_tuple!(A, RA, a);
impl_random_spec_tuple!(A B, RA RB, a b);
impl_random_spec_tuple!(A B C, RA RB RC, a b c);
impl_random_spec_tuple!(A B C D, RA RB RC RD, a b c d);
impl_random_spec_tuple!(A B C D E, RA RB RC RD RE, a b c d e);
impl_random_spec_tuple!(A B C D E F, RA RB RC RD RE RF, a b c d e f);
impl_random_spec_tuple!(A B C D E F G, RA RB RC RD RE RF RG, a b c d e f g);
impl_random_spec_tuple!(A B C D E F G H, RA RB RC RD RE RF RG RH, a b c d e f g h);
impl_random_spec_tuple!(A B C D E F G H I, RA RB RC RD RE RF RG RH RI, a b c d e f g h i);
impl_random_spec_tuple!(A B C D E F G H I J, RA RB RC RD RE RF RG RH RI RJ, a b c d e f g h i j);

macro_rules! impl_random_spec_primitive {
    ($($t:ty)*) => {
        $(impl RandomSpec<$t> for $t {
            fn rand(&self, _rng: &mut Xorshift) -> $t {
                *self
            }
        })*
    };
}
impl_random_spec_primitive!(() u8 u16 u32 u64 u128 usize i8 i16 i32 i64 i128 isize bool char);

impl<T, R> RandomSpec<T> for &R
where
    R: RandomSpec<T>,
{
    fn rand(&self, rng: &mut Xorshift) -> T {
        <R as RandomSpec<T>>::rand(self, rng)
    }
}
impl<T, R> RandomSpec<T> for &mut R
where
    R: RandomSpec<T>,
{
    fn rand(&self, rng: &mut Xorshift) -> T {
        <R as RandomSpec<T>>::rand(self, rng)
    }
}

#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
/// Left-close Right-open No Empty Segment
pub struct NotEmptySegment<T>(pub T);
impl<T> RandomSpec<(usize, usize)> for NotEmptySegment<T>
where
    T: RandomSpec<usize>,
{
    fn rand(&self, rng: &mut Xorshift) -> (usize, usize) {
        let n = rng.gen(&self.0) as u64;
        let k = randint_uniform(rng, n);
        let l = randint_uniform(rng, n - k) as usize;
        (l, l + k as usize + 1)
    }
}

#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
pub struct RandRange<Q, T> {
    data: Q,
    _marker: PhantomData<fn() -> T>,
}
impl<Q, T> RandRange<Q, T> {
    pub fn new(data: Q) -> Self {
        Self {
            data,
            _marker: PhantomData,
        }
    }
}
impl<Q, T> RandomSpec<(Bound<T>, Bound<T>)> for RandRange<Q, T>
where
    Q: RandomSpec<T>,
    T: Ord,
{
    fn rand(&self, rng: &mut Xorshift) -> (Bound<T>, Bound<T>) {
        let mut l = rng.gen(&self.data);
        let mut r = rng.gen(&self.data);
        if l > r {
            swap(&mut l, &mut r);
        }
        (
            match rng.rand(3) {
                0 => Bound::Excluded(l),
                1 => Bound::Included(l),
                _ => Bound::Unbounded,
            },
            match rng.rand(3) {
                0 => Bound::Excluded(r),
                1 => Bound::Included(r),
                _ => Bound::Unbounded,
            },
        )
    }
}

#[inline]
fn randint_uniform(rng: &mut Xorshift, k: u64) -> u64 {
    let mut v = rng.rand64();
    if k > 0 {
        v %= k;
    }
    v
}

crates/competitive/src/tree/tree_hash.rs (line 64)

61
62
63
64
65
66
67
68
69
70
71
72
73
74
75
76

    fn hash_rec(&mut self, g: &UndirectedSparseGraph, u: usize, p: usize, d: usize) -> u64 {
        let mut s = 1u64;
        if self.rv.len() <= d {
            self.rv.push(Self::mersenne_mod(self.rng.rand64()));
        }
        for a in g.adjacencies(u) {
            if a.to != p {
                s = Self::mersenne_mul_mod(s, self.hash_rec(g, a.to, u, d + 1));
            }
        }
        s += self.rv[d];
        if s >= Self::MOD {
            s -= Self::MOD;
        }
        s
    }

pub fn rand(&mut self, k: u64) -> u64

Examples found in repository?

crates/competitive/src/tools/xorshift.rs (line 33)

32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56

    pub fn rands(&mut self, k: u64, n: usize) -> Vec<u64> {
        repeat_with(|| self.rand(k)).take(n).collect()
    }

    pub fn randf(&mut self) -> f64 {
        const UPPER_MASK: u64 = 0x3FF0_0000_0000_0000;
        const LOWER_MASK: u64 = 0x000F_FFFF_FFFF_FFFF;
        let x = self.rand64();
        let tmp = UPPER_MASK | (x & LOWER_MASK);
        let result: f64 = f64::from_bits(tmp);
        f64::from_bits(f64::to_bits(result - 1.0) ^ x >> 63)
    }

    pub fn gen_bool(&mut self, p: f64) -> bool {
        self.randf() < p
    }

    pub fn shuffle<T>(&mut self, slice: &mut [T]) {
        let mut n = slice.len();
        while n > 1 {
            let i = self.rand(n as _) as usize;
            n -= 1;
            slice.swap(i, n);
        }
    }

Hide additional examples

crates/competitive/src/heuristic/simurated_annealing.rs (line 77)

69
70
71
72
73
74
75
76
77
78
79

    pub fn is_accepted(&mut self, current_score: f64, next_score: f64) -> bool {
        let diff = if self.is_maximize {
            next_score - current_score
        } else {
            current_score - next_score
        };
        diff >= 0.
            || diff
                > self.log_table[self.rand.rand(Self::LOG_TABLE_SIZE as u64) as usize]
                    * self.temperture
    }

crates/competitive/src/tools/random_generator.rs (line 249)

242
243
244
245
246
247
248
249
250
251
252
253
254
255
256
257
258
259
260

    fn rand(&self, rng: &mut Xorshift) -> (Bound<T>, Bound<T>) {
        let mut l = rng.gen(&self.data);
        let mut r = rng.gen(&self.data);
        if l > r {
            swap(&mut l, &mut r);
        }
        (
            match rng.rand(3) {
                0 => Bound::Excluded(l),
                1 => Bound::Included(l),
                _ => Bound::Unbounded,
            },
            match rng.rand(3) {
                0 => Bound::Excluded(r),
                1 => Bound::Included(r),
                _ => Bound::Unbounded,
            },
        )
    }

crates/competitive/src/tree/generator.rs (line 48)

43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68

        fn rand_inner(n: usize, rng: &mut Xorshift) -> Vec<(usize, usize)> {
            let mut edges = Vec::with_capacity(n.saturating_sub(1));
            if n >= 2 {
                let k = rng.gen(1..n);
                for n in [k, n - k].iter().cloned() {
                    let ty = rng.rand(6);
                    edges.extend(match ty {
                        0 => from_prufer_sequence(
                            n,
                            &rng.gen_iter(0..n)
                                .take(n.saturating_sub(2))
                                .collect::<Vec<usize>>(),
                        ),
                        1 => (1..n).map(|u| (u - 1, u)).collect(),
                        2 => (1..n).map(|u| (0, u)).collect(),
                        _ => rand_inner(n, rng),
                    });
                }
                for (u, v) in edges[k - 1..].iter_mut() {
                    *u += k;
                    *v += k;
                }
                edges.push((rng.gen(0..k), rng.gen(k..n)));
            }
            edges
        }

crates/competitive/src/math/discrete_logarithm.rs (line 202)

182
183
184
185
186
187
188
189
190
191
192
193
194
195
196
197
198
199
200
201
202
203
204
205
206
207
208
209
210
211
212
213
214
215
216
217
218
219
220
221
222
223
224
225
226
227
228
229
230
231
232
233
234
235
236
237
238
239
240
241
242
243
244
245
246
247
248
249
250
251
252
253
254
255
256
257
258
259
260
261
262
263
264
265
266
267
268
269
270
271
272
273
274
275
276
277
278
279
280
281
282
283
284
285
286
287
288
289
290
291
292
293
294
295
296
297
298
299
300
301
302
303
304
305
306
307
308
309
310
311
312

fn index_calculus_for_primitive_root(
    p: u64,
    ord: u64,
    br_primes: &[BarrettReduction<u64>],
    prec: &QdrtPowPrec,
) -> Vec<u64> {
    let br_ord = BarrettReduction::<u128>::new(ord as u128);
    let mul = |x: u64, y: u64| br_ord.rem(x as u128 * y as u128) as u64;
    let sub = |x: u64, y: u64| if x < y { x + ord - y } else { x - y };

    let pc = br_primes.len();
    let mut mat: Vec<Vec<u64>> = vec![];
    let mut rows: Vec<Vec<u64>> = vec![];

    let mut rng = Xorshift::default();
    let br = BarrettReduction::<u128>::new(p as u128);

    for i in 0..pc {
        for ri in 0usize.. {
            let mut row = vec![0u64; pc + 1];
            let mut kk = rng.rand(ord - 1) + 1;
            let mut gkk = prec.pow(kk, &br);
            let mut k = kk;
            let mut gk = gkk;
            while ri >= rows.len() {
                row[pc] = k;
                if factorize_smooth(gk, &mut row, br_primes) {
                    rows.push(row);
                    break;
                }
                if k + kk < ord {
                    k += kk;
                    gk = br.rem(gk as u128 * gkk as u128) as u64;
                } else {
                    kk = rng.rand(ord - 1) + 1;
                    gkk = prec.pow(kk, &br);
                    k = kk;
                    gk = gkk;
                }
            }
            let row = &mut rows[ri];
            for j in 0..i {
                if row[j] != 0 {
                    let b = mul(modinv(mat[j][j], ord), row[j]);
                    for (r, a) in row[j..].iter_mut().zip(&mat[j][j..]) {
                        *r = sub(*r, mul(*a, b));
                    }
                }
                assert_eq!(row[j], 0);
            }
            if gcd(row[i], ord) == 1 {
                let last = rows.len() - 1;
                rows.swap(ri, last);
                mat.push(rows.pop().unwrap());
                break;
            }
        }
    }
    for i in (0..pc).rev() {
        for j in i + 1..pc {
            mat[i][pc] = sub(mat[i][pc], mul(mat[i][j], mat[j][pc]));
        }
        mat[i][pc] = mul(mat[i][pc], modinv(mat[i][i], ord));
    }
    (0..pc).map(|i| (mat[i][pc])).collect()
}

#[derive(Debug)]
struct IndexCalculusWithPrimitiveRoot {
    p: u64,
    ord: u64,
    prec: QdrtPowPrec,
    coeff: Vec<u64>,
}

impl IndexCalculusWithPrimitiveRoot {
    fn new(p: u64, br_primes: &[BarrettReduction<u64>]) -> Self {
        let ord = p - 1;
        let g = primitive_root(p);
        let br = BarrettReduction::<u128>::new(p as u128);
        let prec = QdrtPowPrec::new(g, ord, &br);
        let coeff = index_calculus_for_primitive_root(p, ord, br_primes, &prec);
        Self {
            p,
            ord,
            prec,
            coeff,
        }
    }
    fn index_calculus(&self, a: u64, br_primes: &[BarrettReduction<u64>]) -> Option<u64> {
        let p = self.p;
        let ord = self.ord;
        let br = BarrettReduction::<u128>::new(p as u128);
        let a = br.rem(a as _) as u64;
        if a == 1 {
            return Some(0);
        }
        if p == 2 {
            return None;
        }

        let mut rng = Xorshift::new();
        let mut row = vec![0u64; br_primes.len()];
        let mut kk = rng.rand(ord - 1) + 1;
        let mut gkk = self.prec.pow(kk, &br);
        let mut k = kk;
        let mut gk = br.rem(gkk as u128 * a as u128) as u64;
        loop {
            if factorize_smooth(gk, &mut row, br_primes) {
                let mut res = ord - k;
                for (&c, &r) in self.coeff.iter().zip(&row) {
                    for _ in 0..r {
                        res += c;
                        if res >= ord {
                            res -= ord;
                        }
                    }
                }
                return Some(res);
            }
            if k + kk < ord {
                k += kk;
                gk = br.rem(gk as u128 * gkk as u128) as u64;
            } else {
                kk = rng.rand(ord - 1) + 1;
                gkk = self.prec.pow(kk, &br);
                k = kk;
                gk = br.rem(gkk as u128 * a as u128) as u64;
            }
        }
    }

pub fn rands(&mut self, k: u64, n: usize) -> Vec<u64>

pub fn randf(&mut self) -> f64

Examples found in repository?

crates/competitive/src/tools/xorshift.rs (line 46)

45
46
47

    pub fn gen_bool(&mut self, p: f64) -> bool {
        self.randf() < p
    }

pub fn gen_bool(&mut self, p: f64) -> bool

pub fn shuffle<T>(&mut self, slice: &mut [T])