competitive::tree::link_cut_tree

Struct LinkCutBstSpec

struct LinkCutBstSpec<S>(PhantomData<fn() -> S>);

Tuple Fields
0: PhantomData<fn() -> S>
Implementations

impl<S> LinkCutBstSpec<S>
where S: LinkCutTreeSpec,

unsafe fn toggle( node: NonNull<BstNode<LinkCutData<S>, WithParent<LinkCutData<S>>>>, )

Examples found in repository?

crates/competitive/src/tree/link_cut_tree.rs (line 104)

98    fn top_down(mut node: BstDataMutRef<'_, Self>) {
99        let pointer = node.node;
100        if node.reborrow().into_data().reverse {
101            node.data_mut().reverse = false;
102            let children = unsafe { pointer.as_ref().child };
103            for child in children.into_iter().flatten() {
104                unsafe { Self::toggle(child) };
105            }
106        }
107        let children = unsafe { pointer.as_ref().child };
108        let data = unsafe { &mut (*pointer.as_ptr()).data.inner };
109        let children =
110            children.map(|child| child.map(|child| unsafe { &mut (*child.as_ptr()).data.inner }));
111        S::top_down(data, children);
112    }
113
114    #[inline]
115    fn bottom_up(node: BstDataMutRef<'_, Self>) {
116        let pointer = node.node;
117        let children = unsafe { pointer.as_ref().child };
118        let data = unsafe { &mut (*pointer.as_ptr()).data.inner };
119        let children =
120            children.map(|child| child.map(|child| unsafe { &(*child.as_ptr()).data.inner }));
121        S::bottom_up(data, children);
122    }
123
124    fn merge(_left: Option<BstRoot<Self>>, _right: Option<BstRoot<Self>>) -> Option<BstRoot<Self>> {
125        unreachable!("link-cut trees do not merge auxiliary trees through BstSpec")
126    }
127
128    fn split<Seeker>(
129        _node: Option<BstRoot<Self>>,
130        _seeker: Seeker,
131        _equal_side: EqualSide,
132    ) -> (Option<BstRoot<Self>>, Option<BstRoot<Self>>)
133    where
134        Seeker: BstSeeker<Spec = Self>,
135    {
136        unreachable!("link-cut trees do not split auxiliary trees through BstSpec")
137    }
138}
139
140/// A link-cut forest with stable insertion-order node identifiers.
141///
142/// Its dynamic-tree operations take amortized `O(log n)` time when the spec
143/// hooks take constant time.
144pub struct LinkCutTree<S>
145where
146    S: LinkCutTreeSpec,
147{
148    nodes: Vec<LinkCutPtr<S>>,
149    allocator: MemoryPool<LinkCutNode<S>>,
150}
151
152impl<S> LinkCutTree<S>
153where
154    S: LinkCutTreeSpec,
155{
156    pub fn with_capacity(capacity: usize) -> Self {
157        Self {
158            nodes: Vec::with_capacity(capacity),
159            allocator: MemoryPool::with_capacity(capacity),
160        }
161    }
162
163    /// `edges` must form a tree over the values in iteration order.
164    pub fn from_edges<T>(values: T, edges: &[(usize, usize)]) -> Self
165    where
166        T: IntoIterator<Item = S::Value>,
167    {
168        let tree: Self = values.into_iter().collect();
169        for (child, parent, preferred) in
170            splay_operations::rooted_heavy_order(tree.nodes.len(), edges)
171                .into_iter()
172                .rev()
173        {
174            let child = tree.node(child);
175            let mut parent = tree.node(parent);
176            unsafe {
177                (*child.as_ptr()).parent.parent = Some(parent);
178                if preferred {
179                    parent.as_mut().child[1] = Some(child);
180                } else {
181                    LinkCutBstSpec::<S>::with_two_inner_mut(parent, child, S::attach_virtual);
182                }
183                Self::pull(parent);
184            }
185        }
186        tree
187    }
188
189    pub fn add_node(&mut self, value: S::Value) -> usize {
190        let index = self.nodes.len();
191        let node = self.allocator.allocate(BstNode::new(LinkCutData {
192            inner: S::new(value),
193            index,
194            reverse: false,
195        }));
196        self.nodes.push(node);
197        index
198    }
199
200    #[inline]
201    fn node(&self, index: usize) -> LinkCutPtr<S> {
202        self.nodes[index]
203    }
204
205    #[inline]
206    unsafe fn pull(node: LinkCutPtr<S>) {
207        unsafe {
208            LinkCutBstSpec::<S>::bottom_up(BstDataMutRef::new_unchecked(node));
209        }
210    }
211
212    #[inline]
213    unsafe fn splay(node: LinkCutPtr<S>) {
214        let root = if S::ROOT_TO_NODE_TOP_DOWN {
215            unsafe {
216                splay_operations::with_parent::splay::<LinkCutBstSpec<S>, LinkCutData<S>>(node)
217            }
218        } else {
219            unsafe {
220                splay_operations::with_parent::splay_with_local_top_down::<
221                    LinkCutBstSpec<S>,
222                    LinkCutData<S>,
223                >(node)
224            }
225        };
226        if root != node {
227            unsafe {
228                LinkCutBstSpec::<S>::with_two_inner_mut(root, node, S::transfer_path_parent);
229            }
230        }
231    }
232
233    fn access_node(mut node: LinkCutPtr<S>) {
234        unsafe {
235            Self::splay(node);
236            if let Some(right) = node.as_mut().child[1].take() {
237                LinkCutBstSpec::<S>::with_two_inner_mut(node, right, S::attach_virtual);
238            }
239            Self::pull(node);
240            while let Some(mut parent) = node.as_ref().parent.parent {
241                Self::splay(parent);
242                if let Some(right) = parent.as_mut().child[1].take() {
243                    LinkCutBstSpec::<S>::with_two_inner_mut(parent, right, S::attach_virtual);
244                }
245                LinkCutBstSpec::<S>::with_two_inner_mut(parent, node, S::detach_virtual);
246                parent.as_mut().child[1] = Some(node);
247                node.as_mut().parent.parent = Some(parent);
248                Self::pull(parent);
249                Self::splay(node);
250            }
251        }
252    }
253
254    pub fn get(&mut self, node: usize) -> &S::Value {
255        let node = self.node(node);
256        Self::access_node(node);
257        unsafe { S::value(&node.as_ref().data.inner) }
258    }
259
260    pub fn set(&mut self, node: usize, value: S::Value) {
261        let node = self.node(node);
262        Self::access_node(node);
263        unsafe {
264            *S::value_mut(&mut (*node.as_ptr()).data.inner) = value;
265            Self::pull(node);
266        }
267    }
268
269    pub fn modify<F>(&mut self, node: usize, f: F)
270    where
271        F: FnOnce(&S::Value) -> S::Value,
272    {
273        let node = self.node(node);
274        Self::access_node(node);
275        unsafe {
276            let data = &mut (*node.as_ptr()).data.inner;
277            *S::value_mut(data) = f(S::value(data));
278            Self::pull(node);
279        }
280    }
281
282    pub fn reroot(&mut self, node: usize) {
283        let node = self.node(node);
284        Self::access_node(node);
285        unsafe { LinkCutBstSpec::<S>::toggle(node) };
286    }

unsafe fn with_two_inner_mut<R>( left: NonNull<BstNode<LinkCutData<S>, WithParent<LinkCutData<S>>>>, right: NonNull<BstNode<LinkCutData<S>, WithParent<LinkCutData<S>>>>, f: impl FnOnce(&mut S::Data, &mut S::Data) -> R, ) -> R

Examples found in repository?

crates/competitive/src/tree/link_cut_tree.rs (line 181)

164    pub fn from_edges<T>(values: T, edges: &[(usize, usize)]) -> Self
165    where
166        T: IntoIterator<Item = S::Value>,
167    {
168        let tree: Self = values.into_iter().collect();
169        for (child, parent, preferred) in
170            splay_operations::rooted_heavy_order(tree.nodes.len(), edges)
171                .into_iter()
172                .rev()
173        {
174            let child = tree.node(child);
175            let mut parent = tree.node(parent);
176            unsafe {
177                (*child.as_ptr()).parent.parent = Some(parent);
178                if preferred {
179                    parent.as_mut().child[1] = Some(child);
180                } else {
181                    LinkCutBstSpec::<S>::with_two_inner_mut(parent, child, S::attach_virtual);
182                }
183                Self::pull(parent);
184            }
185        }
186        tree
187    }
188
189    pub fn add_node(&mut self, value: S::Value) -> usize {
190        let index = self.nodes.len();
191        let node = self.allocator.allocate(BstNode::new(LinkCutData {
192            inner: S::new(value),
193            index,
194            reverse: false,
195        }));
196        self.nodes.push(node);
197        index
198    }
199
200    #[inline]
201    fn node(&self, index: usize) -> LinkCutPtr<S> {
202        self.nodes[index]
203    }
204
205    #[inline]
206    unsafe fn pull(node: LinkCutPtr<S>) {
207        unsafe {
208            LinkCutBstSpec::<S>::bottom_up(BstDataMutRef::new_unchecked(node));
209        }
210    }
211
212    #[inline]
213    unsafe fn splay(node: LinkCutPtr<S>) {
214        let root = if S::ROOT_TO_NODE_TOP_DOWN {
215            unsafe {
216                splay_operations::with_parent::splay::<LinkCutBstSpec<S>, LinkCutData<S>>(node)
217            }
218        } else {
219            unsafe {
220                splay_operations::with_parent::splay_with_local_top_down::<
221                    LinkCutBstSpec<S>,
222                    LinkCutData<S>,
223                >(node)
224            }
225        };
226        if root != node {
227            unsafe {
228                LinkCutBstSpec::<S>::with_two_inner_mut(root, node, S::transfer_path_parent);
229            }
230        }
231    }
232
233    fn access_node(mut node: LinkCutPtr<S>) {
234        unsafe {
235            Self::splay(node);
236            if let Some(right) = node.as_mut().child[1].take() {
237                LinkCutBstSpec::<S>::with_two_inner_mut(node, right, S::attach_virtual);
238            }
239            Self::pull(node);
240            while let Some(mut parent) = node.as_ref().parent.parent {
241                Self::splay(parent);
242                if let Some(right) = parent.as_mut().child[1].take() {
243                    LinkCutBstSpec::<S>::with_two_inner_mut(parent, right, S::attach_virtual);
244                }
245                LinkCutBstSpec::<S>::with_two_inner_mut(parent, node, S::detach_virtual);
246                parent.as_mut().child[1] = Some(node);
247                node.as_mut().parent.parent = Some(parent);
248                Self::pull(parent);
249                Self::splay(node);
250            }
251        }
252    }
253
254    pub fn get(&mut self, node: usize) -> &S::Value {
255        let node = self.node(node);
256        Self::access_node(node);
257        unsafe { S::value(&node.as_ref().data.inner) }
258    }
259
260    pub fn set(&mut self, node: usize, value: S::Value) {
261        let node = self.node(node);
262        Self::access_node(node);
263        unsafe {
264            *S::value_mut(&mut (*node.as_ptr()).data.inner) = value;
265            Self::pull(node);
266        }
267    }
268
269    pub fn modify<F>(&mut self, node: usize, f: F)
270    where
271        F: FnOnce(&S::Value) -> S::Value,
272    {
273        let node = self.node(node);
274        Self::access_node(node);
275        unsafe {
276            let data = &mut (*node.as_ptr()).data.inner;
277            *S::value_mut(data) = f(S::value(data));
278            Self::pull(node);
279        }
280    }
281
282    pub fn reroot(&mut self, node: usize) {
283        let node = self.node(node);
284        Self::access_node(node);
285        unsafe { LinkCutBstSpec::<S>::toggle(node) };
286    }
287
288    /// `child` and `parent` must belong to different trees.
289    pub fn link(&mut self, child: usize, parent: usize) {
290        assert_ne!(child, parent);
291        self.reroot(child);
292        let child = self.node(child);
293        let parent = self.node(parent);
294        Self::access_node(parent);
295        unsafe {
296            (*child.as_ptr()).parent.parent = Some(parent);
297            LinkCutBstSpec::<S>::with_two_inner_mut(parent, child, S::attach_virtual);
298            Self::pull(parent);
299        }
300    }