competitive::tree::top_tree

Struct TopBstSpec

struct TopBstSpec<S, A>(PhantomData<fn() -> (S, A)>);

Tuple Fields
0: PhantomData<fn() -> (S, A)>
Implementations

impl<S, A> TopBstSpec<S, A>
where S: TopTreeSpec, A: TopTreeAction<S>,

fn is_unit(action: &A::Action) -> bool

Examples found in repository?

crates/competitive/src/tree/top_tree.rs (line 132)

131    unsafe fn apply_all(mut node: TopPtr<S, A>, action: &A::Action) {
132        if Self::is_unit(action) {
133            return;
134        }
135        let data = unsafe { &mut node.as_mut().data };
136        Self::compose(&mut data.heavy_action, action);
137        Self::compose(&mut data.light_action, action);
138        A::act_info(&mut data.info, action);
139        A::act_path(&mut data.sum, action);
140        A::act_path_light(&mut data.sum, action);
141    }
142}
143
144impl<S, A> BstSpec for TopBstSpec<S, A>
145where
146    S: TopTreeSpec,
147    A: TopTreeAction<S>,
148{
149    type Parent = WithParent<Self::Data>;
150    type Data = TopTreeData<S, A>;
151
152    #[inline]
153    fn top_down(mut node: BstDataMutRef<'_, Self>) {
154        let pointer = node.node;
155        if node.reborrow().into_data().reverse {
156            node.data_mut().reverse = false;
157            for child in unsafe { pointer.as_ref().child }.into_iter().flatten() {
158                unsafe { Self::toggle(child) };
159            }
160        }
161
162        if !Self::is_unit(&node.reborrow().into_data().heavy_action) {
163            let action = replace(
164                &mut node.data_mut().heavy_action,
165                <A::ActionMonoid as Unital>::unit(),
166            );
167            for child in unsafe { pointer.as_ref().child }.into_iter().flatten() {
168                unsafe { Self::apply_heavy(child, &action) };
169            }
170        }
171        if !Self::is_unit(&node.reborrow().into_data().light_action) {
172            let action = replace(
173                &mut node.data_mut().light_action,
174                <A::ActionMonoid as Unital>::unit(),
175            );
176            for child in unsafe { pointer.as_ref().child }.into_iter().flatten() {
177                unsafe { Self::apply_light(child, &action) };
178            }
179            if let Some(light) = node.reborrow().into_data().light {
180                unsafe { RakeBstSpec::<S, A>::apply(light, &action) };
181            }
182        }
183    }
184
185    #[inline]
186    fn bottom_up(node: BstDataMutRef<'_, Self>) {
187        let pointer = node.node;
188        let data = unsafe { &mut (*pointer.as_ptr()).data };
189        let mut sum = if let Some(light) = data.light {
190            S::add_vertex(unsafe { &light.as_ref().data.sum }, &data.info)
191        } else {
192            S::vertex(&data.info)
193        };
194        if let Some(left) = unsafe { pointer.as_ref().child[0] } {
195            sum = S::compress(unsafe { &left.as_ref().data.sum }, &sum);
196        }
197        if let Some(right) = unsafe { pointer.as_ref().child[1] } {
198            sum = S::compress(&sum, unsafe { &right.as_ref().data.sum });
199        }
200        data.sum = sum;
201    }
202
203    fn merge(_left: Option<BstRoot<Self>>, _right: Option<BstRoot<Self>>) -> Option<BstRoot<Self>> {
204        unreachable!("top trees do not merge auxiliary trees through BstSpec")
205    }
206
207    fn split<Seeker>(
208        _node: Option<BstRoot<Self>>,
209        _seeker: Seeker,
210        _equal_side: EqualSide,
211    ) -> (Option<BstRoot<Self>>, Option<BstRoot<Self>>)
212    where
213        Seeker: BstSeeker<Spec = Self>,
214    {
215        unreachable!("top trees do not split auxiliary trees through BstSpec")
216    }
217}
218
219impl<S, A> RakeBstSpec<S, A>
220where
221    S: TopTreeSpec,
222    A: TopTreeAction<S>,
223{
224    #[inline]
225    unsafe fn apply(mut node: RakePtr<S, A>, action: &A::Action) {
226        let data = unsafe { &mut node.as_mut().data };
227        A::act_point(&mut data.key, action);
228        A::act_point(&mut data.sum, action);
229        TopBstSpec::<S, A>::compose(&mut data.action, action);
230        TopBstSpec::<S, A>::compose(&mut data.buffer, action);
231    }
232}
233
234impl<S, A> BstSpec for RakeBstSpec<S, A>
235where
236    S: TopTreeSpec,
237    A: TopTreeAction<S>,
238{
239    type Parent = WithParent<Self::Data>;
240    type Data = RakeData<S, A>;
241
242    #[inline]
243    fn top_down(mut node: BstDataMutRef<'_, Self>) {
244        if TopBstSpec::<S, A>::is_unit(&node.reborrow().into_data().action) {
245            return;
246        }
247        let pointer = node.node;
248        let action = replace(
249            &mut node.data_mut().action,
250            <A::ActionMonoid as Unital>::unit(),
251        );
252        for child in unsafe { pointer.as_ref().child }.into_iter().flatten() {
253            unsafe { Self::apply(child, &action) };
254        }
255    }
256
257    #[inline]
258    fn bottom_up(node: BstDataMutRef<'_, Self>) {
259        let pointer = node.node;
260        let data = unsafe { &mut (*pointer.as_ptr()).data };
261        let mut sum = data.key.clone();
262        if let Some(left) = unsafe { pointer.as_ref().child[0] } {
263            sum = S::rake(&sum, unsafe { &left.as_ref().data.sum });
264        }
265        if let Some(right) = unsafe { pointer.as_ref().child[1] } {
266            sum = S::rake(&sum, unsafe { &right.as_ref().data.sum });
267        }
268        data.sum = sum;
269    }
270
271    fn merge(_left: Option<BstRoot<Self>>, _right: Option<BstRoot<Self>>) -> Option<BstRoot<Self>> {
272        unreachable!("rake trees use their dedicated merge operation")
273    }
274
275    fn split<Seeker>(
276        _node: Option<BstRoot<Self>>,
277        _seeker: Seeker,
278        _equal_side: EqualSide,
279    ) -> (Option<BstRoot<Self>>, Option<BstRoot<Self>>)
280    where
281        Seeker: BstSeeker<Spec = Self>,
282    {
283        unreachable!("rake trees do not use BstSpec::split")
284    }
285}
286
287/// A self-adjusting top tree, also called a strong link-cut tree.
288///
289/// This is not a classical worst-case-balanced top tree. Circular order and
290/// `select` are not supported.
291pub struct TopTree<S, A = NoTopTreeAction>
292where
293    S: TopTreeSpec,
294    A: TopTreeAction<S>,
295{
296    nodes: Vec<TopPtr<S, A>>,
297    node_allocator: MemoryPool<TopNode<S, A>>,
298    rake_allocator: MemoryPool<RakeNode<S, A>>,
299}
300
301impl<S, A> TopTree<S, A>
302where
303    S: TopTreeSpec,
304    A: TopTreeAction<S>,
305{
306    pub fn with_capacity(capacity: usize) -> Self {
307        Self {
308            nodes: Vec::with_capacity(capacity),
309            node_allocator: MemoryPool::with_capacity(capacity),
310            rake_allocator: MemoryPool::with_capacity(capacity),
311        }
312    }
313
314    pub fn add_node(&mut self, info: S::Info) -> usize {
315        let index = self.nodes.len();
316        let sum = S::vertex(&info);
317        let node = self.node_allocator.allocate(BstNode::new(TopTreeData {
318            info,
319            sum,
320            reverse: false,
321            light: None,
322            belong: None,
323            heavy_action: <A::ActionMonoid as Unital>::unit(),
324            light_action: <A::ActionMonoid as Unital>::unit(),
325            index,
326        }));
327        self.nodes.push(node);
328        index
329    }
330
331    fn node(&self, index: usize) -> TopPtr<S, A> {
332        self.nodes[index]
333    }
334
335    #[inline]
336    unsafe fn pull_top(node: TopPtr<S, A>) {
337        unsafe { TopBstSpec::<S, A>::bottom_up(BstDataMutRef::new_unchecked(node)) };
338    }
339
340    #[inline]
341    unsafe fn pull_rake(node: RakePtr<S, A>) {
342        unsafe { RakeBstSpec::<S, A>::bottom_up(BstDataMutRef::new_unchecked(node)) };
343    }
344
345    #[inline]
346    unsafe fn splay_top(node: TopPtr<S, A>) {
347        let root = unsafe {
348            splay_operations::with_parent::splay::<TopBstSpec<S, A>, TopTreeData<S, A>>(node)
349        };
350        if root != node {
351            unsafe {
352                (*node.as_ptr()).data.belong = (*root.as_ptr()).data.belong.take();
353            }
354        }
355    }
356
357    #[inline]
358    unsafe fn splay_rake(node: RakePtr<S, A>) {
359        unsafe {
360            splay_operations::with_parent::splay_with_local_top_down::<
361                RakeBstSpec<S, A>,
362                RakeData<S, A>,
363            >(node)
364        };
365    }
366
367    unsafe fn rake_rightmost(mut node: RakePtr<S, A>) -> RakePtr<S, A> {
368        loop {
369            unsafe { RakeBstSpec::<S, A>::top_down(BstDataMutRef::new_unchecked(node)) };
370            match unsafe { node.as_ref().child[1] } {
371                Some(right) => node = right,
372                None => return node,
373            }
374        }
375    }
376
377    unsafe fn rake_insert(
378        &mut self,
379        root: Option<RakePtr<S, A>>,
380        key: S::Point,
381    ) -> (RakePtr<S, A>, RakePtr<S, A>) {
382        let node = self.rake_allocator.allocate(BstNode::new(RakeData {
383            sum: key.clone(),
384            key,
385            action: <A::ActionMonoid as Unital>::unit(),
386            buffer: <A::ActionMonoid as Unital>::unit(),
387        }));
388        if let Some(root) = root {
389            let mut rightmost = unsafe { Self::rake_rightmost(root) };
390            unsafe {
391                Self::splay_rake(rightmost);
392                rightmost.as_mut().child[1] = Some(node);
393                (*node.as_ptr()).parent.parent = Some(rightmost);
394                Self::pull_rake(rightmost);
395            }
396            (rightmost, node)
397        } else {
398            (node, node)
399        }
400    }
401
402    unsafe fn rake_remove(
403        &mut self,
404        mut node: RakePtr<S, A>,
405    ) -> (Option<RakePtr<S, A>>, A::Action) {
406        unsafe {
407            Self::splay_rake(node);
408            RakeBstSpec::<S, A>::top_down(BstDataMutRef::new_unchecked(node));
409        }
410        let left = unsafe { node.as_mut().child[0].take() };
411        let right = unsafe { node.as_mut().child[1].take() };
412        for mut child in [left, right].into_iter().flatten() {
413            unsafe { child.as_mut().parent.parent = None };
414        }
415        let root = match (left, right) {
416            (None, right) => right,
417            (left, None) => left,
418            (Some(left), Some(right)) => {
419                let mut root = unsafe { Self::rake_rightmost(left) };
420                unsafe {
421                    Self::splay_rake(root);
422                    root.as_mut().child[1] = Some(right);
423                    (*right.as_ptr()).parent.parent = Some(root);
424                    Self::pull_rake(root);
425                }
426                Some(root)
427            }
428        };
429        let node = self.rake_allocator.deallocate(node);
430        (root, node.data.buffer)
431    }
432
433    fn access_node(&mut self, node: TopPtr<S, A>) {
434        unsafe {
435            let mut previous: Option<TopPtr<S, A>> = None;
436            let mut current = Some(node);
437            while let Some(mut cursor) = current {
438                Self::splay_top(cursor);
439                let next = cursor.as_ref().parent.parent;
440                if let Some(right) = cursor.as_mut().child[1].take() {
441                    let point = S::add_edge(&right.as_ref().data.sum);
442                    let (light, entry) = self.rake_insert(cursor.as_ref().data.light, point);
443                    cursor.as_mut().data.light = Some(light);
444                    (*right.as_ptr()).data.belong = Some(entry);
445                }
446                if let Some(previous) = previous {
447                    let entry = (*previous.as_ptr())
448                        .data
449                        .belong
450                        .take()
451                        .expect("a virtual path must have a rake-tree entry");
452                    let (light, action) = self.rake_remove(entry);
453                    cursor.as_mut().data.light = light;
454                    TopBstSpec::<S, A>::apply_all(previous, &action);
455                    cursor.as_mut().child[1] = Some(previous);
456                    (*previous.as_ptr()).parent.parent = Some(cursor);
457                }
458                Self::pull_top(cursor);
459                previous = Some(cursor);
460                current = next;
461            }
462            Self::splay_top(node);
463        }
464    }
465
466    pub fn get(&mut self, node: usize) -> &S::Info {
467        let node = self.node(node);
468        self.access_node(node);
469        unsafe { &node.as_ref().data.info }
470    }
471
472    pub fn set(&mut self, node: usize, info: S::Info) {
473        let mut node = self.node(node);
474        self.access_node(node);
475        unsafe {
476            node.as_mut().data.info = info;
477            Self::pull_top(node);
478        }
479    }
480
481    pub fn reroot(&mut self, node: usize) {
482        let node = self.node(node);
483        self.access_node(node);
484        unsafe { TopBstSpec::<S, A>::toggle(node) };
485    }
486
487    /// `child` and `parent` must belong to different trees.
488    pub fn link(&mut self, child: usize, parent: usize) {
489        assert_ne!(child, parent);
490        self.reroot(child);
491        let child = self.node(child);
492        let mut parent = self.node(parent);
493        self.access_node(parent);
494        unsafe {
495            (*child.as_ptr()).parent.parent = Some(parent);
496            let point = S::add_edge(&child.as_ref().data.sum);
497            let (light, entry) = self.rake_insert(parent.as_ref().data.light, point);
498            parent.as_mut().data.light = Some(light);
499            (*child.as_ptr()).data.belong = Some(entry);
500            Self::pull_top(parent);
501        }
502    }
503
504    /// `(u, v)` must be an edge.
505    pub fn cut(&mut self, u: usize, v: usize) {
506        assert_ne!(u, v);
507        self.reroot(u);
508        let mut v = self.node(v);
509        self.access_node(v);
510        unsafe {
511            let mut left = v.as_mut().child[0]
512                .take()
513                .expect("the specified edge must exist");
514            left.as_mut().parent.parent = None;
515            Self::pull_top(v);
516        }
517    }
518
519    pub fn root(&mut self, node: usize) -> usize {
520        let mut root = self.node(node);
521        self.access_node(root);
522        unsafe {
523            loop {
524                TopBstSpec::<S, A>::top_down(BstDataMutRef::new_unchecked(root));
525                match root.as_ref().child[0] {
526                    Some(left) => root = left,
527                    None => break,
528                }
529            }
530            Self::splay_top(root);
531            root.as_ref().data.index
532        }
533    }
534
535    pub fn is_connected(&mut self, u: usize, v: usize) -> bool {
536        self.root(u) == self.root(v)
537    }
538
539    /// `u` and `v` must be connected.
540    pub fn fold_path(&mut self, u: usize, v: usize) -> S::Path {
541        self.reroot(u);
542        let v = self.node(v);
543        self.access_node(v);
544        unsafe { v.as_ref().data.sum.clone() }
545    }
546
547    /// `u` and `v` must be connected.
548    pub fn update_path(&mut self, u: usize, v: usize, action: &A::Action) {
549        self.reroot(u);
550        let v = self.node(v);
551        self.access_node(v);
552        if !TopBstSpec::<S, A>::is_unit(action) {
553            unsafe { TopBstSpec::<S, A>::apply_heavy(v, action) };
554        }
555    }

fn compose(target: &mut A::Action, action: &A::Action)

Examples found in repository?

crates/competitive/src/tree/top_tree.rs (line 118)

116    unsafe fn apply_heavy(mut node: TopPtr<S, A>, action: &A::Action) {
117        let data = unsafe { &mut node.as_mut().data };
118        Self::compose(&mut data.heavy_action, action);
119        A::act_info(&mut data.info, action);
120        A::act_path(&mut data.sum, action);
121    }
122
123    #[inline]
124    unsafe fn apply_light(mut node: TopPtr<S, A>, action: &A::Action) {
125        let data = unsafe { &mut node.as_mut().data };
126        Self::compose(&mut data.light_action, action);
127        A::act_path_light(&mut data.sum, action);
128    }
129
130    #[inline]
131    unsafe fn apply_all(mut node: TopPtr<S, A>, action: &A::Action) {
132        if Self::is_unit(action) {
133            return;
134        }
135        let data = unsafe { &mut node.as_mut().data };
136        Self::compose(&mut data.heavy_action, action);
137        Self::compose(&mut data.light_action, action);
138        A::act_info(&mut data.info, action);
139        A::act_path(&mut data.sum, action);
140        A::act_path_light(&mut data.sum, action);
141    }
142}
143
144impl<S, A> BstSpec for TopBstSpec<S, A>
145where
146    S: TopTreeSpec,
147    A: TopTreeAction<S>,
148{
149    type Parent = WithParent<Self::Data>;
150    type Data = TopTreeData<S, A>;
151
152    #[inline]
153    fn top_down(mut node: BstDataMutRef<'_, Self>) {
154        let pointer = node.node;
155        if node.reborrow().into_data().reverse {
156            node.data_mut().reverse = false;
157            for child in unsafe { pointer.as_ref().child }.into_iter().flatten() {
158                unsafe { Self::toggle(child) };
159            }
160        }
161
162        if !Self::is_unit(&node.reborrow().into_data().heavy_action) {
163            let action = replace(
164                &mut node.data_mut().heavy_action,
165                <A::ActionMonoid as Unital>::unit(),
166            );
167            for child in unsafe { pointer.as_ref().child }.into_iter().flatten() {
168                unsafe { Self::apply_heavy(child, &action) };
169            }
170        }
171        if !Self::is_unit(&node.reborrow().into_data().light_action) {
172            let action = replace(
173                &mut node.data_mut().light_action,
174                <A::ActionMonoid as Unital>::unit(),
175            );
176            for child in unsafe { pointer.as_ref().child }.into_iter().flatten() {
177                unsafe { Self::apply_light(child, &action) };
178            }
179            if let Some(light) = node.reborrow().into_data().light {
180                unsafe { RakeBstSpec::<S, A>::apply(light, &action) };
181            }
182        }
183    }
184
185    #[inline]
186    fn bottom_up(node: BstDataMutRef<'_, Self>) {
187        let pointer = node.node;
188        let data = unsafe { &mut (*pointer.as_ptr()).data };
189        let mut sum = if let Some(light) = data.light {
190            S::add_vertex(unsafe { &light.as_ref().data.sum }, &data.info)
191        } else {
192            S::vertex(&data.info)
193        };
194        if let Some(left) = unsafe { pointer.as_ref().child[0] } {
195            sum = S::compress(unsafe { &left.as_ref().data.sum }, &sum);
196        }
197        if let Some(right) = unsafe { pointer.as_ref().child[1] } {
198            sum = S::compress(&sum, unsafe { &right.as_ref().data.sum });
199        }
200        data.sum = sum;
201    }
202
203    fn merge(_left: Option<BstRoot<Self>>, _right: Option<BstRoot<Self>>) -> Option<BstRoot<Self>> {
204        unreachable!("top trees do not merge auxiliary trees through BstSpec")
205    }
206
207    fn split<Seeker>(
208        _node: Option<BstRoot<Self>>,
209        _seeker: Seeker,
210        _equal_side: EqualSide,
211    ) -> (Option<BstRoot<Self>>, Option<BstRoot<Self>>)
212    where
213        Seeker: BstSeeker<Spec = Self>,
214    {
215        unreachable!("top trees do not split auxiliary trees through BstSpec")
216    }
217}
218
219impl<S, A> RakeBstSpec<S, A>
220where
221    S: TopTreeSpec,
222    A: TopTreeAction<S>,
223{
224    #[inline]
225    unsafe fn apply(mut node: RakePtr<S, A>, action: &A::Action) {
226        let data = unsafe { &mut node.as_mut().data };
227        A::act_point(&mut data.key, action);
228        A::act_point(&mut data.sum, action);
229        TopBstSpec::<S, A>::compose(&mut data.action, action);
230        TopBstSpec::<S, A>::compose(&mut data.buffer, action);
231    }

unsafe fn toggle( node: NonNull<BstNode<TopTreeData<S, A>, WithParent<TopTreeData<S, A>>>>, )

Examples found in repository?

crates/competitive/src/tree/top_tree.rs (line 158)

153    fn top_down(mut node: BstDataMutRef<'_, Self>) {
154        let pointer = node.node;
155        if node.reborrow().into_data().reverse {
156            node.data_mut().reverse = false;
157            for child in unsafe { pointer.as_ref().child }.into_iter().flatten() {
158                unsafe { Self::toggle(child) };
159            }
160        }
161
162        if !Self::is_unit(&node.reborrow().into_data().heavy_action) {
163            let action = replace(
164                &mut node.data_mut().heavy_action,
165                <A::ActionMonoid as Unital>::unit(),
166            );
167            for child in unsafe { pointer.as_ref().child }.into_iter().flatten() {
168                unsafe { Self::apply_heavy(child, &action) };
169            }
170        }
171        if !Self::is_unit(&node.reborrow().into_data().light_action) {
172            let action = replace(
173                &mut node.data_mut().light_action,
174                <A::ActionMonoid as Unital>::unit(),
175            );
176            for child in unsafe { pointer.as_ref().child }.into_iter().flatten() {
177                unsafe { Self::apply_light(child, &action) };
178            }
179            if let Some(light) = node.reborrow().into_data().light {
180                unsafe { RakeBstSpec::<S, A>::apply(light, &action) };
181            }
182        }
183    }
184
185    #[inline]
186    fn bottom_up(node: BstDataMutRef<'_, Self>) {
187        let pointer = node.node;
188        let data = unsafe { &mut (*pointer.as_ptr()).data };
189        let mut sum = if let Some(light) = data.light {
190            S::add_vertex(unsafe { &light.as_ref().data.sum }, &data.info)
191        } else {
192            S::vertex(&data.info)
193        };
194        if let Some(left) = unsafe { pointer.as_ref().child[0] } {
195            sum = S::compress(unsafe { &left.as_ref().data.sum }, &sum);
196        }
197        if let Some(right) = unsafe { pointer.as_ref().child[1] } {
198            sum = S::compress(&sum, unsafe { &right.as_ref().data.sum });
199        }
200        data.sum = sum;
201    }
202
203    fn merge(_left: Option<BstRoot<Self>>, _right: Option<BstRoot<Self>>) -> Option<BstRoot<Self>> {
204        unreachable!("top trees do not merge auxiliary trees through BstSpec")
205    }
206
207    fn split<Seeker>(
208        _node: Option<BstRoot<Self>>,
209        _seeker: Seeker,
210        _equal_side: EqualSide,
211    ) -> (Option<BstRoot<Self>>, Option<BstRoot<Self>>)
212    where
213        Seeker: BstSeeker<Spec = Self>,
214    {
215        unreachable!("top trees do not split auxiliary trees through BstSpec")
216    }
217}
218
219impl<S, A> RakeBstSpec<S, A>
220where
221    S: TopTreeSpec,
222    A: TopTreeAction<S>,
223{
224    #[inline]
225    unsafe fn apply(mut node: RakePtr<S, A>, action: &A::Action) {
226        let data = unsafe { &mut node.as_mut().data };
227        A::act_point(&mut data.key, action);
228        A::act_point(&mut data.sum, action);
229        TopBstSpec::<S, A>::compose(&mut data.action, action);
230        TopBstSpec::<S, A>::compose(&mut data.buffer, action);
231    }
232}
233
234impl<S, A> BstSpec for RakeBstSpec<S, A>
235where
236    S: TopTreeSpec,
237    A: TopTreeAction<S>,
238{
239    type Parent = WithParent<Self::Data>;
240    type Data = RakeData<S, A>;
241
242    #[inline]
243    fn top_down(mut node: BstDataMutRef<'_, Self>) {
244        if TopBstSpec::<S, A>::is_unit(&node.reborrow().into_data().action) {
245            return;
246        }
247        let pointer = node.node;
248        let action = replace(
249            &mut node.data_mut().action,
250            <A::ActionMonoid as Unital>::unit(),
251        );
252        for child in unsafe { pointer.as_ref().child }.into_iter().flatten() {
253            unsafe { Self::apply(child, &action) };
254        }
255    }
256
257    #[inline]
258    fn bottom_up(node: BstDataMutRef<'_, Self>) {
259        let pointer = node.node;
260        let data = unsafe { &mut (*pointer.as_ptr()).data };
261        let mut sum = data.key.clone();
262        if let Some(left) = unsafe { pointer.as_ref().child[0] } {
263            sum = S::rake(&sum, unsafe { &left.as_ref().data.sum });
264        }
265        if let Some(right) = unsafe { pointer.as_ref().child[1] } {
266            sum = S::rake(&sum, unsafe { &right.as_ref().data.sum });
267        }
268        data.sum = sum;
269    }
270
271    fn merge(_left: Option<BstRoot<Self>>, _right: Option<BstRoot<Self>>) -> Option<BstRoot<Self>> {
272        unreachable!("rake trees use their dedicated merge operation")
273    }
274
275    fn split<Seeker>(
276        _node: Option<BstRoot<Self>>,
277        _seeker: Seeker,
278        _equal_side: EqualSide,
279    ) -> (Option<BstRoot<Self>>, Option<BstRoot<Self>>)
280    where
281        Seeker: BstSeeker<Spec = Self>,
282    {
283        unreachable!("rake trees do not use BstSpec::split")
284    }
285}
286
287/// A self-adjusting top tree, also called a strong link-cut tree.
288///
289/// This is not a classical worst-case-balanced top tree. Circular order and
290/// `select` are not supported.
291pub struct TopTree<S, A = NoTopTreeAction>
292where
293    S: TopTreeSpec,
294    A: TopTreeAction<S>,
295{
296    nodes: Vec<TopPtr<S, A>>,
297    node_allocator: MemoryPool<TopNode<S, A>>,
298    rake_allocator: MemoryPool<RakeNode<S, A>>,
299}
300
301impl<S, A> TopTree<S, A>
302where
303    S: TopTreeSpec,
304    A: TopTreeAction<S>,
305{
306    pub fn with_capacity(capacity: usize) -> Self {
307        Self {
308            nodes: Vec::with_capacity(capacity),
309            node_allocator: MemoryPool::with_capacity(capacity),
310            rake_allocator: MemoryPool::with_capacity(capacity),
311        }
312    }
313
314    pub fn add_node(&mut self, info: S::Info) -> usize {
315        let index = self.nodes.len();
316        let sum = S::vertex(&info);
317        let node = self.node_allocator.allocate(BstNode::new(TopTreeData {
318            info,
319            sum,
320            reverse: false,
321            light: None,
322            belong: None,
323            heavy_action: <A::ActionMonoid as Unital>::unit(),
324            light_action: <A::ActionMonoid as Unital>::unit(),
325            index,
326        }));
327        self.nodes.push(node);
328        index
329    }
330
331    fn node(&self, index: usize) -> TopPtr<S, A> {
332        self.nodes[index]
333    }
334
335    #[inline]
336    unsafe fn pull_top(node: TopPtr<S, A>) {
337        unsafe { TopBstSpec::<S, A>::bottom_up(BstDataMutRef::new_unchecked(node)) };
338    }
339
340    #[inline]
341    unsafe fn pull_rake(node: RakePtr<S, A>) {
342        unsafe { RakeBstSpec::<S, A>::bottom_up(BstDataMutRef::new_unchecked(node)) };
343    }
344
345    #[inline]
346    unsafe fn splay_top(node: TopPtr<S, A>) {
347        let root = unsafe {
348            splay_operations::with_parent::splay::<TopBstSpec<S, A>, TopTreeData<S, A>>(node)
349        };
350        if root != node {
351            unsafe {
352                (*node.as_ptr()).data.belong = (*root.as_ptr()).data.belong.take();
353            }
354        }
355    }
356
357    #[inline]
358    unsafe fn splay_rake(node: RakePtr<S, A>) {
359        unsafe {
360            splay_operations::with_parent::splay_with_local_top_down::<
361                RakeBstSpec<S, A>,
362                RakeData<S, A>,
363            >(node)
364        };
365    }
366
367    unsafe fn rake_rightmost(mut node: RakePtr<S, A>) -> RakePtr<S, A> {
368        loop {
369            unsafe { RakeBstSpec::<S, A>::top_down(BstDataMutRef::new_unchecked(node)) };
370            match unsafe { node.as_ref().child[1] } {
371                Some(right) => node = right,
372                None => return node,
373            }
374        }
375    }
376
377    unsafe fn rake_insert(
378        &mut self,
379        root: Option<RakePtr<S, A>>,
380        key: S::Point,
381    ) -> (RakePtr<S, A>, RakePtr<S, A>) {
382        let node = self.rake_allocator.allocate(BstNode::new(RakeData {
383            sum: key.clone(),
384            key,
385            action: <A::ActionMonoid as Unital>::unit(),
386            buffer: <A::ActionMonoid as Unital>::unit(),
387        }));
388        if let Some(root) = root {
389            let mut rightmost = unsafe { Self::rake_rightmost(root) };
390            unsafe {
391                Self::splay_rake(rightmost);
392                rightmost.as_mut().child[1] = Some(node);
393                (*node.as_ptr()).parent.parent = Some(rightmost);
394                Self::pull_rake(rightmost);
395            }
396            (rightmost, node)
397        } else {
398            (node, node)
399        }
400    }
401
402    unsafe fn rake_remove(
403        &mut self,
404        mut node: RakePtr<S, A>,
405    ) -> (Option<RakePtr<S, A>>, A::Action) {
406        unsafe {
407            Self::splay_rake(node);
408            RakeBstSpec::<S, A>::top_down(BstDataMutRef::new_unchecked(node));
409        }
410        let left = unsafe { node.as_mut().child[0].take() };
411        let right = unsafe { node.as_mut().child[1].take() };
412        for mut child in [left, right].into_iter().flatten() {
413            unsafe { child.as_mut().parent.parent = None };
414        }
415        let root = match (left, right) {
416            (None, right) => right,
417            (left, None) => left,
418            (Some(left), Some(right)) => {
419                let mut root = unsafe { Self::rake_rightmost(left) };
420                unsafe {
421                    Self::splay_rake(root);
422                    root.as_mut().child[1] = Some(right);
423                    (*right.as_ptr()).parent.parent = Some(root);
424                    Self::pull_rake(root);
425                }
426                Some(root)
427            }
428        };
429        let node = self.rake_allocator.deallocate(node);
430        (root, node.data.buffer)
431    }
432
433    fn access_node(&mut self, node: TopPtr<S, A>) {
434        unsafe {
435            let mut previous: Option<TopPtr<S, A>> = None;
436            let mut current = Some(node);
437            while let Some(mut cursor) = current {
438                Self::splay_top(cursor);
439                let next = cursor.as_ref().parent.parent;
440                if let Some(right) = cursor.as_mut().child[1].take() {
441                    let point = S::add_edge(&right.as_ref().data.sum);
442                    let (light, entry) = self.rake_insert(cursor.as_ref().data.light, point);
443                    cursor.as_mut().data.light = Some(light);
444                    (*right.as_ptr()).data.belong = Some(entry);
445                }
446                if let Some(previous) = previous {
447                    let entry = (*previous.as_ptr())
448                        .data
449                        .belong
450                        .take()
451                        .expect("a virtual path must have a rake-tree entry");
452                    let (light, action) = self.rake_remove(entry);
453                    cursor.as_mut().data.light = light;
454                    TopBstSpec::<S, A>::apply_all(previous, &action);
455                    cursor.as_mut().child[1] = Some(previous);
456                    (*previous.as_ptr()).parent.parent = Some(cursor);
457                }
458                Self::pull_top(cursor);
459                previous = Some(cursor);
460                current = next;
461            }
462            Self::splay_top(node);
463        }
464    }
465
466    pub fn get(&mut self, node: usize) -> &S::Info {
467        let node = self.node(node);
468        self.access_node(node);
469        unsafe { &node.as_ref().data.info }
470    }
471
472    pub fn set(&mut self, node: usize, info: S::Info) {
473        let mut node = self.node(node);
474        self.access_node(node);
475        unsafe {
476            node.as_mut().data.info = info;
477            Self::pull_top(node);
478        }
479    }
480
481    pub fn reroot(&mut self, node: usize) {
482        let node = self.node(node);
483        self.access_node(node);
484        unsafe { TopBstSpec::<S, A>::toggle(node) };
485    }

unsafe fn apply_heavy( node: NonNull<BstNode<TopTreeData<S, A>, WithParent<TopTreeData<S, A>>>>, action: &A::Action, )

Examples found in repository?

crates/competitive/src/tree/top_tree.rs (line 168)

153    fn top_down(mut node: BstDataMutRef<'_, Self>) {
154        let pointer = node.node;
155        if node.reborrow().into_data().reverse {
156            node.data_mut().reverse = false;
157            for child in unsafe { pointer.as_ref().child }.into_iter().flatten() {
158                unsafe { Self::toggle(child) };
159            }
160        }
161
162        if !Self::is_unit(&node.reborrow().into_data().heavy_action) {
163            let action = replace(
164                &mut node.data_mut().heavy_action,
165                <A::ActionMonoid as Unital>::unit(),
166            );
167            for child in unsafe { pointer.as_ref().child }.into_iter().flatten() {
168                unsafe { Self::apply_heavy(child, &action) };
169            }
170        }
171        if !Self::is_unit(&node.reborrow().into_data().light_action) {
172            let action = replace(
173                &mut node.data_mut().light_action,
174                <A::ActionMonoid as Unital>::unit(),
175            );
176            for child in unsafe { pointer.as_ref().child }.into_iter().flatten() {
177                unsafe { Self::apply_light(child, &action) };
178            }
179            if let Some(light) = node.reborrow().into_data().light {
180                unsafe { RakeBstSpec::<S, A>::apply(light, &action) };
181            }
182        }
183    }
184
185    #[inline]
186    fn bottom_up(node: BstDataMutRef<'_, Self>) {
187        let pointer = node.node;
188        let data = unsafe { &mut (*pointer.as_ptr()).data };
189        let mut sum = if let Some(light) = data.light {
190            S::add_vertex(unsafe { &light.as_ref().data.sum }, &data.info)
191        } else {
192            S::vertex(&data.info)
193        };
194        if let Some(left) = unsafe { pointer.as_ref().child[0] } {
195            sum = S::compress(unsafe { &left.as_ref().data.sum }, &sum);
196        }
197        if let Some(right) = unsafe { pointer.as_ref().child[1] } {
198            sum = S::compress(&sum, unsafe { &right.as_ref().data.sum });
199        }
200        data.sum = sum;
201    }
202
203    fn merge(_left: Option<BstRoot<Self>>, _right: Option<BstRoot<Self>>) -> Option<BstRoot<Self>> {
204        unreachable!("top trees do not merge auxiliary trees through BstSpec")
205    }
206
207    fn split<Seeker>(
208        _node: Option<BstRoot<Self>>,
209        _seeker: Seeker,
210        _equal_side: EqualSide,
211    ) -> (Option<BstRoot<Self>>, Option<BstRoot<Self>>)
212    where
213        Seeker: BstSeeker<Spec = Self>,
214    {
215        unreachable!("top trees do not split auxiliary trees through BstSpec")
216    }
217}
218
219impl<S, A> RakeBstSpec<S, A>
220where
221    S: TopTreeSpec,
222    A: TopTreeAction<S>,
223{
224    #[inline]
225    unsafe fn apply(mut node: RakePtr<S, A>, action: &A::Action) {
226        let data = unsafe { &mut node.as_mut().data };
227        A::act_point(&mut data.key, action);
228        A::act_point(&mut data.sum, action);
229        TopBstSpec::<S, A>::compose(&mut data.action, action);
230        TopBstSpec::<S, A>::compose(&mut data.buffer, action);
231    }
232}
233
234impl<S, A> BstSpec for RakeBstSpec<S, A>
235where
236    S: TopTreeSpec,
237    A: TopTreeAction<S>,
238{
239    type Parent = WithParent<Self::Data>;
240    type Data = RakeData<S, A>;
241
242    #[inline]
243    fn top_down(mut node: BstDataMutRef<'_, Self>) {
244        if TopBstSpec::<S, A>::is_unit(&node.reborrow().into_data().action) {
245            return;
246        }
247        let pointer = node.node;
248        let action = replace(
249            &mut node.data_mut().action,
250            <A::ActionMonoid as Unital>::unit(),
251        );
252        for child in unsafe { pointer.as_ref().child }.into_iter().flatten() {
253            unsafe { Self::apply(child, &action) };
254        }
255    }
256
257    #[inline]
258    fn bottom_up(node: BstDataMutRef<'_, Self>) {
259        let pointer = node.node;
260        let data = unsafe { &mut (*pointer.as_ptr()).data };
261        let mut sum = data.key.clone();
262        if let Some(left) = unsafe { pointer.as_ref().child[0] } {
263            sum = S::rake(&sum, unsafe { &left.as_ref().data.sum });
264        }
265        if let Some(right) = unsafe { pointer.as_ref().child[1] } {
266            sum = S::rake(&sum, unsafe { &right.as_ref().data.sum });
267        }
268        data.sum = sum;
269    }
270
271    fn merge(_left: Option<BstRoot<Self>>, _right: Option<BstRoot<Self>>) -> Option<BstRoot<Self>> {
272        unreachable!("rake trees use their dedicated merge operation")
273    }
274
275    fn split<Seeker>(
276        _node: Option<BstRoot<Self>>,
277        _seeker: Seeker,
278        _equal_side: EqualSide,
279    ) -> (Option<BstRoot<Self>>, Option<BstRoot<Self>>)
280    where
281        Seeker: BstSeeker<Spec = Self>,
282    {
283        unreachable!("rake trees do not use BstSpec::split")
284    }
285}
286
287/// A self-adjusting top tree, also called a strong link-cut tree.
288///
289/// This is not a classical worst-case-balanced top tree. Circular order and
290/// `select` are not supported.
291pub struct TopTree<S, A = NoTopTreeAction>
292where
293    S: TopTreeSpec,
294    A: TopTreeAction<S>,
295{
296    nodes: Vec<TopPtr<S, A>>,
297    node_allocator: MemoryPool<TopNode<S, A>>,
298    rake_allocator: MemoryPool<RakeNode<S, A>>,
299}
300
301impl<S, A> TopTree<S, A>
302where
303    S: TopTreeSpec,
304    A: TopTreeAction<S>,
305{
306    pub fn with_capacity(capacity: usize) -> Self {
307        Self {
308            nodes: Vec::with_capacity(capacity),
309            node_allocator: MemoryPool::with_capacity(capacity),
310            rake_allocator: MemoryPool::with_capacity(capacity),
311        }
312    }
313
314    pub fn add_node(&mut self, info: S::Info) -> usize {
315        let index = self.nodes.len();
316        let sum = S::vertex(&info);
317        let node = self.node_allocator.allocate(BstNode::new(TopTreeData {
318            info,
319            sum,
320            reverse: false,
321            light: None,
322            belong: None,
323            heavy_action: <A::ActionMonoid as Unital>::unit(),
324            light_action: <A::ActionMonoid as Unital>::unit(),
325            index,
326        }));
327        self.nodes.push(node);
328        index
329    }
330
331    fn node(&self, index: usize) -> TopPtr<S, A> {
332        self.nodes[index]
333    }
334
335    #[inline]
336    unsafe fn pull_top(node: TopPtr<S, A>) {
337        unsafe { TopBstSpec::<S, A>::bottom_up(BstDataMutRef::new_unchecked(node)) };
338    }
339
340    #[inline]
341    unsafe fn pull_rake(node: RakePtr<S, A>) {
342        unsafe { RakeBstSpec::<S, A>::bottom_up(BstDataMutRef::new_unchecked(node)) };
343    }
344
345    #[inline]
346    unsafe fn splay_top(node: TopPtr<S, A>) {
347        let root = unsafe {
348            splay_operations::with_parent::splay::<TopBstSpec<S, A>, TopTreeData<S, A>>(node)
349        };
350        if root != node {
351            unsafe {
352                (*node.as_ptr()).data.belong = (*root.as_ptr()).data.belong.take();
353            }
354        }
355    }
356
357    #[inline]
358    unsafe fn splay_rake(node: RakePtr<S, A>) {
359        unsafe {
360            splay_operations::with_parent::splay_with_local_top_down::<
361                RakeBstSpec<S, A>,
362                RakeData<S, A>,
363            >(node)
364        };
365    }
366
367    unsafe fn rake_rightmost(mut node: RakePtr<S, A>) -> RakePtr<S, A> {
368        loop {
369            unsafe { RakeBstSpec::<S, A>::top_down(BstDataMutRef::new_unchecked(node)) };
370            match unsafe { node.as_ref().child[1] } {
371                Some(right) => node = right,
372                None => return node,
373            }
374        }
375    }
376
377    unsafe fn rake_insert(
378        &mut self,
379        root: Option<RakePtr<S, A>>,
380        key: S::Point,
381    ) -> (RakePtr<S, A>, RakePtr<S, A>) {
382        let node = self.rake_allocator.allocate(BstNode::new(RakeData {
383            sum: key.clone(),
384            key,
385            action: <A::ActionMonoid as Unital>::unit(),
386            buffer: <A::ActionMonoid as Unital>::unit(),
387        }));
388        if let Some(root) = root {
389            let mut rightmost = unsafe { Self::rake_rightmost(root) };
390            unsafe {
391                Self::splay_rake(rightmost);
392                rightmost.as_mut().child[1] = Some(node);
393                (*node.as_ptr()).parent.parent = Some(rightmost);
394                Self::pull_rake(rightmost);
395            }
396            (rightmost, node)
397        } else {
398            (node, node)
399        }
400    }
401
402    unsafe fn rake_remove(
403        &mut self,
404        mut node: RakePtr<S, A>,
405    ) -> (Option<RakePtr<S, A>>, A::Action) {
406        unsafe {
407            Self::splay_rake(node);
408            RakeBstSpec::<S, A>::top_down(BstDataMutRef::new_unchecked(node));
409        }
410        let left = unsafe { node.as_mut().child[0].take() };
411        let right = unsafe { node.as_mut().child[1].take() };
412        for mut child in [left, right].into_iter().flatten() {
413            unsafe { child.as_mut().parent.parent = None };
414        }
415        let root = match (left, right) {
416            (None, right) => right,
417            (left, None) => left,
418            (Some(left), Some(right)) => {
419                let mut root = unsafe { Self::rake_rightmost(left) };
420                unsafe {
421                    Self::splay_rake(root);
422                    root.as_mut().child[1] = Some(right);
423                    (*right.as_ptr()).parent.parent = Some(root);
424                    Self::pull_rake(root);
425                }
426                Some(root)
427            }
428        };
429        let node = self.rake_allocator.deallocate(node);
430        (root, node.data.buffer)
431    }
432
433    fn access_node(&mut self, node: TopPtr<S, A>) {
434        unsafe {
435            let mut previous: Option<TopPtr<S, A>> = None;
436            let mut current = Some(node);
437            while let Some(mut cursor) = current {
438                Self::splay_top(cursor);
439                let next = cursor.as_ref().parent.parent;
440                if let Some(right) = cursor.as_mut().child[1].take() {
441                    let point = S::add_edge(&right.as_ref().data.sum);
442                    let (light, entry) = self.rake_insert(cursor.as_ref().data.light, point);
443                    cursor.as_mut().data.light = Some(light);
444                    (*right.as_ptr()).data.belong = Some(entry);
445                }
446                if let Some(previous) = previous {
447                    let entry = (*previous.as_ptr())
448                        .data
449                        .belong
450                        .take()
451                        .expect("a virtual path must have a rake-tree entry");
452                    let (light, action) = self.rake_remove(entry);
453                    cursor.as_mut().data.light = light;
454                    TopBstSpec::<S, A>::apply_all(previous, &action);
455                    cursor.as_mut().child[1] = Some(previous);
456                    (*previous.as_ptr()).parent.parent = Some(cursor);
457                }
458                Self::pull_top(cursor);
459                previous = Some(cursor);
460                current = next;
461            }
462            Self::splay_top(node);
463        }
464    }
465
466    pub fn get(&mut self, node: usize) -> &S::Info {
467        let node = self.node(node);
468        self.access_node(node);
469        unsafe { &node.as_ref().data.info }
470    }
471
472    pub fn set(&mut self, node: usize, info: S::Info) {
473        let mut node = self.node(node);
474        self.access_node(node);
475        unsafe {
476            node.as_mut().data.info = info;
477            Self::pull_top(node);
478        }
479    }
480
481    pub fn reroot(&mut self, node: usize) {
482        let node = self.node(node);
483        self.access_node(node);
484        unsafe { TopBstSpec::<S, A>::toggle(node) };
485    }
486
487    /// `child` and `parent` must belong to different trees.
488    pub fn link(&mut self, child: usize, parent: usize) {
489        assert_ne!(child, parent);
490        self.reroot(child);
491        let child = self.node(child);
492        let mut parent = self.node(parent);
493        self.access_node(parent);
494        unsafe {
495            (*child.as_ptr()).parent.parent = Some(parent);
496            let point = S::add_edge(&child.as_ref().data.sum);
497            let (light, entry) = self.rake_insert(parent.as_ref().data.light, point);
498            parent.as_mut().data.light = Some(light);
499            (*child.as_ptr()).data.belong = Some(entry);
500            Self::pull_top(parent);
501        }
502    }
503
504    /// `(u, v)` must be an edge.
505    pub fn cut(&mut self, u: usize, v: usize) {
506        assert_ne!(u, v);
507        self.reroot(u);
508        let mut v = self.node(v);
509        self.access_node(v);
510        unsafe {
511            let mut left = v.as_mut().child[0]
512                .take()
513                .expect("the specified edge must exist");
514            left.as_mut().parent.parent = None;
515            Self::pull_top(v);
516        }
517    }
518
519    pub fn root(&mut self, node: usize) -> usize {
520        let mut root = self.node(node);
521        self.access_node(root);
522        unsafe {
523            loop {
524                TopBstSpec::<S, A>::top_down(BstDataMutRef::new_unchecked(root));
525                match root.as_ref().child[0] {
526                    Some(left) => root = left,
527                    None => break,
528                }
529            }
530            Self::splay_top(root);
531            root.as_ref().data.index
532        }
533    }
534
535    pub fn is_connected(&mut self, u: usize, v: usize) -> bool {
536        self.root(u) == self.root(v)
537    }
538
539    /// `u` and `v` must be connected.
540    pub fn fold_path(&mut self, u: usize, v: usize) -> S::Path {
541        self.reroot(u);
542        let v = self.node(v);
543        self.access_node(v);
544        unsafe { v.as_ref().data.sum.clone() }
545    }
546
547    /// `u` and `v` must be connected.
548    pub fn update_path(&mut self, u: usize, v: usize, action: &A::Action) {
549        self.reroot(u);
550        let v = self.node(v);
551        self.access_node(v);
552        if !TopBstSpec::<S, A>::is_unit(action) {
553            unsafe { TopBstSpec::<S, A>::apply_heavy(v, action) };
554        }
555    }

unsafe fn apply_light( node: NonNull<BstNode<TopTreeData<S, A>, WithParent<TopTreeData<S, A>>>>, action: &A::Action, )

Examples found in repository?

crates/competitive/src/tree/top_tree.rs (line 177)

153    fn top_down(mut node: BstDataMutRef<'_, Self>) {
154        let pointer = node.node;
155        if node.reborrow().into_data().reverse {
156            node.data_mut().reverse = false;
157            for child in unsafe { pointer.as_ref().child }.into_iter().flatten() {
158                unsafe { Self::toggle(child) };
159            }
160        }
161
162        if !Self::is_unit(&node.reborrow().into_data().heavy_action) {
163            let action = replace(
164                &mut node.data_mut().heavy_action,
165                <A::ActionMonoid as Unital>::unit(),
166            );
167            for child in unsafe { pointer.as_ref().child }.into_iter().flatten() {
168                unsafe { Self::apply_heavy(child, &action) };
169            }
170        }
171        if !Self::is_unit(&node.reborrow().into_data().light_action) {
172            let action = replace(
173                &mut node.data_mut().light_action,
174                <A::ActionMonoid as Unital>::unit(),
175            );
176            for child in unsafe { pointer.as_ref().child }.into_iter().flatten() {
177                unsafe { Self::apply_light(child, &action) };
178            }
179            if let Some(light) = node.reborrow().into_data().light {
180                unsafe { RakeBstSpec::<S, A>::apply(light, &action) };
181            }
182        }
183    }

unsafe fn apply_all( node: NonNull<BstNode<TopTreeData<S, A>, WithParent<TopTreeData<S, A>>>>, action: &A::Action, )

Examples found in repository?

crates/competitive/src/tree/top_tree.rs (line 454)

433    fn access_node(&mut self, node: TopPtr<S, A>) {
434        unsafe {
435            let mut previous: Option<TopPtr<S, A>> = None;
436            let mut current = Some(node);
437            while let Some(mut cursor) = current {
438                Self::splay_top(cursor);
439                let next = cursor.as_ref().parent.parent;
440                if let Some(right) = cursor.as_mut().child[1].take() {
441                    let point = S::add_edge(&right.as_ref().data.sum);
442                    let (light, entry) = self.rake_insert(cursor.as_ref().data.light, point);
443                    cursor.as_mut().data.light = Some(light);
444                    (*right.as_ptr()).data.belong = Some(entry);
445                }
446                if let Some(previous) = previous {
447                    let entry = (*previous.as_ptr())
448                        .data
449                        .belong
450                        .take()
451                        .expect("a virtual path must have a rake-tree entry");
452                    let (light, action) = self.rake_remove(entry);
453                    cursor.as_mut().data.light = light;
454                    TopBstSpec::<S, A>::apply_all(previous, &action);
455                    cursor.as_mut().child[1] = Some(previous);
456                    (*previous.as_ptr()).parent.parent = Some(cursor);
457                }
458                Self::pull_top(cursor);
459                previous = Some(cursor);
460                current = next;
461            }
462            Self::splay_top(node);
463        }
464    }
465
466    pub fn get(&mut self, node: usize) -> &S::Info {
467        let node = self.node(node);
468        self.access_node(node);
469        unsafe { &node.as_ref().data.info }
470    }
471
472    pub fn set(&mut self, node: usize, info: S::Info) {
473        let mut node = self.node(node);
474        self.access_node(node);
475        unsafe {
476            node.as_mut().data.info = info;
477            Self::pull_top(node);
478        }
479    }
480
481    pub fn reroot(&mut self, node: usize) {
482        let node = self.node(node);
483        self.access_node(node);
484        unsafe { TopBstSpec::<S, A>::toggle(node) };
485    }
486
487    /// `child` and `parent` must belong to different trees.
488    pub fn link(&mut self, child: usize, parent: usize) {
489        assert_ne!(child, parent);
490        self.reroot(child);
491        let child = self.node(child);
492        let mut parent = self.node(parent);
493        self.access_node(parent);
494        unsafe {
495            (*child.as_ptr()).parent.parent = Some(parent);
496            let point = S::add_edge(&child.as_ref().data.sum);
497            let (light, entry) = self.rake_insert(parent.as_ref().data.light, point);
498            parent.as_mut().data.light = Some(light);
499            (*child.as_ptr()).data.belong = Some(entry);
500            Self::pull_top(parent);
501        }
502    }
503
504    /// `(u, v)` must be an edge.
505    pub fn cut(&mut self, u: usize, v: usize) {
506        assert_ne!(u, v);
507        self.reroot(u);
508        let mut v = self.node(v);
509        self.access_node(v);
510        unsafe {
511            let mut left = v.as_mut().child[0]
512                .take()
513                .expect("the specified edge must exist");
514            left.as_mut().parent.parent = None;
515            Self::pull_top(v);
516        }
517    }
518
519    pub fn root(&mut self, node: usize) -> usize {
520        let mut root = self.node(node);
521        self.access_node(root);
522        unsafe {
523            loop {
524                TopBstSpec::<S, A>::top_down(BstDataMutRef::new_unchecked(root));
525                match root.as_ref().child[0] {
526                    Some(left) => root = left,
527                    None => break,
528                }
529            }
530            Self::splay_top(root);
531            root.as_ref().data.index
532        }
533    }
534
535    pub fn is_connected(&mut self, u: usize, v: usize) -> bool {
536        self.root(u) == self.root(v)
537    }
538
539    /// `u` and `v` must be connected.
540    pub fn fold_path(&mut self, u: usize, v: usize) -> S::Path {
541        self.reroot(u);
542        let v = self.node(v);
543        self.access_node(v);
544        unsafe { v.as_ref().data.sum.clone() }
545    }
546
547    /// `u` and `v` must be connected.
548    pub fn update_path(&mut self, u: usize, v: usize, action: &A::Action) {
549        self.reroot(u);
550        let v = self.node(v);
551        self.access_node(v);
552        if !TopBstSpec::<S, A>::is_unit(action) {
553            unsafe { TopBstSpec::<S, A>::apply_heavy(v, action) };
554        }
555    }
556
557    fn detach_left<R>(mut node: TopPtr<S, A>, f: impl FnOnce(TopPtr<S, A>) -> R) -> R {
558        unsafe {
559            let left = node.as_mut().child[0].take();
560            if let Some(mut left) = left {
561                left.as_mut().parent.parent = None;
562            }
563            Self::pull_top(node);
564            let result = f(node);
565            node.as_mut().child[0] = left;
566            if let Some(mut left) = left {
567                left.as_mut().parent.parent = Some(node);
568            }
569            Self::pull_top(node);
570            result
571        }
572    }
573
574    /// `(node, parent)` must be an edge.
575    pub fn fold_subtree(&mut self, node: usize, parent: usize) -> S::Path {
576        self.reroot(parent);
577        let node = self.node(node);
578        self.access_node(node);
579        Self::detach_left(node, |node| unsafe { node.as_ref().data.sum.clone() })
580    }
581
582    /// `(node, parent)` must be an edge.
583    pub fn update_subtree(&mut self, node: usize, parent: usize, action: &A::Action) {
584        self.reroot(parent);
585        let node = self.node(node);
586        self.access_node(node);
587        Self::detach_left(node, |node| unsafe {
588            TopBstSpec::<S, A>::apply_all(node, action);
589            TopBstSpec::<S, A>::top_down(BstDataMutRef::new_unchecked(node));
590        });
591    }