Module competitive::algorithm

algorithm

Structs

BabyStepGiantStep

$\min{1\le i \le n | x^i=a}$

BitDp

Combinations

ConvexHullTrick

DivisorTransform

IndependentSubSet

MultipleTransform

PartisanGameAnalyzer

PartisanGamer

RhoPath

$P_0 =$ init, $P_{i+1} = next(P_i)$

SlideMinimum

SqrtDecompositionBuckets

SubsetTransform

Subsets

SupersetTransform

XorBasis

Basis of xor operation.

Traits

Bisect

binary search helper

MoSolver

PartisanGame

SliceBisectExt

binary search for slice

SqrtDecomposition

Trisect

ternary search helper

Functions

binary_search

binary search for monotone segment

combinations

choose distinct r elements from n in sorted order

combinations_with_replacement

choose r elements from n in sorted order

floor_kernel

$y = \left\lfloor\frac{n}{x}\right\rfloor$

parallel_binary_search

permutations

choose distinct r elements from n in any order

product

choose r elements from n independently

run_length_encoding

return: [(elem, length)]

slide_minimum

ternary_search

like (left..=right).min_by_key(f)