competitive
Module algorithm

algorithm

Structs

BitDp

Combinations

ConvexHullTrick

EsperEstimator

EsperSolver

ImpartialGameAnalyzer

ImpartialGamer

IndependentSubSet

PartisanGameAnalyzer

PartisanGamer

RhoPath

$P_0 =$ init, $P_{i+1} = next(P_i)$

SlideMinimum

SqrtDecompositionBuckets

Subsets

XorBasis

Basis of xor operation.

ZeroSumGameAnalyzer

ZeroSumGamer

Traits

Bisect

binary search helper

ImpartialGame

PartisanGame

SliceBisectExt

binary search for slice

SliceCombinationsExt

SliceSortExt

SqrtDecomposition

ZeroSumGame

Functions

baby_step_giant_step

$\min{0\le i < n | x^i=y}$

binary_search

binary search for monotone segment

floor_kernel

$y = \left\lfloor\frac{n}{x}\right\rfloor$

parallel_binary_search

run_length_encoding

return: [(elem, length)]

slide_minimum

solve_01_on_tree

ternary_search

Returns the element that gives the minimum value from the strictly concave up function and the minimum value.