Struct competitive::data_structure::LazySegmentTree

pub struct LazySegmentTree<M>where
    M: MonoidAction,{ /* private fields */ }

Implementations

impl<M> LazySegmentTree<M>where M: MonoidAction,

pub fn new(n: usize) -> Self

Examples found in repository?

crates/aizu_online_judge/src/dsl/dsl_2_d.rs (line 10)

6
7
8
9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24

pub fn dsl_2_d(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q);
    let mut seg = LazySegmentTree::<RangeMinRangeUpdate<_>>::new(n);
    for _ in 0..q {
        match scanner.scan::<usize>() {
            0 => {
                scan!(scanner, s, t, x: i32);
                seg.update(s, t + 1, Some(x));
            }
            1 => {
                scan!(scanner, i);
                writeln!(writer, "{}", seg.fold(i, i + 1)).ok();
            }
            _ => panic!("unknown query"),
        }
    }
}

Hide additional examples

crates/aizu_online_judge/src/dsl/dsl_2_f.rs (line 10)

6
7
8
9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24

pub fn dsl_2_f(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q);
    let mut seg = LazySegmentTree::<RangeMinRangeUpdate<_>>::new(n);
    for _ in 0..q {
        match scanner.scan::<usize>() {
            0 => {
                scan!(scanner, s, t, x: i32);
                seg.update(s, t + 1, Some(x));
            }
            1 => {
                scan!(scanner, s, t);
                writeln!(writer, "{}", seg.fold(s, t + 1)).ok();
            }
            _ => panic!("unknown query"),
        }
    }
}

pub fn from_vec(v: Vec<M::Agg>) -> Self

Examples found in repository?

crates/aizu_online_judge/src/dsl/dsl_2_h.rs (line 10)

6
7
8
9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24

pub fn dsl_2_h(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q);
    let mut seg = LazySegmentTree::<RangeMinRangeAdd<_>>::from_vec(vec![0; n]);
    for _ in 0..q {
        match scanner.scan::<usize>() {
            0 => {
                scan!(scanner, s, t, x: i64);
                seg.update(s, t + 1, x);
            }
            1 => {
                scan!(scanner, s, t);
                writeln!(writer, "{}", seg.fold(s, t + 1)).ok();
            }
            _ => panic!("unknown query"),
        }
    }
}

Hide additional examples

crates/aizu_online_judge/src/dsl/dsl_2_e.rs (line 10)

6
7
8
9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24

pub fn dsl_2_e(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q);
    let mut seg = LazySegmentTree::<RangeSumRangeAdd<_>>::from_vec(vec![(0, 1); n]);
    for _ in 0..q {
        match scanner.scan::<usize>() {
            0 => {
                scan!(scanner, s, t, x: u64);
                seg.update(s - 1, t, x);
            }
            1 => {
                scan!(scanner, i);
                writeln!(writer, "{}", seg.fold(i - 1, i).0).ok();
            }
            _ => panic!("unknown query"),
        }
    }
}

crates/aizu_online_judge/src/dsl/dsl_2_g.rs (line 10)

6
7
8
9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24

pub fn dsl_2_g(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q);
    let mut seg = LazySegmentTree::<RangeSumRangeAdd<_>>::from_vec(vec![(0, 1); n]);
    for _ in 0..q {
        match scanner.scan::<usize>() {
            0 => {
                scan!(scanner, s, t, x: u64);
                seg.update(s - 1, t, x);
            }
            1 => {
                scan!(scanner, s, t);
                writeln!(writer, "{}", seg.fold(s - 1, t).0).ok();
            }
            _ => panic!("unknown query"),
        }
    }
}

crates/aizu_online_judge/src/dsl/dsl_2_i.rs (line 10)

6
7
8
9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24

pub fn dsl_2_i(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q);
    let mut seg = LazySegmentTree::<RangeSumRangeUpdate<_>>::from_vec(vec![(0, 1); n]);
    for _ in 0..q {
        match scanner.scan::<usize>() {
            0 => {
                scan!(scanner, s, t, x: i64);
                seg.update(s, t + 1, Some(x));
            }
            1 => {
                scan!(scanner, s, t);
                writeln!(writer, "{}", seg.fold(s, t + 1).0).ok();
            }
            _ => panic!("unknown query"),
        }
    }
}

crates/library_checker/src/datastructure/range_affine_range_sum.rs (lines 14-16)

10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30

pub fn range_affine_range_sum(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q, a: [MInt998244353]);
    let mut seg = LazySegmentTree::<RangeSumRangeLinear<_>>::from_vec(
        a.take(n).map(|x| (x, MInt998244353::one())).collect::<_>(),
    );
    for _ in 0..q {
        match scanner.scan::<usize>() {
            0 => {
                scan!(scanner, l, r, bc: (MInt998244353, MInt998244353));
                seg.update(l, r, bc);
            }
            1 => {
                scan!(scanner, l, r);
                writeln!(writer, "{}", seg.fold(l, r).0).ok();
            }
            _ => panic!("unknown query"),
        }
    }
}

crates/aizu_online_judge/src/grl/grl_5_e.rs (line 24)

12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41

pub fn grl_5_e(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, c: [SizedCollect<usize>]);
    let edges = c
        .take(n)
        .enumerate()
        .flat_map(|(u, it)| it.into_iter().map(move |v| (u, v)))
        .collect();
    let mut graph = UndirectedSparseGraph::from_edges(n, edges);
    let hld = HeavyLightDecomposition::new(0, &mut graph);
    type M = (AdditiveOperation<u64>, AdditiveOperation<u64>);
    let mut seg = LazySegmentTree::<RangeSumRangeAdd<_>>::from_vec(vec![(0u64, 1u64); n]);

    scan!(scanner, q);
    for _ in 0..q {
        match scanner.scan::<usize>() {
            0 => {
                scan!(scanner, v, w: u64);
                hld.update(0, v, true, |l, r| seg.update(l, r, w));
            }
            1 => {
                scan!(scanner, u);
                let ans = hld.query::<M, _>(0, u, true, |l, r| seg.fold(l, r)).0;
                writeln!(writer, "{}", ans).ok();
            }
            _ => panic!("unknown query"),
        }
    }
}

Additional examples can be found in:

crates/library_checker/src/datastructure/range_chmin_chmax_add_range_sum.rs

pub fn update(&mut self, l: usize, r: usize, x: M::Act)

Examples found in repository?

crates/aizu_online_judge/src/dsl/dsl_2_d.rs (line 15)

6
7
8
9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24

pub fn dsl_2_d(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q);
    let mut seg = LazySegmentTree::<RangeMinRangeUpdate<_>>::new(n);
    for _ in 0..q {
        match scanner.scan::<usize>() {
            0 => {
                scan!(scanner, s, t, x: i32);
                seg.update(s, t + 1, Some(x));
            }
            1 => {
                scan!(scanner, i);
                writeln!(writer, "{}", seg.fold(i, i + 1)).ok();
            }
            _ => panic!("unknown query"),
        }
    }
}

Hide additional examples

crates/aizu_online_judge/src/dsl/dsl_2_f.rs (line 15)

6
7
8
9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24

pub fn dsl_2_f(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q);
    let mut seg = LazySegmentTree::<RangeMinRangeUpdate<_>>::new(n);
    for _ in 0..q {
        match scanner.scan::<usize>() {
            0 => {
                scan!(scanner, s, t, x: i32);
                seg.update(s, t + 1, Some(x));
            }
            1 => {
                scan!(scanner, s, t);
                writeln!(writer, "{}", seg.fold(s, t + 1)).ok();
            }
            _ => panic!("unknown query"),
        }
    }
}

crates/aizu_online_judge/src/dsl/dsl_2_h.rs (line 15)

6
7
8
9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24

pub fn dsl_2_h(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q);
    let mut seg = LazySegmentTree::<RangeMinRangeAdd<_>>::from_vec(vec![0; n]);
    for _ in 0..q {
        match scanner.scan::<usize>() {
            0 => {
                scan!(scanner, s, t, x: i64);
                seg.update(s, t + 1, x);
            }
            1 => {
                scan!(scanner, s, t);
                writeln!(writer, "{}", seg.fold(s, t + 1)).ok();
            }
            _ => panic!("unknown query"),
        }
    }
}

crates/aizu_online_judge/src/dsl/dsl_2_e.rs (line 15)

6
7
8
9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24

pub fn dsl_2_e(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q);
    let mut seg = LazySegmentTree::<RangeSumRangeAdd<_>>::from_vec(vec![(0, 1); n]);
    for _ in 0..q {
        match scanner.scan::<usize>() {
            0 => {
                scan!(scanner, s, t, x: u64);
                seg.update(s - 1, t, x);
            }
            1 => {
                scan!(scanner, i);
                writeln!(writer, "{}", seg.fold(i - 1, i).0).ok();
            }
            _ => panic!("unknown query"),
        }
    }
}

crates/aizu_online_judge/src/dsl/dsl_2_g.rs (line 15)

6
7
8
9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24

pub fn dsl_2_g(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q);
    let mut seg = LazySegmentTree::<RangeSumRangeAdd<_>>::from_vec(vec![(0, 1); n]);
    for _ in 0..q {
        match scanner.scan::<usize>() {
            0 => {
                scan!(scanner, s, t, x: u64);
                seg.update(s - 1, t, x);
            }
            1 => {
                scan!(scanner, s, t);
                writeln!(writer, "{}", seg.fold(s - 1, t).0).ok();
            }
            _ => panic!("unknown query"),
        }
    }
}

crates/aizu_online_judge/src/dsl/dsl_2_i.rs (line 15)

6
7
8
9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24

pub fn dsl_2_i(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q);
    let mut seg = LazySegmentTree::<RangeSumRangeUpdate<_>>::from_vec(vec![(0, 1); n]);
    for _ in 0..q {
        match scanner.scan::<usize>() {
            0 => {
                scan!(scanner, s, t, x: i64);
                seg.update(s, t + 1, Some(x));
            }
            1 => {
                scan!(scanner, s, t);
                writeln!(writer, "{}", seg.fold(s, t + 1).0).ok();
            }
            _ => panic!("unknown query"),
        }
    }
}

Additional examples can be found in:

crates/library_checker/src/datastructure/range_affine_range_sum.rs

crates/aizu_online_judge/src/grl/grl_5_e.rs

crates/library_checker/src/datastructure/range_chmin_chmax_add_range_sum.rs

pub fn fold(&mut self, l: usize, r: usize) -> M::Agg

Examples found in repository?

crates/competitive/src/data_structure/lazy_segment_tree.rs (line 156)

155
156
157
158
159
160

    pub fn get(&mut self, k: usize) -> M::Agg {
        self.fold(k, k + 1)
    }
    pub fn fold_all(&mut self) -> M::Agg {
        self.fold(0, self.n)
    }

Hide additional examples

crates/aizu_online_judge/src/dsl/dsl_2_d.rs (line 19)

6
7
8
9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24

pub fn dsl_2_d(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q);
    let mut seg = LazySegmentTree::<RangeMinRangeUpdate<_>>::new(n);
    for _ in 0..q {
        match scanner.scan::<usize>() {
            0 => {
                scan!(scanner, s, t, x: i32);
                seg.update(s, t + 1, Some(x));
            }
            1 => {
                scan!(scanner, i);
                writeln!(writer, "{}", seg.fold(i, i + 1)).ok();
            }
            _ => panic!("unknown query"),
        }
    }
}

crates/aizu_online_judge/src/dsl/dsl_2_f.rs (line 19)

6
7
8
9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24

pub fn dsl_2_f(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q);
    let mut seg = LazySegmentTree::<RangeMinRangeUpdate<_>>::new(n);
    for _ in 0..q {
        match scanner.scan::<usize>() {
            0 => {
                scan!(scanner, s, t, x: i32);
                seg.update(s, t + 1, Some(x));
            }
            1 => {
                scan!(scanner, s, t);
                writeln!(writer, "{}", seg.fold(s, t + 1)).ok();
            }
            _ => panic!("unknown query"),
        }
    }
}

crates/aizu_online_judge/src/dsl/dsl_2_h.rs (line 19)

6
7
8
9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24

pub fn dsl_2_h(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q);
    let mut seg = LazySegmentTree::<RangeMinRangeAdd<_>>::from_vec(vec![0; n]);
    for _ in 0..q {
        match scanner.scan::<usize>() {
            0 => {
                scan!(scanner, s, t, x: i64);
                seg.update(s, t + 1, x);
            }
            1 => {
                scan!(scanner, s, t);
                writeln!(writer, "{}", seg.fold(s, t + 1)).ok();
            }
            _ => panic!("unknown query"),
        }
    }
}

crates/aizu_online_judge/src/dsl/dsl_2_e.rs (line 19)

6
7
8
9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24

pub fn dsl_2_e(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q);
    let mut seg = LazySegmentTree::<RangeSumRangeAdd<_>>::from_vec(vec![(0, 1); n]);
    for _ in 0..q {
        match scanner.scan::<usize>() {
            0 => {
                scan!(scanner, s, t, x: u64);
                seg.update(s - 1, t, x);
            }
            1 => {
                scan!(scanner, i);
                writeln!(writer, "{}", seg.fold(i - 1, i).0).ok();
            }
            _ => panic!("unknown query"),
        }
    }
}

crates/aizu_online_judge/src/dsl/dsl_2_g.rs (line 19)

6
7
8
9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24

pub fn dsl_2_g(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q);
    let mut seg = LazySegmentTree::<RangeSumRangeAdd<_>>::from_vec(vec![(0, 1); n]);
    for _ in 0..q {
        match scanner.scan::<usize>() {
            0 => {
                scan!(scanner, s, t, x: u64);
                seg.update(s - 1, t, x);
            }
            1 => {
                scan!(scanner, s, t);
                writeln!(writer, "{}", seg.fold(s - 1, t).0).ok();
            }
            _ => panic!("unknown query"),
        }
    }
}

Additional examples can be found in:

crates/aizu_online_judge/src/dsl/dsl_2_i.rs

crates/library_checker/src/datastructure/range_affine_range_sum.rs

crates/aizu_online_judge/src/grl/grl_5_e.rs

crates/library_checker/src/datastructure/range_chmin_chmax_add_range_sum.rs

pub fn set(&mut self, k: usize, x: M::Agg)

pub fn get(&mut self, k: usize) -> M::Agg

pub fn fold_all(&mut self) -> M::Agg

pub fn position_acc<P>(&mut self, l: usize, r: usize, p: P) -> Option<usize>where P: Fn(&M::Agg) -> bool,

Returns the first index that satisfies a accumlative predicate.

pub fn rposition_acc<P>(&mut self, l: usize, r: usize, p: P) -> Option<usize>where P: Fn(&M::Agg) -> bool,

Returns the last index that satisfies a accumlative predicate.