Struct competitive::data_structure::SegmentTree

pub struct SegmentTree<M>
where
    M: Monoid,
{ /* private fields */ }

Implementations

impl<M> SegmentTree<M>
where M: Monoid,

pub fn new(n: usize) -> Self

Examples found in repository?

crates/aizu_online_judge/src/dsl/dsl_2_a.rs (line 10)

6
7
8
9
10
11
12
13
14
15
16
17
18
19

pub fn dsl_2_a(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q);
    let mut seg = SegmentTree::<MinOperation<_>>::new(n);
    for _ in 0..q {
        scan!(scanner, ty, x, y);
        if ty == 0 {
            seg.set(x, y as i32);
        } else {
            writeln!(writer, "{}", seg.fold(x..=y)).ok();
        }
    }
}

Hide additional examples

crates/aizu_online_judge/src/dsl/dsl_2_b.rs (line 10)

6
7
8
9
10
11
12
13
14
15
16
17
18
19

pub fn dsl_2_b(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q);
    let mut seg = SegmentTree::<AdditiveOperation<_>>::new(n);
    for _ in 0..q {
        scan!(scanner, ty, x: Usize1, y);
        if ty == 0 {
            seg.update(x, y as i32);
        } else {
            writeln!(writer, "{}", seg.fold(x..y)).ok();
        }
    }
}

pub fn from_vec(v: Vec<M::T>) -> Self

Examples found in repository?

crates/library_checker/src/datastructure/staticrmq.rs (line 24)

20
21
22
23
24
25
26
27
28

pub fn staticrmq_segment_tree(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q, a: [u64; n], lr: [(usize, usize)]);
    let seg = SegmentTree::<MinOperation<_>>::from_vec(a);
    for (l, r) in lr.take(q) {
        writeln!(writer, "{}", seg.fold(l..r)).ok();
    }
}

Hide additional examples

crates/library_checker/src/datastructure/point_add_range_sum.rs (line 37)

33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51

pub fn point_add_range_sum_segment_tree(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q, a: [i64; n]);
    let mut seg = SegmentTree::<AdditiveOperation<_>>::from_vec(a);
    for _ in 0..q {
        match scanner.scan::<usize>() {
            0 => {
                scan!(scanner, p, x: i64);
                seg.update(p, x);
            }
            1 => {
                scan!(scanner, l, r);
                writeln!(writer, "{}", seg.fold(l..r)).ok();
            }
            _ => panic!("unknown query"),
        }
    }
}

crates/library_checker/src/datastructure/point_set_range_composite.rs (line 14)

10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29

pub fn point_set_range_composite(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q, ab: [(MInt998244353, MInt998244353); n]);
    let mut seg = SegmentTree::<LinearOperation<_>>::from_vec(ab);
    for _ in 0..q {
        match scanner.scan::<usize>() {
            0 => {
                scan!(scanner, p, cd: (MInt998244353, MInt998244353));
                seg.set(p, cd);
            }
            1 => {
                scan!(scanner, l, r, x: MInt998244353);
                let (a, b) = seg.fold(l..r);
                writeln!(writer, "{}", a * x + b).ok();
            }
            _ => panic!("unknown query"),
        }
    }
}

crates/library_checker/src/datastructure/vertex_add_subtree_sum.rs (line 21)

9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35

pub fn vertex_add_subtree_sum(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q, a: [u64; n], p: [usize]);
    let edges = p.take(n - 1).enumerate().map(|(i, p)| (i + 1, p)).collect();
    let graph = UndirectedSparseGraph::from_edges(n, edges);
    let mut et = EulerTourForVertex::new(&graph);
    et.subtree_vertex_tour(0, n);
    let mut b = vec![0; n];
    for i in 0..n {
        b[et.vidx[i].0] = a[i];
    }
    let mut seg = SegmentTree::<AdditiveOperation<_>>::from_vec(b);
    for _ in 0..q {
        match scanner.scan::<usize>() {
            0 => {
                scan!(scanner, u, x: u64);
                et.subtree_update(u, x, |k, x| seg.update(k, x));
            }
            1 => {
                scan!(scanner, u);
                writeln!(writer, "{}", et.subtree_query(u, |l, r| seg.fold(l..r))).ok();
            }
            _ => panic!("unknown query"),
        }
    }
}

crates/library_checker/src/datastructure/vertex_set_path_composite.rs (line 21)

12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44

pub fn vertex_set_path_composite(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q, ab: [(MInt998244353, MInt998244353); n], (mut graph, _): @TreeGraphScanner::<usize, ()>::new(n));
    let hld = HeavyLightDecomposition::new(0, &mut graph);
    let mut nab = vec![(MInt998244353::default(), MInt998244353::default()); n];
    for i in 0..n {
        nab[hld.vidx[i]] = ab[i];
    }
    let mut seg1 = SegmentTree::<LinearOperation<_>>::from_vec(nab.clone());
    let mut seg2 = SegmentTree::<ReverseOperation<LinearOperation<_>>>::from_vec(nab);
    for _ in 0..q {
        match scanner.scan::<usize>() {
            0 => {
                scan!(scanner, p, cd: (MInt998244353, MInt998244353));
                seg1.set(hld.vidx[p], cd);
                seg2.set(hld.vidx[p], cd);
            }
            1 => {
                scan!(scanner, u, v, x: MInt998244353);
                let (a, b) = hld.query_noncom::<LinearOperation<_>, _, _>(
                    u,
                    v,
                    false,
                    |l, r| seg1.fold(l..r),
                    |l, r| seg2.fold(l..r),
                );
                writeln!(writer, "{}", a * x + b).ok();
            }
            _ => panic!("unknown query"),
        }
    }
}

pub fn set(&mut self, k: usize, x: M::T)

Examples found in repository?

crates/aizu_online_judge/src/dsl/dsl_2_a.rs (line 14)

6
7
8
9
10
11
12
13
14
15
16
17
18
19

pub fn dsl_2_a(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q);
    let mut seg = SegmentTree::<MinOperation<_>>::new(n);
    for _ in 0..q {
        scan!(scanner, ty, x, y);
        if ty == 0 {
            seg.set(x, y as i32);
        } else {
            writeln!(writer, "{}", seg.fold(x..=y)).ok();
        }
    }
}

Hide additional examples

crates/library_checker/src/datastructure/point_set_range_composite.rs (line 19)

10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29

pub fn point_set_range_composite(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q, ab: [(MInt998244353, MInt998244353); n]);
    let mut seg = SegmentTree::<LinearOperation<_>>::from_vec(ab);
    for _ in 0..q {
        match scanner.scan::<usize>() {
            0 => {
                scan!(scanner, p, cd: (MInt998244353, MInt998244353));
                seg.set(p, cd);
            }
            1 => {
                scan!(scanner, l, r, x: MInt998244353);
                let (a, b) = seg.fold(l..r);
                writeln!(writer, "{}", a * x + b).ok();
            }
            _ => panic!("unknown query"),
        }
    }
}

crates/library_checker/src/datastructure/vertex_set_path_composite.rs (line 27)

12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44

pub fn vertex_set_path_composite(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q, ab: [(MInt998244353, MInt998244353); n], (mut graph, _): @TreeGraphScanner::<usize, ()>::new(n));
    let hld = HeavyLightDecomposition::new(0, &mut graph);
    let mut nab = vec![(MInt998244353::default(), MInt998244353::default()); n];
    for i in 0..n {
        nab[hld.vidx[i]] = ab[i];
    }
    let mut seg1 = SegmentTree::<LinearOperation<_>>::from_vec(nab.clone());
    let mut seg2 = SegmentTree::<ReverseOperation<LinearOperation<_>>>::from_vec(nab);
    for _ in 0..q {
        match scanner.scan::<usize>() {
            0 => {
                scan!(scanner, p, cd: (MInt998244353, MInt998244353));
                seg1.set(hld.vidx[p], cd);
                seg2.set(hld.vidx[p], cd);
            }
            1 => {
                scan!(scanner, u, v, x: MInt998244353);
                let (a, b) = hld.query_noncom::<LinearOperation<_>, _, _>(
                    u,
                    v,
                    false,
                    |l, r| seg1.fold(l..r),
                    |l, r| seg2.fold(l..r),
                );
                writeln!(writer, "{}", a * x + b).ok();
            }
            _ => panic!("unknown query"),
        }
    }
}

pub fn update(&mut self, k: usize, x: M::T)

Examples found in repository?

crates/aizu_online_judge/src/dsl/dsl_2_b.rs (line 14)

6
7
8
9
10
11
12
13
14
15
16
17
18
19

pub fn dsl_2_b(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q);
    let mut seg = SegmentTree::<AdditiveOperation<_>>::new(n);
    for _ in 0..q {
        scan!(scanner, ty, x: Usize1, y);
        if ty == 0 {
            seg.update(x, y as i32);
        } else {
            writeln!(writer, "{}", seg.fold(x..y)).ok();
        }
    }
}

Hide additional examples

crates/library_checker/src/datastructure/point_add_range_sum.rs (line 42)

33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51

pub fn point_add_range_sum_segment_tree(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q, a: [i64; n]);
    let mut seg = SegmentTree::<AdditiveOperation<_>>::from_vec(a);
    for _ in 0..q {
        match scanner.scan::<usize>() {
            0 => {
                scan!(scanner, p, x: i64);
                seg.update(p, x);
            }
            1 => {
                scan!(scanner, l, r);
                writeln!(writer, "{}", seg.fold(l..r)).ok();
            }
            _ => panic!("unknown query"),
        }
    }
}

crates/library_checker/src/datastructure/vertex_add_subtree_sum.rs (line 26)

9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35

pub fn vertex_add_subtree_sum(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q, a: [u64; n], p: [usize]);
    let edges = p.take(n - 1).enumerate().map(|(i, p)| (i + 1, p)).collect();
    let graph = UndirectedSparseGraph::from_edges(n, edges);
    let mut et = EulerTourForVertex::new(&graph);
    et.subtree_vertex_tour(0, n);
    let mut b = vec![0; n];
    for i in 0..n {
        b[et.vidx[i].0] = a[i];
    }
    let mut seg = SegmentTree::<AdditiveOperation<_>>::from_vec(b);
    for _ in 0..q {
        match scanner.scan::<usize>() {
            0 => {
                scan!(scanner, u, x: u64);
                et.subtree_update(u, x, |k, x| seg.update(k, x));
            }
            1 => {
                scan!(scanner, u);
                writeln!(writer, "{}", et.subtree_query(u, |l, r| seg.fold(l..r))).ok();
            }
            _ => panic!("unknown query"),
        }
    }
}

pub fn get(&self, k: usize) -> M::T

pub fn fold<R>(&self, range: R) -> M::T
where R: RangeBounds<usize>,

Examples found in repository?

crates/library_checker/src/datastructure/staticrmq.rs (line 26)

20
21
22
23
24
25
26
27
28

pub fn staticrmq_segment_tree(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q, a: [u64; n], lr: [(usize, usize)]);
    let seg = SegmentTree::<MinOperation<_>>::from_vec(a);
    for (l, r) in lr.take(q) {
        writeln!(writer, "{}", seg.fold(l..r)).ok();
    }
}

Hide additional examples

crates/aizu_online_judge/src/dsl/dsl_2_a.rs (line 16)

6
7
8
9
10
11
12
13
14
15
16
17
18
19

pub fn dsl_2_a(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q);
    let mut seg = SegmentTree::<MinOperation<_>>::new(n);
    for _ in 0..q {
        scan!(scanner, ty, x, y);
        if ty == 0 {
            seg.set(x, y as i32);
        } else {
            writeln!(writer, "{}", seg.fold(x..=y)).ok();
        }
    }
}

crates/aizu_online_judge/src/dsl/dsl_2_b.rs (line 16)

6
7
8
9
10
11
12
13
14
15
16
17
18
19

pub fn dsl_2_b(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q);
    let mut seg = SegmentTree::<AdditiveOperation<_>>::new(n);
    for _ in 0..q {
        scan!(scanner, ty, x: Usize1, y);
        if ty == 0 {
            seg.update(x, y as i32);
        } else {
            writeln!(writer, "{}", seg.fold(x..y)).ok();
        }
    }
}

crates/library_checker/src/datastructure/point_add_range_sum.rs (line 46)

33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51

pub fn point_add_range_sum_segment_tree(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q, a: [i64; n]);
    let mut seg = SegmentTree::<AdditiveOperation<_>>::from_vec(a);
    for _ in 0..q {
        match scanner.scan::<usize>() {
            0 => {
                scan!(scanner, p, x: i64);
                seg.update(p, x);
            }
            1 => {
                scan!(scanner, l, r);
                writeln!(writer, "{}", seg.fold(l..r)).ok();
            }
            _ => panic!("unknown query"),
        }
    }
}

crates/library_checker/src/datastructure/point_set_range_composite.rs (line 23)

10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29

pub fn point_set_range_composite(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q, ab: [(MInt998244353, MInt998244353); n]);
    let mut seg = SegmentTree::<LinearOperation<_>>::from_vec(ab);
    for _ in 0..q {
        match scanner.scan::<usize>() {
            0 => {
                scan!(scanner, p, cd: (MInt998244353, MInt998244353));
                seg.set(p, cd);
            }
            1 => {
                scan!(scanner, l, r, x: MInt998244353);
                let (a, b) = seg.fold(l..r);
                writeln!(writer, "{}", a * x + b).ok();
            }
            _ => panic!("unknown query"),
        }
    }
}

crates/library_checker/src/datastructure/vertex_add_subtree_sum.rs (line 30)

9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35

pub fn vertex_add_subtree_sum(reader: impl Read, mut writer: impl Write) {
    let s = read_all_unchecked(reader);
    let mut scanner = Scanner::new(&s);
    scan!(scanner, n, q, a: [u64; n], p: [usize]);
    let edges = p.take(n - 1).enumerate().map(|(i, p)| (i + 1, p)).collect();
    let graph = UndirectedSparseGraph::from_edges(n, edges);
    let mut et = EulerTourForVertex::new(&graph);
    et.subtree_vertex_tour(0, n);
    let mut b = vec![0; n];
    for i in 0..n {
        b[et.vidx[i].0] = a[i];
    }
    let mut seg = SegmentTree::<AdditiveOperation<_>>::from_vec(b);
    for _ in 0..q {
        match scanner.scan::<usize>() {
            0 => {
                scan!(scanner, u, x: u64);
                et.subtree_update(u, x, |k, x| seg.update(k, x));
            }
            1 => {
                scan!(scanner, u);
                writeln!(writer, "{}", et.subtree_query(u, |l, r| seg.fold(l..r))).ok();
            }
            _ => panic!("unknown query"),
        }
    }
}

Additional examples can be found in:

crates/library_checker/src/datastructure/vertex_set_path_composite.rs

pub fn position_acc<R, F>(&self, range: R, f: F) -> Option<usize>
where R: RangeBounds<usize>, F: Fn(&M::T) -> bool,

Returns the first index that satisfies a accumlative predicate.

pub fn rposition_acc<R, F>(&self, range: R, f: F) -> Option<usize>
where R: RangeBounds<usize>, F: Fn(&M::T) -> bool,

Returns the last index that satisfies a accumlative predicate.

pub fn as_slice(&self) -> &[M::T]

impl<M> SegmentTree<M>
where M: AbelianMonoid,

pub fn fold_all(&self) -> M::T