Trait competitive::data_structure::RankSelectDictionaries

pub trait RankSelectDictionaries {
    // Required methods
    fn bit_length(&self) -> usize;
    fn access(&self, k: usize) -> bool;

    // Provided methods
    fn rank1(&self, k: usize) -> usize { ... }
    fn rank0(&self, k: usize) -> usize { ... }
    fn select1(&self, k: usize) -> Option<usize> { ... }
    fn select0(&self, k: usize) -> Option<usize> { ... }
}

rank_i(select_i(k)) = k rank_i(select_i(k) + 1) = k + 1

Required Methods

fn bit_length(&self) -> usize

fn access(&self, k: usize) -> bool

get k-th bit

Provided Methods

fn rank1(&self, k: usize) -> usize

the number of 1 in [0, k)

Examples found in repository?

crates/competitive/src/data_structure/bit_vector.rs (line 15)

14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33

    fn rank0(&self, k: usize) -> usize {
        k - self.rank1(k)
    }
    /// index of k-th 1
    fn select1(&self, k: usize) -> Option<usize> {
        let n = self.bit_length();
        if self.rank1(n) <= k {
            return None;
        }
        let (mut l, mut r) = (0, n);
        while r - l > 1 {
            let m = (l + r) / 2;
            if self.rank1(m) <= k {
                l = m;
            } else {
                r = m;
            }
        }
        Some(l)
    }

Hide additional examples

crates/competitive/src/data_structure/wavelet_matrix.rs (line 38)

29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70
71
72
73
74
75
76
77
78
79
80
81
82
83
84
85
86
87
88
89
90
91
92
93
94
95
96
97
98
99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122
123
124
125
126
127
128
129
130
131
132
133
134
135
136
137
138
139
140
141
142
143
144
145
146
147
148
149
150
151
152
153
154
155
156
157
158
159
160
161
162
163
164
165
166
167
168
169

    pub fn new_with_init<T, F>(v: Vec<T>, bit_length: usize, mut f: F) -> Self
    where
        T: Clone + RankSelectDictionaries,
        F: FnMut(usize, usize, T),
    {
        let this = Self::new(v.clone(), bit_length);
        for (mut k, v) in v.into_iter().enumerate() {
            for (d, &(c, ref b)) in this.table.iter().rev().enumerate().rev() {
                if v.access(d) {
                    k = c + b.rank1(k);
                } else {
                    k = b.rank0(k);
                }
                f(d, k, v.clone());
            }
        }
        this
    }
    /// get k-th value
    pub fn access(&self, mut k: usize) -> usize {
        let mut val = 0;
        for (d, &(c, ref b)) in self.table.iter().rev().enumerate().rev() {
            if b.access(k) {
                k = c + b.rank1(k);
                val |= 1 << d;
            } else {
                k = b.rank0(k);
            }
        }
        val
    }
    /// the number of val in range
    pub fn rank(&self, val: usize, mut range: Range<usize>) -> usize {
        for (d, &(c, ref b)) in self.table.iter().rev().enumerate().rev() {
            if val.access(d) {
                range.start = c + b.rank1(range.start);
                range.end = c + b.rank1(range.end);
            } else {
                range.start = b.rank0(range.start);
                range.end = b.rank0(range.end);
            }
        }
        range.end - range.start
    }
    /// index of k-th val
    pub fn select(&self, val: usize, k: usize) -> Option<usize> {
        if self.rank(val, 0..self.len) <= k {
            return None;
        }
        let mut i = 0;
        for (d, &(c, ref b)) in self.table.iter().rev().enumerate().rev() {
            if val.access(d) {
                i = c + b.rank1(i);
            } else {
                i = b.rank0(i);
            }
        }
        i += k;
        for &(c, ref b) in self.table.iter().rev() {
            if i >= c {
                i = b.select1(i - c).unwrap();
            } else {
                i = b.select0(i).unwrap();
            }
        }
        Some(i)
    }
    /// get k-th smallest value in range
    pub fn quantile(&self, mut range: Range<usize>, mut k: usize) -> usize {
        let mut val = 0;
        for (d, &(c, ref b)) in self.table.iter().rev().enumerate().rev() {
            let z = b.rank0(range.end) - b.rank0(range.start);
            if z <= k {
                k -= z;
                val |= 1 << d;
                range.start = c + b.rank1(range.start);
                range.end = c + b.rank1(range.end);
            } else {
                range.start = b.rank0(range.start);
                range.end = b.rank0(range.end);
            }
        }
        val
    }
    /// get k-th smallest value out of range
    pub fn quantile_outer(&self, mut range: Range<usize>, mut k: usize) -> usize {
        let mut val = 0;
        let mut orange = 0..self.len;
        for (d, &(c, ref b)) in self.table.iter().rev().enumerate().rev() {
            let z = b.rank0(orange.end) - b.rank0(orange.start) + b.rank0(range.start)
                - b.rank0(range.end);
            if z <= k {
                k -= z;
                val |= 1 << d;
                range.start = c + b.rank1(range.start);
                range.end = c + b.rank1(range.end);
                orange.start = c + b.rank1(orange.start);
                orange.end = c + b.rank1(orange.end);
            } else {
                range.start = b.rank0(range.start);
                range.end = b.rank0(range.end);
                orange.start = b.rank0(orange.start);
                orange.end = b.rank0(orange.end);
            }
        }
        val
    }
    /// the number of value less than val in range
    pub fn rank_lessthan(&self, val: usize, mut range: Range<usize>) -> usize {
        let mut res = 0;
        for (d, &(c, ref b)) in self.table.iter().rev().enumerate().rev() {
            if val.access(d) {
                res += b.rank0(range.end) - b.rank0(range.start);
                range.start = c + b.rank1(range.start);
                range.end = c + b.rank1(range.end);
            } else {
                range.start = b.rank0(range.start);
                range.end = b.rank0(range.end);
            }
        }
        res
    }
    /// the number of valrange in range
    pub fn rank_range(&self, valrange: Range<usize>, range: Range<usize>) -> usize {
        self.rank_lessthan(valrange.end, range.clone()) - self.rank_lessthan(valrange.start, range)
    }
    pub fn query_less_than<F>(&self, val: usize, mut range: Range<usize>, mut f: F)
    where
        F: FnMut(usize, Range<usize>),
    {
        for (d, &(c, ref b)) in self.table.iter().rev().enumerate().rev() {
            if val.access(d) {
                f(d, b.rank0(range.start)..b.rank0(range.end));
                range.start = c + b.rank1(range.start);
                range.end = c + b.rank1(range.end);
            } else {
                range.start = b.rank0(range.start);
                range.end = b.rank0(range.end);
            }
        }
    }

fn rank0(&self, k: usize) -> usize

the number of 0 in [0, k)

Examples found in repository?

crates/competitive/src/data_structure/bit_vector.rs (line 37)

35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50

    fn select0(&self, k: usize) -> Option<usize> {
        let n = self.bit_length();
        if self.rank0(n) <= k {
            return None;
        }
        let (mut l, mut r) = (0, n);
        while r - l > 1 {
            let m = (l + r) / 2;
            if self.rank0(m) <= k {
                l = m;
            } else {
                r = m;
            }
        }
        Some(l)
    }

Hide additional examples

crates/competitive/src/data_structure/wavelet_matrix.rs (line 19)

11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70
71
72
73
74
75
76
77
78
79
80
81
82
83
84
85
86
87
88
89
90
91
92
93
94
95
96
97
98
99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122
123
124
125
126
127
128
129
130
131
132
133
134
135
136
137
138
139
140
141
142
143
144
145
146
147
148
149
150
151
152
153
154
155
156
157
158
159
160
161
162
163
164
165
166
167
168
169

    pub fn new<T>(mut v: Vec<T>, bit_length: usize) -> Self
    where
        T: Clone + RankSelectDictionaries,
    {
        let len = v.len();
        let mut table = Vec::new();
        for d in (0..bit_length).rev() {
            let b: BitVector = v.iter().map(|x| x.access(d)).collect();
            table.push((b.rank0(len), b));
            v = v
                .iter()
                .filter(|&x| !x.access(d))
                .chain(v.iter().filter(|&x| x.access(d)))
                .cloned()
                .collect();
        }
        Self { len, table }
    }
    pub fn new_with_init<T, F>(v: Vec<T>, bit_length: usize, mut f: F) -> Self
    where
        T: Clone + RankSelectDictionaries,
        F: FnMut(usize, usize, T),
    {
        let this = Self::new(v.clone(), bit_length);
        for (mut k, v) in v.into_iter().enumerate() {
            for (d, &(c, ref b)) in this.table.iter().rev().enumerate().rev() {
                if v.access(d) {
                    k = c + b.rank1(k);
                } else {
                    k = b.rank0(k);
                }
                f(d, k, v.clone());
            }
        }
        this
    }
    /// get k-th value
    pub fn access(&self, mut k: usize) -> usize {
        let mut val = 0;
        for (d, &(c, ref b)) in self.table.iter().rev().enumerate().rev() {
            if b.access(k) {
                k = c + b.rank1(k);
                val |= 1 << d;
            } else {
                k = b.rank0(k);
            }
        }
        val
    }
    /// the number of val in range
    pub fn rank(&self, val: usize, mut range: Range<usize>) -> usize {
        for (d, &(c, ref b)) in self.table.iter().rev().enumerate().rev() {
            if val.access(d) {
                range.start = c + b.rank1(range.start);
                range.end = c + b.rank1(range.end);
            } else {
                range.start = b.rank0(range.start);
                range.end = b.rank0(range.end);
            }
        }
        range.end - range.start
    }
    /// index of k-th val
    pub fn select(&self, val: usize, k: usize) -> Option<usize> {
        if self.rank(val, 0..self.len) <= k {
            return None;
        }
        let mut i = 0;
        for (d, &(c, ref b)) in self.table.iter().rev().enumerate().rev() {
            if val.access(d) {
                i = c + b.rank1(i);
            } else {
                i = b.rank0(i);
            }
        }
        i += k;
        for &(c, ref b) in self.table.iter().rev() {
            if i >= c {
                i = b.select1(i - c).unwrap();
            } else {
                i = b.select0(i).unwrap();
            }
        }
        Some(i)
    }
    /// get k-th smallest value in range
    pub fn quantile(&self, mut range: Range<usize>, mut k: usize) -> usize {
        let mut val = 0;
        for (d, &(c, ref b)) in self.table.iter().rev().enumerate().rev() {
            let z = b.rank0(range.end) - b.rank0(range.start);
            if z <= k {
                k -= z;
                val |= 1 << d;
                range.start = c + b.rank1(range.start);
                range.end = c + b.rank1(range.end);
            } else {
                range.start = b.rank0(range.start);
                range.end = b.rank0(range.end);
            }
        }
        val
    }
    /// get k-th smallest value out of range
    pub fn quantile_outer(&self, mut range: Range<usize>, mut k: usize) -> usize {
        let mut val = 0;
        let mut orange = 0..self.len;
        for (d, &(c, ref b)) in self.table.iter().rev().enumerate().rev() {
            let z = b.rank0(orange.end) - b.rank0(orange.start) + b.rank0(range.start)
                - b.rank0(range.end);
            if z <= k {
                k -= z;
                val |= 1 << d;
                range.start = c + b.rank1(range.start);
                range.end = c + b.rank1(range.end);
                orange.start = c + b.rank1(orange.start);
                orange.end = c + b.rank1(orange.end);
            } else {
                range.start = b.rank0(range.start);
                range.end = b.rank0(range.end);
                orange.start = b.rank0(orange.start);
                orange.end = b.rank0(orange.end);
            }
        }
        val
    }
    /// the number of value less than val in range
    pub fn rank_lessthan(&self, val: usize, mut range: Range<usize>) -> usize {
        let mut res = 0;
        for (d, &(c, ref b)) in self.table.iter().rev().enumerate().rev() {
            if val.access(d) {
                res += b.rank0(range.end) - b.rank0(range.start);
                range.start = c + b.rank1(range.start);
                range.end = c + b.rank1(range.end);
            } else {
                range.start = b.rank0(range.start);
                range.end = b.rank0(range.end);
            }
        }
        res
    }
    /// the number of valrange in range
    pub fn rank_range(&self, valrange: Range<usize>, range: Range<usize>) -> usize {
        self.rank_lessthan(valrange.end, range.clone()) - self.rank_lessthan(valrange.start, range)
    }
    pub fn query_less_than<F>(&self, val: usize, mut range: Range<usize>, mut f: F)
    where
        F: FnMut(usize, Range<usize>),
    {
        for (d, &(c, ref b)) in self.table.iter().rev().enumerate().rev() {
            if val.access(d) {
                f(d, b.rank0(range.start)..b.rank0(range.end));
                range.start = c + b.rank1(range.start);
                range.end = c + b.rank1(range.end);
            } else {
                range.start = b.rank0(range.start);
                range.end = b.rank0(range.end);
            }
        }
    }

fn select1(&self, k: usize) -> Option<usize>

index of k-th 1

Examples found in repository?

crates/competitive/src/data_structure/bit_vector.rs (line 115)

100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116

    fn select1(&self, mut k: usize) -> Option<usize> {
        let (mut l, mut r) = (0, self.data.len());
        if self.sum <= k {
            return None;
        }
        while r - l > 1 {
            let m = (l + r) / 2;
            if self.data[m].1 <= k {
                l = m;
            } else {
                r = m;
            }
        }
        let (bit, sum) = self.data[l];
        k -= sum;
        Some(l * Self::WORD_SIZE + bit.select1(k).unwrap())
    }

Hide additional examples

crates/competitive/src/data_structure/wavelet_matrix.rs (line 89)

74
75
76
77
78
79
80
81
82
83
84
85
86
87
88
89
90
91
92
93
94
95

    pub fn select(&self, val: usize, k: usize) -> Option<usize> {
        if self.rank(val, 0..self.len) <= k {
            return None;
        }
        let mut i = 0;
        for (d, &(c, ref b)) in self.table.iter().rev().enumerate().rev() {
            if val.access(d) {
                i = c + b.rank1(i);
            } else {
                i = b.rank0(i);
            }
        }
        i += k;
        for &(c, ref b) in self.table.iter().rev() {
            if i >= c {
                i = b.select1(i - c).unwrap();
            } else {
                i = b.select0(i).unwrap();
            }
        }
        Some(i)
    }

fn select0(&self, k: usize) -> Option<usize>

index of k-th 0

Examples found in repository?

crates/competitive/src/data_structure/bit_vector.rs (line 132)

117
118
119
120
121
122
123
124
125
126
127
128
129
130
131
132
133

    fn select0(&self, mut k: usize) -> Option<usize> {
        let (mut l, mut r) = (0, self.data.len());
        if r * Self::WORD_SIZE - self.sum <= k {
            return None;
        }
        while r - l > 1 {
            let m = (l + r) / 2;
            if m * Self::WORD_SIZE - self.data[m].1 <= k {
                l = m;
            } else {
                r = m;
            }
        }
        let (bit, sum) = self.data[l];
        k -= l * Self::WORD_SIZE - sum;
        Some(l * Self::WORD_SIZE + bit.select0(k).unwrap())
    }

Hide additional examples

crates/competitive/src/data_structure/wavelet_matrix.rs (line 91)

74
75
76
77
78
79
80
81
82
83
84
85
86
87
88
89
90
91
92
93
94
95

    pub fn select(&self, val: usize, k: usize) -> Option<usize> {
        if self.rank(val, 0..self.len) <= k {
            return None;
        }
        let mut i = 0;
        for (d, &(c, ref b)) in self.table.iter().rev().enumerate().rev() {
            if val.access(d) {
                i = c + b.rank1(i);
            } else {
                i = b.rank0(i);
            }
        }
        i += k;
        for &(c, ref b) in self.table.iter().rev() {
            if i >= c {
                i = b.select1(i - c).unwrap();
            } else {
                i = b.select0(i).unwrap();
            }
        }
        Some(i)
    }