Trait competitive::num::RangeBoundsExt

pub trait RangeBoundsExt<T> {
    // Required methods
    fn start_bound_included_checked(&self) -> Option<T>;
    fn start_bound_excluded_checked(&self) -> Option<T>;
    fn end_bound_included_checked(&self) -> Option<T>;
    fn end_bound_excluded_checked(&self) -> Option<T>;
    fn start_bound_included(&self) -> T;
    fn start_bound_excluded(&self) -> T;
    fn end_bound_included(&self) -> T;
    fn end_bound_excluded(&self) -> T;

    // Provided methods
    fn to_range_checked(&self) -> Option<Range<T>> { ... }
    fn to_range(&self) -> Range<T> { ... }
    fn to_range_inclusive_checked(&self) -> Option<RangeInclusive<T>> { ... }
    fn to_range_inclusive(&self) -> RangeInclusive<T> { ... }
}

Required Methods

fn start_bound_included_checked(&self) -> Option<T>

fn start_bound_excluded_checked(&self) -> Option<T>

fn end_bound_included_checked(&self) -> Option<T>

fn end_bound_excluded_checked(&self) -> Option<T>

fn start_bound_included(&self) -> T

fn start_bound_excluded(&self) -> T

fn end_bound_included(&self) -> T

fn end_bound_excluded(&self) -> T

Provided Methods

fn to_range_checked(&self) -> Option<Range<T>>

fn to_range(&self) -> Range<T>

Examples found in repository?

crates/competitive/src/data_structure/segment_tree.rs (line 87)

83
84
85
86
87
88
89
90
91
92
93
94
95
96
97
98
99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122
123
124
125
126
127
128
129
130
131
132
133
134
135
136
137
138
139
140
141
142
143
144
145
146
147
148
149
150
151
152
153
154
155
156
157
158
159
160
161
162
163
164
165
166
167
168
169
170
171
172
173
174
175
176
177
178
179
180
181
182
183
184
185
186
187
188
189
190
191
192
193
194
195
196
197
198
199
200
201
202
203
204
205
206
207
208
209
210
211
212
213
214

    pub fn fold<R>(&self, range: R) -> M::T
    where
        R: RangeBounds<usize>,
    {
        let range = range.to_range();
        debug_assert!(range.end <= self.n);
        let mut l = range.start + self.n;
        let mut r = range.end + self.n;
        let mut vl = M::unit();
        let mut vr = M::unit();
        while l < r {
            if l & 1 != 0 {
                vl = M::operate(&vl, &self.seg[l]);
                l += 1;
            }
            if r & 1 != 0 {
                r -= 1;
                vr = M::operate(&self.seg[r], &vr);
            }
            l /= 2;
            r /= 2;
        }
        M::operate(&vl, &vr)
    }
    fn bisect_perfect<F>(&self, mut pos: usize, mut acc: M::T, f: F) -> (usize, M::T)
    where
        F: Fn(&M::T) -> bool,
    {
        while pos < self.n {
            pos <<= 1;
            let nacc = M::operate(&acc, &self.seg[pos]);
            if !f(&nacc) {
                acc = nacc;
                pos += 1;
            }
        }
        (pos - self.n, acc)
    }
    fn rbisect_perfect<F>(&self, mut pos: usize, mut acc: M::T, f: F) -> (usize, M::T)
    where
        F: Fn(&M::T) -> bool,
    {
        while pos < self.n {
            pos = pos * 2 + 1;
            let nacc = M::operate(&self.seg[pos], &acc);
            if !f(&nacc) {
                acc = nacc;
                pos -= 1;
            }
        }
        (pos - self.n, acc)
    }
    /// Returns the first index that satisfies a accumlative predicate.
    pub fn position_acc<R, F>(&self, range: R, f: F) -> Option<usize>
    where
        R: RangeBounds<usize>,
        F: Fn(&M::T) -> bool,
    {
        let range = range.to_range();
        debug_assert!(range.end <= self.n);
        let mut l = range.start + self.n;
        let r = range.end + self.n;
        let mut k = 0usize;
        let mut acc = M::unit();
        while l < r >> k {
            if l & 1 != 0 {
                let nacc = M::operate(&acc, &self.seg[l]);
                if f(&nacc) {
                    return Some(self.bisect_perfect(l, acc, f).0);
                }
                acc = nacc;
                l += 1;
            }
            l >>= 1;
            k += 1;
        }
        for k in (0..k).rev() {
            let r = r >> k;
            if r & 1 != 0 {
                let nacc = M::operate(&acc, &self.seg[r - 1]);
                if f(&nacc) {
                    return Some(self.bisect_perfect(r - 1, acc, f).0);
                }
                acc = nacc;
            }
        }
        None
    }
    /// Returns the last index that satisfies a accumlative predicate.
    pub fn rposition_acc<R, F>(&self, range: R, f: F) -> Option<usize>
    where
        R: RangeBounds<usize>,
        F: Fn(&M::T) -> bool,
    {
        let range = range.to_range();
        debug_assert!(range.end <= self.n);
        let mut l = range.start + self.n;
        let mut r = range.end + self.n;
        let mut c = 0usize;
        let mut k = 0usize;
        let mut acc = M::unit();
        while l >> k < r {
            c <<= 1;
            if l & 1 << k != 0 {
                l += 1 << k;
                c += 1;
            }
            if r & 1 != 0 {
                r -= 1;
                let nacc = M::operate(&self.seg[r], &acc);
                if f(&nacc) {
                    return Some(self.rbisect_perfect(r, acc, f).0);
                }
                acc = nacc;
            }
            r >>= 1;
            k += 1;
        }
        for k in (0..k).rev() {
            if c & 1 != 0 {
                l -= 1 << k;
                let l = l >> k;
                let nacc = M::operate(&self.seg[l], &acc);
                if f(&nacc) {
                    return Some(self.rbisect_perfect(l, acc, f).0);
                }
                acc = nacc;
            }
            c >>= 1;
        }
        None
    }

Hide additional examples

crates/competitive/src/data_structure/segment_tree_map.rs (line 93)

89
90
91
92
93
94
95
96
97
98
99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122
123
124
125
126
127
128
129
130
131
132
133
134
135
136
137
138
139
140
141
142
143
144
145
146
147
148
149
150
151
152
153
154
155
156
157
158
159
160
161
162
163
164
165
166
167
168
169
170
171
172
173
174
175
176
177
178
179
180
181
182
183
184
185
186
187
188
189
190
191
192
193
194
195
196
197
198
199
200
201
202
203
204
205
206
207
208
209
210
211
212
213
214
215
216
217
218
219
220

    pub fn fold<R>(&self, range: R) -> M::T
    where
        R: RangeBounds<usize>,
    {
        let range = range.to_range();
        debug_assert!(range.end <= self.n);
        let mut l = range.start + self.n;
        let mut r = range.end + self.n;
        let mut vl = M::unit();
        let mut vr = M::unit();
        while l < r {
            if l & 1 != 0 {
                vl = M::operate(&vl, self.get_ref(l));
                l += 1;
            }
            if r & 1 != 0 {
                r -= 1;
                vr = M::operate(self.get_ref(r), &vr);
            }
            l /= 2;
            r /= 2;
        }
        M::operate(&vl, &vr)
    }
    fn bisect_perfect<F>(&self, mut pos: usize, mut acc: M::T, f: F) -> (usize, M::T)
    where
        F: Fn(&M::T) -> bool,
    {
        while pos < self.n {
            pos <<= 1;
            let nacc = M::operate(&acc, self.get_ref(pos));
            if !f(&nacc) {
                acc = nacc;
                pos += 1;
            }
        }
        (pos - self.n, acc)
    }
    fn rbisect_perfect<F>(&self, mut pos: usize, mut acc: M::T, f: F) -> (usize, M::T)
    where
        F: Fn(&M::T) -> bool,
    {
        while pos < self.n {
            pos = pos * 2 + 1;
            let nacc = M::operate(self.get_ref(pos), &acc);
            if !f(&nacc) {
                acc = nacc;
                pos -= 1;
            }
        }
        (pos - self.n, acc)
    }
    /// Returns the first index that satisfies a accumlative predicate.
    pub fn position_acc<R, F>(&self, range: R, f: F) -> Option<usize>
    where
        R: RangeBounds<usize>,
        F: Fn(&M::T) -> bool,
    {
        let range = range.to_range();
        debug_assert!(range.end <= self.n);
        let mut l = range.start + self.n;
        let r = range.end + self.n;
        let mut k = 0usize;
        let mut acc = M::unit();
        while l < r >> k {
            if l & 1 != 0 {
                let nacc = M::operate(&acc, self.get_ref(l));
                if f(&nacc) {
                    return Some(self.bisect_perfect(l, acc, f).0);
                }
                acc = nacc;
                l += 1;
            }
            l >>= 1;
            k += 1;
        }
        for k in (0..k).rev() {
            let r = r >> k;
            if r & 1 != 0 {
                let nacc = M::operate(&acc, self.get_ref(r - 1));
                if f(&nacc) {
                    return Some(self.bisect_perfect(r - 1, acc, f).0);
                }
                acc = nacc;
            }
        }
        None
    }
    /// Returns the last index that satisfies a accumlative predicate.
    pub fn rposition_acc<R, F>(&self, range: R, f: F) -> Option<usize>
    where
        R: RangeBounds<usize>,
        F: Fn(&M::T) -> bool,
    {
        let range = range.to_range();
        debug_assert!(range.end <= self.n);
        let mut l = range.start + self.n;
        let mut r = range.end + self.n;
        let mut c = 0usize;
        let mut k = 0usize;
        let mut acc = M::unit();
        while l >> k < r {
            c <<= 1;
            if l & 1 << k != 0 {
                l += 1 << k;
                c += 1;
            }
            if r & 1 != 0 {
                r -= 1;
                let nacc = M::operate(self.get_ref(r), &acc);
                if f(&nacc) {
                    return Some(self.rbisect_perfect(r, acc, f).0);
                }
                acc = nacc;
            }
            r >>= 1;
            k += 1;
        }
        for k in (0..k).rev() {
            if c & 1 != 0 {
                l -= 1 << k;
                let l = l >> k;
                let nacc = M::operate(self.get_ref(l), &acc);
                if f(&nacc) {
                    return Some(self.rbisect_perfect(l, acc, f).0);
                }
                acc = nacc;
            }
            c >>= 1;
        }
        None
    }

fn to_range_inclusive_checked(&self) -> Option<RangeInclusive<T>>

fn to_range_inclusive(&self) -> RangeInclusive<T>