Trait competitive::num::RangeBoundsExt

pub trait RangeBoundsExt<T> {

    // Required methods
    fn start_bound_included_checked(&self) -> Option<T>;
    fn start_bound_excluded_checked(&self) -> Option<T>;
    fn end_bound_included_checked(&self) -> Option<T>;
    fn end_bound_excluded_checked(&self) -> Option<T>;
    fn start_bound_included(&self) -> T;
    fn start_bound_excluded(&self) -> T;
    fn end_bound_included(&self) -> T;
    fn end_bound_excluded(&self) -> T;
    fn start_bound_included_bounded(&self, lb: T) -> Option<T>
       where T: Ord;
    fn start_bound_excluded_bounded(&self, lb: T) -> Option<T>
       where T: Ord;
    fn end_bound_included_bounded(&self, ub: T) -> Option<T>
       where T: Ord;
    fn end_bound_excluded_bounded(&self, ub: T) -> Option<T>
       where T: Ord;

    // Provided methods
    fn to_range_checked(&self) -> Option<Range<T>> { ... }
    fn to_range(&self) -> Range<T> { ... }
    fn to_range_bounded(&self, min: T, max: T) -> Option<Range<T>>
       where T: Ord { ... }
    fn to_range_inclusive_checked(&self) -> Option<RangeInclusive<T>> { ... }
    fn to_range_inclusive(&self) -> RangeInclusive<T> { ... }
    fn to_range_inclusive_bounded(
        &self,
        min: T,
        max: T,
    ) -> Option<RangeInclusive<T>>
       where T: Ord { ... }

}

Required Methods

fn start_bound_included_checked(&self) -> Option<T>

fn start_bound_excluded_checked(&self) -> Option<T>

fn end_bound_included_checked(&self) -> Option<T>

fn end_bound_excluded_checked(&self) -> Option<T>

fn start_bound_included(&self) -> T

fn start_bound_excluded(&self) -> T

fn end_bound_included(&self) -> T

fn end_bound_excluded(&self) -> T

fn start_bound_included_bounded(&self, lb: T) -> Option<T>
where T: Ord,

fn start_bound_excluded_bounded(&self, lb: T) -> Option<T>
where T: Ord,

fn end_bound_included_bounded(&self, ub: T) -> Option<T>
where T: Ord,

fn end_bound_excluded_bounded(&self, ub: T) -> Option<T>
where T: Ord,

Provided Methods

fn to_range_checked(&self) -> Option<Range<T>>

fn to_range(&self) -> Range<T>

Examples found in repository?

crates/competitive/src/data_structure/segment_tree_map.rs (line 93)

89
90
91
92
93
94
95
96
97
98
99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122
123
124
125
126
127
128
129
130
131
132
133
134
135
136
137
138
139
140
141
142
143
144
145
146
147
148
149
150
151
152
153
154
155
156
157
158
159
160
161
162
163
164
165
166
167
168
169
170
171
172
173
174
175
176
177
178
179
180
181
182
183
184
185
186
187
188
189
190
191
192
193
194
195
196
197
198
199
200
201
202
203
204
205
206
207
208
209
210
211
212
213
214
215
216
217
218
219
220

    pub fn fold<R>(&self, range: R) -> M::T
    where
        R: RangeBounds<usize>,
    {
        let range = range.to_range();
        debug_assert!(range.end <= self.n);
        let mut l = range.start + self.n;
        let mut r = range.end + self.n;
        let mut vl = M::unit();
        let mut vr = M::unit();
        while l < r {
            if l & 1 != 0 {
                vl = M::operate(&vl, self.get_ref(l));
                l += 1;
            }
            if r & 1 != 0 {
                r -= 1;
                vr = M::operate(self.get_ref(r), &vr);
            }
            l /= 2;
            r /= 2;
        }
        M::operate(&vl, &vr)
    }
    fn bisect_perfect<F>(&self, mut pos: usize, mut acc: M::T, f: F) -> (usize, M::T)
    where
        F: Fn(&M::T) -> bool,
    {
        while pos < self.n {
            pos <<= 1;
            let nacc = M::operate(&acc, self.get_ref(pos));
            if !f(&nacc) {
                acc = nacc;
                pos += 1;
            }
        }
        (pos - self.n, acc)
    }
    fn rbisect_perfect<F>(&self, mut pos: usize, mut acc: M::T, f: F) -> (usize, M::T)
    where
        F: Fn(&M::T) -> bool,
    {
        while pos < self.n {
            pos = pos * 2 + 1;
            let nacc = M::operate(self.get_ref(pos), &acc);
            if !f(&nacc) {
                acc = nacc;
                pos -= 1;
            }
        }
        (pos - self.n, acc)
    }
    /// Returns the first index that satisfies a accumlative predicate.
    pub fn position_acc<R, F>(&self, range: R, f: F) -> Option<usize>
    where
        R: RangeBounds<usize>,
        F: Fn(&M::T) -> bool,
    {
        let range = range.to_range();
        debug_assert!(range.end <= self.n);
        let mut l = range.start + self.n;
        let r = range.end + self.n;
        let mut k = 0usize;
        let mut acc = M::unit();
        while l < r >> k {
            if l & 1 != 0 {
                let nacc = M::operate(&acc, self.get_ref(l));
                if f(&nacc) {
                    return Some(self.bisect_perfect(l, acc, f).0);
                }
                acc = nacc;
                l += 1;
            }
            l >>= 1;
            k += 1;
        }
        for k in (0..k).rev() {
            let r = r >> k;
            if r & 1 != 0 {
                let nacc = M::operate(&acc, self.get_ref(r - 1));
                if f(&nacc) {
                    return Some(self.bisect_perfect(r - 1, acc, f).0);
                }
                acc = nacc;
            }
        }
        None
    }
    /// Returns the last index that satisfies a accumlative predicate.
    pub fn rposition_acc<R, F>(&self, range: R, f: F) -> Option<usize>
    where
        R: RangeBounds<usize>,
        F: Fn(&M::T) -> bool,
    {
        let range = range.to_range();
        debug_assert!(range.end <= self.n);
        let mut l = range.start + self.n;
        let mut r = range.end + self.n;
        let mut c = 0usize;
        let mut k = 0usize;
        let mut acc = M::unit();
        while l >> k < r {
            c <<= 1;
            if l & 1 << k != 0 {
                l += 1 << k;
                c += 1;
            }
            if r & 1 != 0 {
                r -= 1;
                let nacc = M::operate(self.get_ref(r), &acc);
                if f(&nacc) {
                    return Some(self.rbisect_perfect(r, acc, f).0);
                }
                acc = nacc;
            }
            r >>= 1;
            k += 1;
        }
        for k in (0..k).rev() {
            if c & 1 != 0 {
                l -= 1 << k;
                let l = l >> k;
                let nacc = M::operate(self.get_ref(l), &acc);
                if f(&nacc) {
                    return Some(self.rbisect_perfect(l, acc, f).0);
                }
                acc = nacc;
            }
            c >>= 1;
        }
        None
    }

fn to_range_bounded(&self, min: T, max: T) -> Option<Range<T>>
where T: Ord,

Examples found in repository?

crates/competitive/src/data_structure/accumulate.rs (line 76)

71
72
73
74
75
76
77
78
79
80
81
82
83
84
85
86
87
88
89
90
91
92
93
94
95
96
97
98
99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122
123
124
125
126
127
128
129
130
131
132
133
134
135
136
137
138
139
140
141
142
143
144
145
146
147
148
149
150
151
152
153
154
155
156
157
158
159
160
161
162
163
164
165
166
167
168
169
170
171
172
173
174
175
176
177
178
179
180
181
182
183
184
185
186
187
188
189
190
191
192
193
194
195
196
197
198
199
200
201
202
203
204
205
206
207
208
209
210
211
212
213
214
215
216
217
218
219
220
221
222
223
224
225
226
227
228
229
230
231
232
233
234
235
236
237
238
239
240
241
242
243
244
245
246
247
248
249
250
251
252
253
254
255
256
257
258
259
260
261
262
263
264
265
266
267
268
269
270
271
272
273
274
275
276
277
278
279
280
281
282
283
284
285
286
287
288
289
290
291
292
293
294
295
296
297
298
299
300
301
302
303
304
305
306
307
308
309
310
311

    pub fn fold<R>(&self, range: R) -> M::T
    where
        R: RangeBounds<usize>,
    {
        let n = self.data.len() - 1;
        let range = range.to_range_bounded(0, n).expect("invalid range");
        let (l, r) = (range.start, range.end);
        assert!(l <= r, "bad range [{}, {})", l, r);
        M::operate(&M::inverse(unsafe { self.data.get_unchecked(l) }), unsafe {
            self.data.get_unchecked(r)
        })
    }
}

/// 2-dimensional accumlated data
pub struct Accumulate2d<M>
where
    M: AbelianMonoid,
{
    h: usize,
    w: usize,
    data: Vec<M::T>,
}

impl<M> Debug for Accumulate2d<M>
where
    M: AbelianMonoid,
    M::T: Debug,
{
    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
        f.debug_struct("Accumulate2d")
            .field("h", &self.h)
            .field("w", &self.w)
            .field("data", &self.data)
            .finish()
    }
}

impl<M> Accumulate2d<M>
where
    M: AbelianMonoid,
{
    pub fn new(arr2d: &[Vec<M::T>]) -> Self {
        let h = arr2d.len();
        assert!(h > 0);
        let w = arr2d[0].len();
        assert!(w > 0);
        let w1 = w + 1;
        let mut data = Vec::with_capacity((h + 1) * w1);
        data.resize_with(w1, M::unit);
        for (i, arr) in arr2d.iter().enumerate() {
            assert_eq!(w, arr.len(), "expected 2d array");
            let mut acc = M::unit();
            for (j, x) in arr.iter().enumerate() {
                let y = M::operate(&acc, x);
                data.push(M::operate(&acc, unsafe { data.get_unchecked(w1 * i + j) }));
                acc = y;
            }
            data.push(M::operate(&acc, unsafe { data.get_unchecked(w1 * i + w) }));
        }
        Self { h, w, data }
    }
    pub fn from_fn<F>(h: usize, w: usize, mut f: F) -> Self
    where
        F: FnMut(usize, usize) -> M::T,
    {
        let w1 = w + 1;
        let mut data = Vec::with_capacity((h + 1) * w1);
        data.resize_with(w1, M::unit);
        for i in 0..h {
            let mut acc = M::unit();
            for j in 0..w {
                let y = M::operate(&acc, &f(i, j));
                data.push(M::operate(&acc, unsafe { data.get_unchecked(w1 * i + j) }));
                acc = y;
            }
            data.push(M::operate(&acc, unsafe { data.get_unchecked(w1 * i + w) }));
        }
        Self { h, w, data }
    }
    /// Return fold of \[0, x\) × \[0, y\)
    pub fn accumulate(&self, x: usize, y: usize) -> M::T {
        let h1 = self.h + 1;
        let w1 = self.w + 1;
        assert!(
            x < h1,
            "index out of range: the first len is {} but the index is {}",
            h1,
            x
        );
        assert!(
            y < w1,
            "index out of range: the second len is {} but the index is {}",
            w1,
            y
        );
        unsafe { self.data.get_unchecked(w1 * x + y) }.clone()
    }
}

impl<M> Accumulate2d<M>
where
    M: AbelianGroup,
{
    /// Return fold of range
    pub fn fold<R0, R1>(&self, range0: R0, range1: R1) -> M::T
    where
        R0: RangeBounds<usize>,
        R1: RangeBounds<usize>,
    {
        let range0 = range0.to_range_bounded(0, self.h).expect("invalid range");
        let range1 = range1.to_range_bounded(0, self.w).expect("invalid range");
        let (xl, xr) = (range0.start, range0.end);
        let (yl, yr) = (range1.start, range1.end);
        assert!(xl <= xr, "bad range [{}, {})", xl, xr);
        assert!(yl <= yr, "bad range [{}, {})", yl, yr);
        let w1 = self.w + 1;
        unsafe {
            M::rinv_operate(
                &M::operate(
                    self.data.get_unchecked(w1 * xl + yl),
                    self.data.get_unchecked(w1 * xr + yr),
                ),
                &M::operate(
                    self.data.get_unchecked(w1 * xl + yr),
                    self.data.get_unchecked(w1 * xr + yl),
                ),
            )
        }
    }
}

pub struct AccumulateKd<const K: usize, M>
where
    M: AbelianMonoid,
{
    dim: [usize; K],
    offset: [usize; K],
    data: Vec<M::T>,
}

impl<const K: usize, M> Debug for AccumulateKd<K, M>
where
    M: AbelianMonoid,
    M::T: Debug,
{
    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
        f.debug_struct("AccumulateKd")
            .field("dim", &self.dim)
            .field("offset", &self.offset)
            .field("data", &self.data)
            .finish()
    }
}

impl<const K: usize, M> AccumulateKd<K, M>
where
    M: AbelianMonoid,
{
    pub fn from_fn(dim: [usize; K], mut f: impl FnMut([usize; K]) -> M::T) -> Self {
        fn fill<const K: usize, T>(
            dim: &[usize; K],
            offset: &[usize; K],
            data: &mut [T],
            f: &mut impl FnMut([usize; K]) -> T,
            mut index: [usize; K],
            pos: usize,
        ) {
            if pos < K {
                for i in 0..dim[pos] {
                    index[pos] = i;
                    fill(dim, offset, data, f, index, pos + 1);
                }
            } else {
                let i: usize = index.iter().zip(offset).map(|(x, y)| (x + 1) * y).sum();
                data[i] = f(index);
            }
        }

        let mut offset = [1; K];
        for d in (1..K).rev() {
            offset[d - 1] = offset[d] * (dim[d] + 1);
        }
        let size = offset[0] * (dim[0] + 1);
        let mut data = vec![M::unit(); size];
        fill(&dim, &offset, &mut data, &mut f, [0; K], 0);
        for d in 0..K {
            for i in 1..size {
                if i / offset[d] % (dim[d] + 1) != 0 {
                    data[i] = M::operate(&data[i], &data[i - offset[d]]);
                }
            }
        }
        Self { dim, offset, data }
    }
    pub fn accumulate(&self, x: [usize; K]) -> M::T {
        for (d, x) in x.into_iter().enumerate() {
            assert!(
                x <= self.dim[d],
                "index out of range: the len is {} but the index is {}",
                self.dim[d] + 1,
                x
            );
        }
        let p: usize = x.iter().zip(&self.offset).map(|(x, y)| x * y).sum();
        unsafe { self.data.get_unchecked(p) }.clone()
    }
}

impl<const K: usize, M> AccumulateKd<K, M>
where
    M: AbelianGroup,
{
    pub fn fold<R>(&self, ranges: [R; K]) -> M::T
    where
        R: RangeBounds<usize>,
    {
        let ranges: [_; K] = std::array::from_fn(|i| {
            let range = ranges[i]
                .to_range_bounded(0, self.dim[i])
                .expect("invalid range");
            let (l, r) = (range.start, range.end);
            assert!(l <= r, "bad range [{}, {})", l, r);
            [l, r]
        });
        let mut acc = M::unit();
        for bit in 0..1 << K {
            let p: usize = ranges
                .iter()
                .zip(&self.offset)
                .enumerate()
                .map(|(d, (range, offset))| range[bit >> d & 1 ^ 1] * offset)
                .sum();
            if bit.count_ones() & 1 == 0 {
                acc = M::operate(&acc, unsafe { self.data.get_unchecked(p) });
            } else {
                acc = M::rinv_operate(&acc, unsafe { self.data.get_unchecked(p) });
            }
        }
        acc
    }

Hide additional examples

crates/competitive/src/data_structure/segment_tree.rs (line 90)

86
87
88
89
90
91
92
93
94
95
96
97
98
99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122
123
124
125
126
127
128
129
130
131
132
133
134
135
136
137
138
139
140
141
142
143
144
145
146
147
148
149
150
151
152
153
154
155
156
157
158
159
160
161
162
163
164
165
166
167
168
169
170
171
172
173
174
175
176
177
178
179
180
181
182
183
184
185
186
187
188
189
190
191
192
193
194
195
196
197
198
199
200
201
202
203
204
205
206
207
208
209
210
211
212
213
214

    pub fn fold<R>(&self, range: R) -> M::T
    where
        R: RangeBounds<usize>,
    {
        let range = range.to_range_bounded(0, self.n).expect("invalid range");
        let mut l = range.start + self.n;
        let mut r = range.end + self.n;
        let mut vl = M::unit();
        let mut vr = M::unit();
        while l < r {
            if l & 1 != 0 {
                vl = M::operate(&vl, &self.seg[l]);
                l += 1;
            }
            if r & 1 != 0 {
                r -= 1;
                vr = M::operate(&self.seg[r], &vr);
            }
            l /= 2;
            r /= 2;
        }
        M::operate(&vl, &vr)
    }
    fn bisect_perfect<F>(&self, mut pos: usize, mut acc: M::T, f: F) -> (usize, M::T)
    where
        F: Fn(&M::T) -> bool,
    {
        while pos < self.n {
            pos <<= 1;
            let nacc = M::operate(&acc, &self.seg[pos]);
            if !f(&nacc) {
                acc = nacc;
                pos += 1;
            }
        }
        (pos - self.n, acc)
    }
    fn rbisect_perfect<F>(&self, mut pos: usize, mut acc: M::T, f: F) -> (usize, M::T)
    where
        F: Fn(&M::T) -> bool,
    {
        while pos < self.n {
            pos = pos * 2 + 1;
            let nacc = M::operate(&self.seg[pos], &acc);
            if !f(&nacc) {
                acc = nacc;
                pos -= 1;
            }
        }
        (pos - self.n, acc)
    }
    /// Returns the first index that satisfies a accumlative predicate.
    pub fn position_acc<R, F>(&self, range: R, f: F) -> Option<usize>
    where
        R: RangeBounds<usize>,
        F: Fn(&M::T) -> bool,
    {
        let range = range.to_range_bounded(0, self.n).expect("invalid range");
        let mut l = range.start + self.n;
        let r = range.end + self.n;
        let mut k = 0usize;
        let mut acc = M::unit();
        while l < r >> k {
            if l & 1 != 0 {
                let nacc = M::operate(&acc, &self.seg[l]);
                if f(&nacc) {
                    return Some(self.bisect_perfect(l, acc, f).0);
                }
                acc = nacc;
                l += 1;
            }
            l >>= 1;
            k += 1;
        }
        for k in (0..k).rev() {
            let r = r >> k;
            if r & 1 != 0 {
                let nacc = M::operate(&acc, &self.seg[r - 1]);
                if f(&nacc) {
                    return Some(self.bisect_perfect(r - 1, acc, f).0);
                }
                acc = nacc;
            }
        }
        None
    }
    /// Returns the last index that satisfies a accumlative predicate.
    pub fn rposition_acc<R, F>(&self, range: R, f: F) -> Option<usize>
    where
        R: RangeBounds<usize>,
        F: Fn(&M::T) -> bool,
    {
        let range = range.to_range_bounded(0, self.n).expect("invalid range");
        let mut l = range.start + self.n;
        let mut r = range.end + self.n;
        let mut c = 0usize;
        let mut k = 0usize;
        let mut acc = M::unit();
        while l >> k < r {
            c <<= 1;
            if l & 1 << k != 0 {
                l += 1 << k;
                c += 1;
            }
            if r & 1 != 0 {
                r -= 1;
                let nacc = M::operate(&self.seg[r], &acc);
                if f(&nacc) {
                    return Some(self.rbisect_perfect(r, acc, f).0);
                }
                acc = nacc;
            }
            r >>= 1;
            k += 1;
        }
        for k in (0..k).rev() {
            if c & 1 != 0 {
                l -= 1 << k;
                let l = l >> k;
                let nacc = M::operate(&self.seg[l], &acc);
                if f(&nacc) {
                    return Some(self.rbisect_perfect(l, acc, f).0);
                }
                acc = nacc;
            }
            c >>= 1;
        }
        None
    }

fn to_range_inclusive_checked(&self) -> Option<RangeInclusive<T>>

fn to_range_inclusive(&self) -> RangeInclusive<T>

fn to_range_inclusive_bounded( &self, min: T, max: T, ) -> Option<RangeInclusive<T>>
where T: Ord,