competitive::tree::heavy_light_decomposition

Struct HeavyLightDecomposition

pub struct HeavyLightDecomposition {
    parent: Vec<usize>,
    size: Vec<usize>,
    head: Vec<usize>,
    index: Vec<usize>,
    order: Vec<usize>,
}

Fields
parent: Vec<usize>size: Vec<usize>head: Vec<usize>index: Vec<usize>order: Vec<usize>
Implementations

impl HeavyLightDecomposition

pub fn new(root: usize, graph: &UndirectedSparseGraph) -> Self

Examples found in repository?

crates/competitive/src/tree/heavy_light_decomposition.rs (line 15)

14    pub fn hld(&self, root: usize) -> HeavyLightDecomposition {
15        HeavyLightDecomposition::new(root, self)
16    }

fn dfs_size(&mut self, u: usize, p: usize, graph: &UndirectedSparseGraph)

Examples found in repository?

crates/competitive/src/tree/heavy_light_decomposition.rs (line 29)

20    pub fn new(root: usize, graph: &UndirectedSparseGraph) -> Self {
21        let n = graph.vertices_size();
22        let mut self_ = Self {
23            parent: vec![n; n],
24            size: vec![0; n],
25            head: vec![n; n],
26            index: vec![0; n],
27            order: Vec::with_capacity(n),
28        };
29        self_.dfs_size(root, n, graph);
30        self_.dfs_hld(root, n, root, graph);
31        self_
32    }
33
34    fn dfs_size(&mut self, u: usize, p: usize, graph: &UndirectedSparseGraph) {
35        self.parent[u] = p;
36        self.size[u] = 1;
37        for a in graph.adjacencies(u) {
38            if a.to != p {
39                self.dfs_size(a.to, u, graph);
40                self.size[u] += self.size[a.to];
41                if self.head[u] == graph.vertices_size()
42                    || self.size[self.head[u]] < self.size[a.to]
43                {
44                    self.head[u] = a.to;
45                }
46            }
47        }
48    }

fn dfs_hld( &mut self, u: usize, p: usize, head: usize, graph: &UndirectedSparseGraph, )

Examples found in repository?

crates/competitive/src/tree/heavy_light_decomposition.rs (line 30)

20    pub fn new(root: usize, graph: &UndirectedSparseGraph) -> Self {
21        let n = graph.vertices_size();
22        let mut self_ = Self {
23            parent: vec![n; n],
24            size: vec![0; n],
25            head: vec![n; n],
26            index: vec![0; n],
27            order: Vec::with_capacity(n),
28        };
29        self_.dfs_size(root, n, graph);
30        self_.dfs_hld(root, n, root, graph);
31        self_
32    }
33
34    fn dfs_size(&mut self, u: usize, p: usize, graph: &UndirectedSparseGraph) {
35        self.parent[u] = p;
36        self.size[u] = 1;
37        for a in graph.adjacencies(u) {
38            if a.to != p {
39                self.dfs_size(a.to, u, graph);
40                self.size[u] += self.size[a.to];
41                if self.head[u] == graph.vertices_size()
42                    || self.size[self.head[u]] < self.size[a.to]
43                {
44                    self.head[u] = a.to;
45                }
46            }
47        }
48    }
49
50    fn dfs_hld(&mut self, u: usize, p: usize, head: usize, graph: &UndirectedSparseGraph) {
51        let heavy = self.head[u];
52        self.head[u] = head;
53        self.index[u] = self.order.len();
54        self.order.push(u);
55        if heavy != graph.vertices_size() {
56            self.dfs_hld(heavy, u, head, graph);
57        }
58        for a in graph.adjacencies(u) {
59            if a.to != p && a.to != heavy {
60                self.dfs_hld(a.to, u, a.to, graph);
61            }
62        }
63    }

pub fn len(&self) -> usize

Examples found in repository?

crates/competitive/src/tree/heavy_light_decomposition.rs (line 82)

81    pub fn parent(&self, v: usize) -> Option<usize> {
82        (self.parent[v] < self.len()).then_some(self.parent[v])
83    }

pub fn is_empty(&self) -> bool

pub fn root(&self) -> usize

pub fn parent(&self, v: usize) -> Option<usize>

Examples found in repository?

crates/competitive/src/tree/heavy_light_decomposition.rs (line 119)

111    pub fn kth_ancestor(&self, mut v: usize, mut k: usize) -> Option<usize> {
112        loop {
113            let head = self.head[v];
114            let chain_len = self.index[v] - self.index[head];
115            if k <= chain_len {
116                return Some(self.order[self.index[v] - k]);
117            }
118            k -= chain_len + 1;
119            v = self.parent(head)?;
120        }
121    }

pub fn index(&self, v: usize) -> usize

Examples found in repository?

crates/library_checker/src/tree/vertex_add_path_sum.rs (line 21)

14pub fn vertex_add_path_sum(reader: impl Read, mut writer: impl Write) {
15    let s = read_all_unchecked(reader);
16    let mut scanner = Scanner::new(&s);
17    scan!(scanner, n, q, a: [i64; n], (graph, _): @TreeGraphScanner::<usize, ()>::new(n));
18    let hld = graph.hld(0);
19    let mut bit = BinaryIndexedTree::<AdditiveOperation<_>>::new(n);
20    for (v, &x) in a.iter().enumerate() {
21        bit.update(hld.index(v), x);
22    }
23    for _ in 0..q {
24        scan!(scanner, query: Query);
25        match query {
26            Query::Add { p, x } => {
27                bit.update(hld.index(p), x);
28            }
29            Query::Sum { u, v } => {
30                let mut sum = 0;
31                hld.path_vertices(u, v, |l, r| sum += bit.fold(l, r));
32                writeln!(writer, "{sum}").ok();
33            }
34        }
35    }
36}

Hide additional examples

crates/library_checker/src/tree/vertex_add_subtree_sum.rs (line 45)

36pub fn vertex_add_subtree_sum_hld(reader: impl Read, mut writer: impl Write) {
37    let s = read_all_unchecked(reader);
38    let mut scanner = Scanner::new(&s);
39    scan!(scanner, n, q, a: [u64; n], p: [usize]);
40    let edges = p.take(n - 1).enumerate().map(|(i, p)| (i + 1, p)).collect();
41    let tree = UndirectedSparseGraph::from_edges(n, edges);
42    let hld = tree.hld(0);
43    let mut b = vec![0; n];
44    for (v, x) in a.into_iter().enumerate() {
45        b[hld.index(v)] = x;
46    }
47    let mut seg = SegmentTree::<AdditiveOperation<_>>::from_vec(b);
48    for _ in 0..q {
49        scan!(scanner, query: Query);
50        match query {
51            Query::Add { u, x } => seg.update(hld.index(u), x),
52            Query::Sum { u } => {
53                writeln!(writer, "{}", seg.fold(hld.subtree_range(u))).ok();
54            }
55        }
56    }
57}

crates/library_checker/src/tree/vertex_set_path_composite.rs (line 24)

17pub fn vertex_set_path_composite(reader: impl Read, mut writer: impl Write) {
18    let s = read_all_unchecked(reader);
19    let mut scanner = Scanner::new(&s);
20    scan!(scanner, n, q, ab: [(MInt998244353, MInt998244353); n], (graph, _): @TreeGraphScanner::<usize, ()>::new(n));
21    let hld = graph.hld(0);
22    let mut nab = vec![(MInt998244353::default(), MInt998244353::default()); n];
23    for i in 0..n {
24        nab[hld.index(i)] = ab[i];
25    }
26    let mut seg1 = SegmentTree::<LinearOperation<_>>::from_vec(nab.clone());
27    let mut seg2 = SegmentTree::<ReverseOperation<LinearOperation<_>>>::from_vec(nab);
28    for _ in 0..q {
29        scan!(scanner, query: Query);
30        match query {
31            Query::Set { p, cd } => {
32                seg1.set(hld.index(p), cd);
33                seg2.set(hld.index(p), cd);
34            }
35            Query::Apply { u, v, x } => {
36                let (a, b) = hld.fold_vertices::<LinearOperation<_>, _, _>(
37                    u,
38                    v,
39                    |l, r| seg1.fold(l..r),
40                    |l, r| seg2.fold(l..r),
41                );
42                writeln!(writer, "{}", a * x + b).ok();
43            }
44        }
45    }
46}

pub fn vertex(&self, index: usize) -> usize

pub fn subtree_size(&self, v: usize) -> usize

pub fn subtree_range(&self, v: usize) -> Range<usize>

Examples found in repository?

crates/competitive/src/tree/heavy_light_decomposition.rs (line 107)

106    pub fn is_ancestor(&self, ancestor: usize, v: usize) -> bool {
107        self.subtree_range(ancestor).contains(&self.index[v])
108    }

Hide additional examples

crates/library_checker/src/tree/vertex_add_subtree_sum.rs (line 53)

36pub fn vertex_add_subtree_sum_hld(reader: impl Read, mut writer: impl Write) {
37    let s = read_all_unchecked(reader);
38    let mut scanner = Scanner::new(&s);
39    scan!(scanner, n, q, a: [u64; n], p: [usize]);
40    let edges = p.take(n - 1).enumerate().map(|(i, p)| (i + 1, p)).collect();
41    let tree = UndirectedSparseGraph::from_edges(n, edges);
42    let hld = tree.hld(0);
43    let mut b = vec![0; n];
44    for (v, x) in a.into_iter().enumerate() {
45        b[hld.index(v)] = x;
46    }
47    let mut seg = SegmentTree::<AdditiveOperation<_>>::from_vec(b);
48    for _ in 0..q {
49        scan!(scanner, query: Query);
50        match query {
51            Query::Add { u, x } => seg.update(hld.index(u), x),
52            Query::Sum { u } => {
53                writeln!(writer, "{}", seg.fold(hld.subtree_range(u))).ok();
54            }
55        }
56    }
57}

pub fn is_ancestor(&self, ancestor: usize, v: usize) -> bool

pub fn kth_ancestor(&self, v: usize, k: usize) -> Option<usize>

Examples found in repository?

crates/competitive/src/tree/heavy_light_decomposition.rs (line 145)

142    pub fn jump(&self, u: usize, v: usize, k: usize) -> Option<usize> {
143        let (up, down) = self.path_lengths(u, v);
144        if k <= up {
145            self.kth_ancestor(u, k)
146        } else if k <= up + down {
147            self.kth_ancestor(v, up + down - k)
148        } else {
149            None
150        }
151    }

pub fn lca(&self, u: usize, v: usize) -> usize

Examples found in repository?

crates/library_checker/src/tree/lca.rs (line 29)

20pub fn lca_hld(reader: impl Read, mut writer: impl Write) {
21    let s = read_all_unchecked(reader);
22    let mut scanner = Scanner::new(&s);
23    scan!(scanner, n, q, p: [usize]);
24    let edges = p.take(n - 1).enumerate().map(|(i, p)| (i + 1, p)).collect();
25    let graph = UndirectedSparseGraph::from_edges(n, edges);
26    let hld = graph.hld(0);
27    for _ in 0..q {
28        scan!(scanner, u, v);
29        writeln!(writer, "{}", hld.lca(u, v)).ok();
30    }
31}

pub fn distance(&self, u: usize, v: usize) -> usize

pub fn jump(&self, u: usize, v: usize, k: usize) -> Option<usize>

Examples found in repository?

crates/library_checker/src/tree/jump_on_tree.rs (line 12)

5pub fn jump_on_tree(reader: impl Read, mut writer: impl Write) {
6    let s = read_all_unchecked(reader);
7    let mut scanner = Scanner::new(&s);
8    scan!(scanner, n, q, (g, _): @TreeGraphScanner::<usize>::new(n));
9    let hld = g.hld(0);
10    for _ in 0..q {
11        scan!(scanner, s, t, i);
12        writeln!(writer, "{}", hld.jump(s, t, i).unwrap_or(!0) as isize).ok();
13    }
14}

fn path_lengths(&self, u: usize, v: usize) -> (usize, usize)

Examples found in repository?

crates/competitive/src/tree/heavy_light_decomposition.rs (line 137)

136    pub fn distance(&self, u: usize, v: usize) -> usize {
137        let (up, down) = self.path_lengths(u, v);
138        up + down
139    }
140
141    #[inline]
142    pub fn jump(&self, u: usize, v: usize, k: usize) -> Option<usize> {
143        let (up, down) = self.path_lengths(u, v);
144        if k <= up {
145            self.kth_ancestor(u, k)
146        } else if k <= up + down {
147            self.kth_ancestor(v, up + down - k)
148        } else {
149            None
150        }
151    }

pub fn path_vertices<F: FnMut(usize, usize)>(&self, u: usize, v: usize, f: F)

Calls f once for each nonempty DFS-index range on the vertex path. The callback order is unspecified.

Examples found in repository?

crates/library_checker/src/tree/vertex_add_path_sum.rs (line 31)

14pub fn vertex_add_path_sum(reader: impl Read, mut writer: impl Write) {
15    let s = read_all_unchecked(reader);
16    let mut scanner = Scanner::new(&s);
17    scan!(scanner, n, q, a: [i64; n], (graph, _): @TreeGraphScanner::<usize, ()>::new(n));
18    let hld = graph.hld(0);
19    let mut bit = BinaryIndexedTree::<AdditiveOperation<_>>::new(n);
20    for (v, &x) in a.iter().enumerate() {
21        bit.update(hld.index(v), x);
22    }
23    for _ in 0..q {
24        scan!(scanner, query: Query);
25        match query {
26            Query::Add { p, x } => {
27                bit.update(hld.index(p), x);
28            }
29            Query::Sum { u, v } => {
30                let mut sum = 0;
31                hld.path_vertices(u, v, |l, r| sum += bit.fold(l, r));
32                writeln!(writer, "{sum}").ok();
33            }
34        }
35    }
36}

pub fn path_edges<F: FnMut(usize, usize)>(&self, u: usize, v: usize, f: F)

Calls f once for each nonempty DFS-index range on the edge path. Each index represents the deeper endpoint of an edge. The callback order is unspecified.

Examples found in repository?

crates/aizu_online_judge/src/grl/grl_5_e.rs (line 33)

15pub fn grl_5_e(reader: impl Read, mut writer: impl Write) {
16    let s = read_all_unchecked(reader);
17    let mut scanner = Scanner::new(&s);
18    scan!(scanner, n, c: [SizedCollect<usize>]);
19    let edges = c
20        .take(n)
21        .enumerate()
22        .flat_map(|(u, it)| it.into_iter().map(move |v| (u, v)))
23        .collect();
24    let graph = UndirectedSparseGraph::from_edges(n, edges);
25    let hld = graph.hld(0);
26    let mut seg = LazySegmentTree::<RangeSumRangeAdd<_>>::from_vec(vec![(0u64, 1u64); n]);
27
28    scan!(scanner, q);
29    for _ in 0..q {
30        scan!(scanner, query: Query);
31        match query {
32            Query::Add { v, w } => {
33                hld.path_edges(0, v, |l, r| seg.update(l..r, w));
34            }
35            Query::Get { u } => {
36                let mut ans = 0;
37                hld.path_edges(0, u, |l, r| ans += seg.fold(l..r).0);
38                writeln!(writer, "{}", ans).ok();
39            }
40        }
41    }
42}

fn path<F: FnMut(usize, usize)>(&self, u: usize, v: usize, is_edge: bool, f: F)

Examples found in repository?

crates/competitive/src/tree/heavy_light_decomposition.rs (line 177)

176    pub fn path_vertices<F: FnMut(usize, usize)>(&self, u: usize, v: usize, f: F) {
177        self.path(u, v, false, f);
178    }
179
180    /// Calls `f` once for each nonempty DFS-index range on the edge path.
181    /// Each index represents the deeper endpoint of an edge. The callback order is unspecified.
182    #[inline]
183    pub fn path_edges<F: FnMut(usize, usize)>(&self, u: usize, v: usize, f: F) {
184        self.path(u, v, true, f);
185    }

pub fn fold_vertices<M: Monoid, F1: FnMut(usize, usize) -> M::T, F2: FnMut(usize, usize) -> M::T>( &self, u: usize, v: usize, forward: F1, reverse: F2, ) -> M::T

Folds a vertex path in u-to-v order. forward folds a DFS-index range from left to right, and reverse folds it from right to left.

Examples found in repository?

crates/library_checker/src/tree/vertex_set_path_composite.rs (lines 36-41)

17pub fn vertex_set_path_composite(reader: impl Read, mut writer: impl Write) {
18    let s = read_all_unchecked(reader);
19    let mut scanner = Scanner::new(&s);
20    scan!(scanner, n, q, ab: [(MInt998244353, MInt998244353); n], (graph, _): @TreeGraphScanner::<usize, ()>::new(n));
21    let hld = graph.hld(0);
22    let mut nab = vec![(MInt998244353::default(), MInt998244353::default()); n];
23    for i in 0..n {
24        nab[hld.index(i)] = ab[i];
25    }
26    let mut seg1 = SegmentTree::<LinearOperation<_>>::from_vec(nab.clone());
27    let mut seg2 = SegmentTree::<ReverseOperation<LinearOperation<_>>>::from_vec(nab);
28    for _ in 0..q {
29        scan!(scanner, query: Query);
30        match query {
31            Query::Set { p, cd } => {
32                seg1.set(hld.index(p), cd);
33                seg2.set(hld.index(p), cd);
34            }
35            Query::Apply { u, v, x } => {
36                let (a, b) = hld.fold_vertices::<LinearOperation<_>, _, _>(
37                    u,
38                    v,
39                    |l, r| seg1.fold(l..r),
40                    |l, r| seg2.fold(l..r),
41                );
42                writeln!(writer, "{}", a * x + b).ok();
43            }
44        }
45    }
46}

pub fn fold_edges<M: Monoid, F1: FnMut(usize, usize) -> M::T, F2: FnMut(usize, usize) -> M::T>( &self, u: usize, v: usize, forward: F1, reverse: F2, ) -> M::T

Folds an edge path in u-to-v order. Each index represents the deeper endpoint of an edge. forward folds a DFS-index range from left to right, and reverse folds it from right to left.

fn fold<M: Monoid, F1: FnMut(usize, usize) -> M::T, F2: FnMut(usize, usize) -> M::T>( &self, u: usize, v: usize, is_edge: bool, forward: F1, reverse: F2, ) -> M::T

Examples found in repository?

crates/competitive/src/tree/heavy_light_decomposition.rs (line 221)

210    pub fn fold_vertices<
211        M: Monoid,
212        F1: FnMut(usize, usize) -> M::T,
213        F2: FnMut(usize, usize) -> M::T,
214    >(
215        &self,
216        u: usize,
217        v: usize,
218        forward: F1,
219        reverse: F2,
220    ) -> M::T {
221        self.fold::<M, _, _>(u, v, false, forward, reverse)
222    }
223
224    /// Folds an edge path in `u`-to-`v` order.
225    /// Each index represents the deeper endpoint of an edge. `forward` folds a DFS-index range
226    /// from left to right, and `reverse` folds it from right to left.
227    #[inline]
228    pub fn fold_edges<
229        M: Monoid,
230        F1: FnMut(usize, usize) -> M::T,
231        F2: FnMut(usize, usize) -> M::T,
232    >(
233        &self,
234        u: usize,
235        v: usize,
236        forward: F1,
237        reverse: F2,
238    ) -> M::T {
239        self.fold::<M, _, _>(u, v, true, forward, reverse)
240    }